% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{absdiff}
\alias{absdiff}
\title{Absolute Difference Between Two Images}
\usage{
absdiff(e1, e2, target = "new")
}
\arguments{
\item{e1}{An \code{\link{Image}} object.}

\item{e2}{An \code{\link{Image}} object.}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{e1} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{e1} but will replace that of \code{target}. Note that
  if \code{target} does not have the same dimensions, colorspace, and
  bitdepth as \code{e1}, nothing will be stored.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{e1} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
This function computes the absolute difference between two
 \code{\link{Image}} objects.
}
\examples{
balloon1 <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon2 <- image(system.file("sample_img/balloon2.png", package = "Rvision"))
plot(absdiff(balloon1, balloon2))

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
