% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{addWeighted}
\alias{addWeighted}
\title{Weighted Sum of Two Images}
\usage{
addWeighted(image1, alpha, image2, beta)
}
\arguments{
\item{image1}{An \code{\link{Image}} object.}

\item{alpha}{Weight of the first image.}

\item{image2}{An \code{\link{Image}} object.}

\item{beta}{Weight of the second image.}
}
\value{
An \code{\link{Image}} object.
}
\description{
This function computes the weighted sum of two
 \code{\link{Image}} objects.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
