% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ximgproc.R
\name{anisotropicDiffusion}
\alias{anisotropicDiffusion}
\title{Anisotropic Diffusion}
\usage{
anisotropicDiffusion(image, alpha = 0.1, K = 1, n_iters = 1, target = "new")
}
\arguments{
\item{image}{An 8-bit (8U) 3-channel \code{\link{Image}} object.}

\item{alpha}{The amount of time to step forward by on each iteration
(normally, it's between 0 and 1; default: 0.1).}

\item{K}{Sensitivity to the edges (default: 1).}

\item{n_iters}{The number of diffusion iterations.}

\item{target}{The location where the results should be stored. It can take 3
values:
\describe{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{image} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  if \code{target} does not have the same dimensions, number of channels, and
  bit depth as \code{image}, an error may be thrown.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{image} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
\code{anisotropicDiffusion} Perona-Malik anisotropic diffusion
 to an \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_smooth <- anisotropicDiffusion(balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{gaussianBlur}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
