% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{arcLength}
\alias{arcLength}
\title{Calculate a Contour Perimeter or a Curve Length}
\usage{
arcLength(curve, closed = TRUE)
}
\arguments{
\item{curve}{An m x 2 matrix of 2D coordinates.}

\item{closed}{A boolean indicating whether the curve is closed (perimeter) or
not (default: TRUE).}
}
\value{
A numerical value.
}
\description{
\code{arcLength} estimates a closed contour perimeter or a curve
 length from a set of 2D coordinates.
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_gray <- changeColorSpace(dots, "GRAY")
dots_bin <- dots_gray < 200
contours <- findContours(dots_bin)
ix <- contours$contours[, 1] == 0
arcLength(contours$contours[ix, 2:3])

}
\seealso{
\code{\link{findContours}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
