% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{as.array.Rcpp_Image}
\alias{as.array.Rcpp_Image}
\alias{as.matrix.Rcpp_Image}
\title{Convert Image to Array or Matrix}
\usage{
as.array(image)
as.matrix(image)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}
}
\value{
A matrix or array of the same dimensions as the \code{\link{Image}}
 object.
}
\examples{
# TODO

}
\seealso{
\code{\link{Image}}, \code{\link{matrix}}, \code{\link{array}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
