% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{border}
\alias{border}
\title{Add Border to Image}
\usage{
border(
  image,
  top,
  bottom = top,
  left = top,
  right = top,
  border_type = "constant",
  border_color = "black"
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{top, bottom, left, right}{The width in pixels of the border on each side
of the image. By default, \code{bottom}, \code{left} and \code{right} are
set to the same value as \code{top}.}

\item{border_type}{A character string indicating the extrapolation method to
use when filling empty pixels created during the transformation. It can be
any of the following:
\itemize{
 \item{"constant" (the default):}{\code{iiiiii|abcdefgh|iiiiii} with \code{i}
  specified by \code{border_value}.}
 \item{"replicate":}{\code{aaaaaa|abcdefgh|hhhhhh}.}
 \item{"reflect":}{\code{fedcba|abcdefgh|hgfedc}.}
 \item{"wrap":}{\code{cdefgh|abcdefgh|abcdef}.}
 \item{"reflect_101":}{\code{gfedcb|abcdefgh|gfedcb}.}
 \item{"transparent":}{\code{uvwxyz|abcdefgh|ijklmn}.}
}}

\item{border_color}{A value or vector of any kind of R color specification
compatible with \code{\link{col2bgr}} representing the color of the border
(default: "black").}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{border} adds a border to an image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_border <- border(balloon, 10)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
