% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{boxPoints}
\alias{boxPoints}
\title{Find Vertices of a Rotated Rectangle}
\usage{
boxPoints(rect)
}
\arguments{
\item{rect}{A list describing a rotated rectangle as created by
\code{\link{minAreaRect}} and \code{\link{fitEllipse}}.}
}
\value{
A matrix containing the coordinates of the four vertices of the
 rotated rectangle in the following order: bottom left, top left, top right,
 and bottom right.
}
\description{
\code{boxPoints} finds the four vertices of a rotated rectangle
 computed by \code{\link{minAreaRect}} or \code{\link{fitEllipse}}.
}
\examples{
rect <- minAreaRect(rnorm(100), rnorm(100))
boxPoints(rect)

}
\seealso{
\code{\link{minAreaRect}}, \code{\link{fitEllipse}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
