% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib3d.R
\name{calibrateCamera}
\alias{calibrateCamera}
\title{Find a Camera's Intrinsic and Extrinsic Parameters}
\usage{
calibrateCamera(
  ref_points,
  img_points,
  nrow,
  ncol,
  fixed_point = 1,
  maxit = 30,
  eps = .Machine$double.eps
)
}
\arguments{
\item{ref_points}{List of matrices of calibration pattern points for each
view, in the calibration pattern coordinate space. Each matrix should have
the same dimensions as the corresponding matrix in \code{img_points}.}

\item{img_points}{List of matrices of the projections of the calibration
pattern points in each view. If \code{fixed_point > 0}, the same calibration
pattern must be used in each view and it must be fully visible. Moreover,
all matrices must have the same dimensions and all points in the calibration
pattern should be roughly close to a plane. The calibration target has to be
rigid, or at least static if the camera (rather than the calibration target)
is shifted when grabbing views. See \code{\link{findChessboardCorners}} and
\code{\link{cornerSubPix}} for more information about generating projection
matrices from images of the calibration pattern.}

\item{nrow, ncol}{The number of rows and columns of the images used to capture
different views of the calibration pattern.}

\item{fixed_point}{The index of the reference point in \code{ref_points[[1]]}
to be fixed (default: 1). Usually the top-right corner point of the
calibration pattern is recommended to be fixed. If \code{fixed_point = 0},
then no point is fixed and a less precise calibration algorithm is then
used.}

\item{maxit}{The maximum number of iterations of the detection algorithm
(default: 30).}

\item{eps}{The desired accuracy or change in parameters at which the
iterative algorithm stops (default: \code{.Machine$double.eps}).}
}
\value{
A list of matrices:
 \itemize{
  \item{\code{camera_matrix}:}{a 3x3 camera intrinsic matrix.}
  \item{\code{dist_coeffs:}}{a single row matrix with 4, 5, 8, 12 or 14
   elements representing distortion coefficients.}
  \item{\code{r_vecs:}}{a 3x\code{length(img_points)} matrix of the rotation
   vectors estimated for each calibration pattern view. Together with the
   translation vectors below, this is equivalent to the position of the
   calibration pattern with respect to the camera coordinate space.}
  \item{\code{t_vecs:}}{a 3x\code{length(img_points)} matrix of the
   translation vectors estimated for each calibration pattern view. Together
   with the rotation vectors above, this is equivalent to the position of the
   calibration pattern with respect to the camera coordinate space.}
  \item{\code{new_ref_points:}}{If \code{fixed_point > 0}, this is an updated
   matrix of calibration pattern points. The coordinates might be scaled
   based on the fixed point defined above. The returned coordinates are
   accurate only if the above mentioned fixed point is accurate. If
   \code{fixed_point = 0}, an empty matrix is returned instead.}
 }
}
\description{
\code{calibrateCamera} finds a camera's intrinsic and extrinsic
 parameters from several views of a calibration pattern.
}
\examples{
# See the help vignette:
\dontrun{ vignette("z8_calib", package = "Rvision") }

}
\references{
Bouguet, J.-Y. (2022). Camera Calibration Toolbox for Matlab.
 CaltechDATA. https://doi.org/10.22002/D1.20164

 Strobl, K. H., & Hirzinger, G. (2011, November). More accurate pinhole
 camera calibration with imperfect planar target. 2011 IEEE International
 Conference on Computer Vision Workshops (ICCV Workshops). 2011 IEEE
 International Conference on Computer Vision Workshops (ICCV Workshops),
 Barcelona, Spain. https://doi.org/10.1109/iccvw.2011.6130369

 Zhang, Z. (2000). A flexible new technique for camera calibration. IEEE
 Transactions on Pattern Analysis and Machine Intelligence, 22(11), 1330–1334.
 https://doi.org/10.1109/34.888718
}
\seealso{
\code{\link{findChessboardCorners}}, \code{\link{cornerSubPix}},
 \code{\link{getOptimalNewCameraMatrix}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
