% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{canny}
\alias{canny}
\title{Canny Edge Detector}
\usage{
canny(image, threshold1, threshold2, aperture_size = 3, L2_gradient = FALSE)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{threshold1}{A numeric indicating the first threshold for the hysteresis
procedure}

\item{threshold2}{A numeric indicating the second threshold for the hysteresis
procedure}

\item{aperture_size}{Aperture size for the Sobel operator (default: 3).}

\item{L2_gradient}{A logical flag, indicating whether a more accurate L2 norm
should be used to calculate the image gradient magnitude, or whether the
default L1 norm is enough (default: FALSE).}
}
\description{
\code{canny} finds edges in an image using the Canny algorithm.
}
\references{
Canny J. A computational approach to edge detection. IEEE Trans
 Pattern Anal Mach Intell. 1986;8: 679–698. doi:10.1109/TPAMI.1986.4767851
}
