% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\name{capacity}
\alias{capacity}
\title{Capacity of a Queue}
\usage{
capacity(x)
}
\arguments{
\item{x}{A \code{\link{Queue}} object.}
}
\value{
A numeric value.
}
\description{
\code{capacity} returns the maximum number of elements that a
 \code{\link{Queue}} object can contain.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
buf <- queue(balloon)
capacity(buf)

}
\seealso{
\code{\link{Queue}}, \code{\link{queue}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
