% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{cartToPolar}
\alias{cartToPolar}
\title{Convert Cartesian Coordinates to Polar}
\usage{
cartToPolar(x, y, magnitude = "new", angle = "new", degree = FALSE)
}
\arguments{
\item{x}{A 32- or 64-bit (32F or 64F) \code{\link{Image}} object
corresponding to the x coordinates of the vector field.}

\item{y}{A 32- or 64-bit (32F or 64F) \code{\link{Image}} object
corresponding to the y coordinates of the vector field.}

\item{magnitude}{The location where the magnitude should be stored. It can
take 2 values:
\describe{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. Note that an error will be thrown if
  \code{magnitude} does not have the same dimensions, number of channels,
  and bit depth as \code{x} and \code{y}.}
}}

\item{angle}{The location where the angle should be stored. It can take 2
values:
\describe{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. Note that an error will be thrown if
  \code{angle} does not have the same dimensions, number of channels,
  and bit depth as \code{x} and \code{y}.}
}}

\item{degree}{A logical indicating whether the angles are measured in radians
(the default) or degrees.}
}
\value{
If \code{magnitude="new"} and \code{angle="new"}, the function returns
 a list containing two \code{\link{Image}} objects. If \code{magnitude} and
 \code{angle} are \code{\link{Image}} objects, the function returns nothing
 and modifies these \code{\link{Image}} objects in place.
}
\description{
\code{cartToPolar} converts the x and y coordinates of a vector
 field (for instance, as generated by \code{\link{spatialGradient}}) and
 computes their polar representation (magnitude and angle).
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
field <- spatialGradient(balloon)
lapply(field, changeBitDepth, bitdepth = "32F", target = "self")
field_converted <- cartToPolar(field$dx, field$dy)

}
\seealso{
\code{\link{polarToCart}}, \code{\link{spatialGradient}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
