% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\docType{data}
\name{cc_table}
\alias{cc_table}
\title{Table of Color Space Conversions}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(cc_table)
}
\description{
A data frame containing all the possible color space conversions
 that can be done with \code{\link{changeColorSpace}}.
}
\examples{
data(cc_table)
cc_table

}
\seealso{
\code{\link{Image}}, \code{\link{changeColorSpace}}
}
\keyword{datasets}
