% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{changeBitDepth}
\alias{changeBitDepth}
\title{Convert image bit depth}
\usage{
changeBitDepth(image, bitdepth, scale = 1, in_place = FALSE)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{bitdepth}{A scalar corresponding to the colorspace the image should
be converted to. Options are "8U", "8S", "16U", "16S", "32S", and "32F"
respectively.}

\item{scale}{A scaling factor (default: 1).}

\item{in_place}{A logical indicating whether the change should be applied to
the image itself (TRUE, faster but destructive) or to a copy of it (FALSE,
the default, slower but non destructive).}
}
\value{
An \code{\link{Image}} object if \code{in_place=FALSE}. Otherwise, it
 returns nothing and modifies \code{image} in place.
}
\description{
This function takes an \code{\link{Image}} object and modifies
 its bit depth (e.g. from 16 bits to 8 bits).
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_16 <- changeBitDepth(balloon, "16U")

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
