% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{changeColorSpace}
\alias{changeColorSpace}
\title{Convert image between colorspaces}
\usage{
changeColorSpace(image, colorspace, in_place = FALSE)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{colorspace}{A string corresponding to the colorspace the image should
be converted to. Options are "BGR" (Blue Green Red image), "BGRA" (BGR image
with Alpha channel), and "GRAY" (grayscale image). Converting from a Bayer
mosaic to BGR using the default algorithm is also possible with "BayerBG2BGR",
"BayerGB2BGR", "BayerRG2BGR", or "BayerGR2BGR".}

\item{in_place}{A logical indicating whether the change should be applied to
the image itself (TRUE, faster but destructive) or to a copy of it (FALSE,
the default, slower but non destructive).}
}
\value{
An \code{\link{Image}} object if \code{in_place=FALSE}. Otherwise, it
 returns nothing and modifies \code{image} in place.
}
\description{
This function takes an \code{\link{Image}} object and converts
 it to another colorspace (e.g BGR to grayscale).
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
grey_balloon <- changeColorSpace(balloon, "GRAY")

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
