% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{click}
\alias{click}
\title{Detect Mouse Click on Image Display}
\usage{
click(image, scale = 1, window_name = "Display")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{scale}{The scaling of the display relative to the image size (default: 1).}

\item{window_name}{A character string representing the name of the display
window (default: "Display").}
}
\value{
A data frame with the following 3 columns:
 \itemize{
  \item{x: }{the x coordinate of the mouse click.}
  \item{y: }{the y coordinate of the mouse click.}
  \item{button: }{the mouse button that was pressed (0: left button; 1: right
   button.)}
 }
}
\description{
\code{click} display an image in a new or existing display
 window and waits for a mouse click on the display window to be detected.
}
\examples{
\dontrun{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
click(balloon)
}

}
\seealso{
\code{\link{Image}}, \code{\link{newDisplay}}, \code{\link{display}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
