% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{cloneImage}
\alias{cloneImage}
\title{Make a copy of an Image object}
\usage{
cloneImage(image)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{\link{Image}} objects are pointers toward C++ objects
 stored in memory. When copying an \code{\link{Image}} object using an
 assignment operator, this creates a copy of the pointer, but not a copy of
 the C++ object. Any operation on the copied \code{\link{Image}} object will
 therefore result in a modification of the orginal \code{\link{Image}} object.
 This function duplicates the original \code{\link{Image}} object instead,
 allowing safe operations on it while maintaining the integrity of the
 original \code{\link{Image}} object.
}
\examples{
# TODO

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
