% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{colorspace}
\alias{colorspace}
\title{The Color Space of an Image}
\usage{
colorspace(x)
}
\arguments{
\item{x}{An \code{\link{Image}} object.}
}
\value{
A character string indicating the color space of the image. For now,
 it can only be one of the following:
 \itemize{
  \item BGR: an image with 3 channels, Blue, Green, and Red.
  \item BGRA: an image with 3 channels, Blue, Green, Red, and Alpha
              (transparency).
  \item GRAY: a grayscale image (1 channel only).
 }
}
\description{
This function returns the color space of an \code{\link{Image}}
 object, that is the range of colors of an image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
colorspace(balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{bitdepth}}.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
