% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{compare}
\alias{compare}
\title{Comparison between Images}
\usage{
compare(e1, e2, comparison, target)
}
\arguments{
\item{e1, e2}{Either 2 \code{\link{Image}} objects or 1 \code{\link{Image}}
object and 1 numeric value/vector. If a vector and its length is less than
the number of channels of the image, then it is recycled to match it.}

\item{comparison}{A character string indicating the comparison to be
performed between \code{e1} on the left and \code{e2} on the right. It can
be any of the followings: "==", "!=", ">", "<", ">=", "<=".}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{e1} if it is an
  \code{\link{Image}} object, otherwise into \code{e2} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{e1} or \code{e2} but will replace that of \code{target}.
  Note that if \code{target} does not have the same dimensions, number of
  channels, and bit depth as \code{e1} (if \code{e1} is an \code{\link{Image}}
  object, \code{e2} otherwise), an error will be thrown.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{e1} in place if it is an \code{\link{Image}} object, otherwise it
 modifies \code{e2} in place. If \code{target} is an \code{\link{Image}}
 object, the function returns nothing and modifies that \code{\link{Image}}
 object in place.
}
\description{
Comparison between Images
}
\examples{
balloon1 <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon2 <- image(system.file("sample_img/balloon2.png", package = "Rvision"))
comp <- compare(balloon1, balloon2, ">=")

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
