% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{computeECC}
\alias{computeECC}
\title{Enhanced Correlation Coefficient Value}
\usage{
computeECC(template, image)
}
\arguments{
\item{template}{A grayscale \code{\link{Image}} object.}

\item{image}{A grayscale \code{\link{Image}} object of the same dimensions as
\code{template}.}
}
\value{
A numerical value.
}
\description{
\code{computeECC} computes the Enhanced Correlation Coefficient
 (ECC) value between two images.
}
\examples{
balloon1 <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon2 <- image(system.file("sample_img/balloon2.png", package = "Rvision"))
computeECC(balloon1, balloon2)

}
\references{
Evangelidis, G. D., and Psarakis, E. Z. (2008). Parametric image
 alignment using enhanced correlation coefficient maximization. IEEE Trans.
 Pattern Anal. Mach. Intell. 30, 1858–1865. doi:10.1109/TPAMI.2008.113.
}
\seealso{
\code{\link{findTransformECC}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
