% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate Images}
\usage{
concatenate(image1, image2, direction = "vertical", target = "new")
}
\arguments{
\item{image1, image2}{\code{\link{Image}} objects with the same bitdepth and
number of channels, and either the same width (if \code{direction = "vertical"})
or the same height (if \code{direction = "horizontal"}).}

\item{direction}{A character string indicating the direction of concatenation.
It can be either \code{"vertical"} (the default) or \code{"horizontal"}.}

\item{target}{The location where the results should be stored. It can take 2
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  \code{target} must have the same bitdepth and number of channels as
  \code{image1} and \code{image2}. If \code{direction = "vertical"}, the
  height of \code{target} must be the sum of the heights of \code{image1}
  and \code{image2}. If \code{direction = "horizontal"}, the width of \code{target}
  must be the sum of the widths of \code{image1} and \code{image2}.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target} is an \code{\link{Image}} object, the function
 returns nothing and modifies that \code{\link{Image}} object in place.
}
\description{
\code{concatenate} concatenates two images into one, either
 vertically or horizontally.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
two_balloons <- concatenate(balloon, balloon)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
