% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{contourArea}
\alias{contourArea}
\title{Area of a Contour}
\usage{
contourArea(x, y, oriented = FALSE)
}
\arguments{
\item{x}{A vector of the x coordinates of the contour vertices.}

\item{y}{A vector of the y coordinates of the contour vertices.}

\item{oriented}{A boolean indicating whether to return the oriented area of
the contour or not (default: FALSE).}
}
\value{
If `oriented = FALSE`, the function return the area in pixels enclosed
 within the contour. If `oriented = TRUE`, the function returns a signed area
 value depending on the contour orientation (clockwise or counter-clokwise).
 Using this feature, you can determine the orientation of a contour by taking
 the sign of the area.
}
\description{
\code{polyArea} computes the surface area of a polygon.
}
\note{
The function will certainly return a wrong result for contours with
 self-intersections.
}
\examples{
contourArea(c(0, 1, 1, 0), c(0, 0, 1, 1))

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
