% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{convexHull}
\alias{convexHull}
\title{Compute the Convex Hull of a Set of Points}
\usage{
convexHull(x, y, clockwise = TRUE)
}
\arguments{
\item{x}{A vector of x coordinates.}

\item{y}{A vector of y coordinates of the same lenght as \code{x}.}

\item{clockwise}{If TRUE (the default), the output convex hull is oriented
clockwise. Otherwise, it is oriented counter-clockwise.}
}
\value{
A vector indicating the index of the points on the convex hull.
}
\description{
\code{convexHull} computes the subset of points which lie on the
 convex hull of the set of points specified.
}
\examples{
convexHull(rnorm(100), rnorm(100))

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
