% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{convexityDefects}
\alias{convexityDefects}
\title{Find the Convexity Defects of a Contour}
\usage{
convexityDefects(contours, id = NULL)
}
\arguments{
\item{contours}{A list of two matrices as produced by \code{\link{findContours}}.}

\item{id}{An optional vector indicating the identity of the specific contours
for which to run the function.}
}
\value{
A matrix with 5 columns:
 \itemize{
   \item{"id": }{the contour identity (indicates the set of points belonging
    to the same contour).}
   \item{"start_index": }{index of the first point of the contour belonging to
    a convexity defect.}
   \item{"end_index": }{index of the last point of the contour belonging to
    a convexity defect.}
   \item{"farthest_pt_index": }{index of the point of the contour belonging to
    a convexity defect and that is the farthest away from the convex hull.}
   \item{"fixpt_depth": }{distance between the farthest contour point and the
    convex hull.}
}
}
\description{
\code{convexityDefects} finds the convexity defects of a contour,
 that is area that do not belong to an object but are located inside of its
 convex hull.
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_gray <- changeColorSpace(dots, "GRAY")
dots_bin <- dots_gray < 200
contours <- findContours(dots_bin)
convexityDefects(contours, id = c(3, 5))

}
\seealso{
\code{\link{findContours}}, \code{\link{convexHull}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
