% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib3d.R
\name{cornerSubPix}
\alias{cornerSubPix}
\title{Refine Corner Locations}
\usage{
cornerSubPix(
  image,
  corners,
  win_size = c(11, 11),
  zero_zone = c(-1, -1),
  maxit = 30,
  eps = 1e-04
)
}
\arguments{
\item{image}{An 8-bit (8U), single-channel (GRAY) \code{\link{Image}} object.}

\item{corners}{A matrix of estimated corner locations as returned by
\code{\link{findChessboardCorners}}.}

\item{win_size}{A vector of the half of the side length of the search window.
For example, if \code{win_size = c(5, 5)} , then a (5∗2+1)×(5∗2+1)=11×11
search window is used (default: \code{c(11, 11)}).}

\item{zero_zone}{A vector of the half of the size of the dead region in the
middle of the search zone. It is used sometimes to avoid possible
singularities in the detection algorithm. The value of \code{c(-1, -1)}
indicates that there is no such a size (the default).}

\item{maxit}{The maximum number of iterations of the detection algorithm
(default: 30).}

\item{eps}{The desired accuracy or change in parameters at which the
iterative algorithm stops (default: 0.0001).}
}
\value{
A \code{(pprow * ppcol)}x\code{2} matrix. If the matrix is empty then no
 chessboard with the indicated dimensions is detected.
}
\description{
\code{cornerSubPix} finds the sub-pixel locations of corners or
 radial saddle points in an image from their estimated locations.
}
\examples{
checkerboard <- image(system.file("sample_img/checkerboard6x9.png", package = "Rvision"))
pts <- findChessboardCorners(checkerboard, 6, 9)
checkerboard_gray <- changeColorSpace(checkerboard, "GRAY")
pts_refined <- cornerSubPix(checkerboard_gray, pts)

}
\references{
Förstner, W., & Gülch, E. (1987). A fast operator for detection
 and precise location of distinct points, corners and centres of circular
 features. Proc. ISPRS Intercommission Conference on Fast Processing of
 Photogrammetric Data, 6, 281–305.
}
\seealso{
\code{\link{findChessboardCorners}}, \code{\link{calibrateCamera}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
