% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{countNonZero}
\alias{countNonZero}
\title{Count Non-Zero Pixels}
\usage{
countNonZero(image)
}
\arguments{
\item{image}{A single-channel \code{\link{Image}} object.}
}
\value{
An integer.
}
\description{
\code{countNonZero} returns the number of non-zero pixels in an
 \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
changeColorSpace(balloon, "GRAY", "self")
countNonZero(balloon > 100)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
