% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{destroyDisplay}
\alias{destroyDisplay}
\alias{destroyAllDisplays}
\title{Destroy \code{\link{Image}} Display}
\usage{
destroyDisplay(window_name)
destroyAllDisplays()
}
\arguments{
\item{window_name}{A character string representing the name of the display
window (default: "Display").}
}
\value{
This function does not return anything.
}
\description{
\code{destroyDisplay} closes a specific existing
 \code{\link{Image}} display window. \code{destroyAllDisplays} all existing
 \code{\link{Image}} display window.
}
\examples{
\dontrun{
newDisplay("Test display")
destroyDisplay("Test display")
destroyAllDisplays()
}
}
\seealso{
\code{\link{Image}}, \code{\link{newDisplay}}, \code{\link{display}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
