% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\name{dim.Rcpp_Queue}
\alias{dim.Rcpp_Queue}
\title{Dimensions of a Queue}
\usage{
\method{dim}{Rcpp_Queue}(x)
}
\arguments{
\item{x}{A \code{\link{Queue}} object.}
}
\value{
A vector with 3 values corresponding to the number of rows, columns
 and length of the queue (in this order).
}
\description{
Retrieve the dimensions a \code{\link{Queue}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
buf <- queue(balloon)
dim(buf)

}
\seealso{
\code{\link{Queue}}, \code{\link{queue}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
