% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R
\name{dim.Rcpp_Stream}
\alias{dim.Rcpp_Stream}
\title{Dimensions of a Stream}
\usage{
\method{dim}{Rcpp_Stream}(x)
}
\arguments{
\item{x}{A \code{\link{Stream}} object.}
}
\value{
A vector with 2 values corresponding to the number of rows and columns
 of the stream (in this order).
}
\description{
Retrieve the dimensions a \code{\link{Stream}} object.
}
\examples{
\dontrun{
live <- stream(0)
sim(live)
release(live)
}

}
\seealso{
\code{\link{Stream}}, \code{\link{stream}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
