% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{dim.Rcpp_Video}
\alias{dim.Rcpp_Video}
\title{Dimensions of a Video}
\usage{
\method{dim}{Rcpp_Video}(x)
}
\arguments{
\item{x}{A \code{\link{Video}} object.}
}
\value{
A vector with 3 values corresponding to the number of rows, columns
 and frames of the video (in this order).
}
\description{
Retrieve the dimensions a \code{\link{Video}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
dim(balloon)

}
\seealso{
\code{\link{Video}}, \code{\link{video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
