% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoWriterClass.R
\name{dim.Rcpp_VideoWriter}
\alias{dim.Rcpp_VideoWriter}
\title{Dimensions of a Video Writer}
\usage{
\method{dim}{Rcpp_VideoWriter}(x)
}
\arguments{
\item{x}{A \code{\link{VideoWriter}} object.}
}
\value{
A vector with 2 values corresponding to the number of rows and columns
 of the video writer (in this order).
}
\description{
Retrieve the dimensions a \code{\link{VideoWriter}} object.
}
\examples{
\dontrun{
writer <- videoWriter("test.mp4", "H264", 25, 1080, 1920)
dim(writer)
release(writer)
}

}
\seealso{
\code{\link{VideoWriter}}, \code{\link{videoWriter}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
