% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoStackClass.R
\name{dim.VideoStack}
\alias{dim.VideoStack}
\title{Dimensions of a VideoStack}
\usage{
\method{dim}{VideoStack}(x)
}
\arguments{
\item{x}{A \code{\link{VideoStack}} object.}
}
\value{
A vector with 3 values corresponding to the number of rows, columns
 and frames of the video stack (in this order).
}
\description{
Retrieve the dimensions a \code{\link{VideoStack}} object.
}
\examples{
path <- system.file("sample_vid/Balloon.mp4", package = "Rvision")
balloonStack <- videoStack(path, path)
dim(balloonStack)

}
\seealso{
\code{\link{VideoStack}}, \code{\link{videoStack}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
