% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display \code{\link{Image}} Object}
\usage{
display(
  image,
  window_name = "Display",
  delay = 25,
  height = 480,
  width = 640,
  interpolation = "linear"
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{window_name}{A character string representing the name of the display
window (default: "Display").}

\item{delay}{The delay in milliseconds during which an image is displayed
before it can be replaced by another image.}

\item{height}{An integer representing the height in pixels of the display
window.}

\item{width}{An integer representing the width in pixels of the display
window.}

\item{interpolation}{A character string representing the type of interpolation
to use if the display size is different from the image size (default: "linear").
See notes in \code{\link{resize}} for all accepted interpolation methods.}
}
\value{
This function does not return anything.
}
\description{
\code{display} displays \code{\link{Image}} objects in special
 windows created by \code{\link{newDisplay}} (or creates it if it does not
 exist yet). This function is faster than the generic \code{\link{plot.Image}}
 function for displaying \code{\link{Image}} objects, but cannot be used in
 combination with base R plotting utilities.
}
\examples{
\dontrun{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
display(balloon, height = nrow(balloon), width = ncol(balloon))
}

}
\seealso{
\code{\link{Image}}, \code{\link{newDisplay}}, \code{\link{destroyDisplay}},
 \code{\link{plot.Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
