% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{distanceTransform}
\alias{distanceTransform}
\title{Distance Transform}
\usage{
distanceTransform(image, distance_type = "L1", mask_size = 3)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{distance_type}{A character string indicating the type of distance
to be calculated. It can be any of the following:
\itemize{
 \item{"L1" (the default):}{\code{distance = |x1-x2| + |y1-y2|}.}
 \item{"L2":}{the simple euclidean distance.}
 \item{"C":}{\code{distance = max(|x1-x2|,|y1-y2|)}.}
 \item{"L12":}{L1-L2 metric. \code{distance = 2(sqrt(1+x*x/2) - 1))}.}
 \item{"FAIR":}{\code{distance = c^2(|x|/c-log(1+|x|/c)), c = 1.3998}.}
 \item{"WELSCH":}{\code{distance = c^2/2(1-exp(-(x/c)^2)), c = 2.9846}.}
 \item{"HUBER":}{\code{distance = |x|<c ? x^2/2 : c(|x|-c/2), c=1.345}.}
}}

\item{mask_size}{A numeric value indicating the size of the distance
transform mask. It can be any of the following:
\itemize{
 \item{0:}{used only to indicate the Felzenszwalb algorithm when
  \code{distance_type = "L2"}.}
 \item{3 (the default):}{3x3 mask.}
 \item{5:}{5x5 mask.}
}}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{distanceTransform} calculates the distance to the closest
 zero pixel for each pixel of the source image.
}
\examples{
file <- system.file("sample_img/balloon1.png", package = "Rvision")
balloon <- changeColorSpace(image(file), "GRAY")
bin <- balloon < 200
dst <- distanceTransform(bin)
plot(dst * round(65536 / 255))

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
