% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawCircle}
\alias{drawCircle}
\title{Draw a Circle on an \code{\link{Image}}}
\usage{
drawCircle(image, x, y, radius, color = "red", thickness = 1)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{x}{A numeric value representing the x coordinate of the center of the
circle.}

\item{y}{A numeric value representing the y coordinate of the center of the
circle.}

\item{radius}{A numeric value representing the radius of the circle.}

\item{color}{Any kind of R color specification compatible with
\code{\link{col2rgb}} representing the color of the circle outline
(default: "red").}

\item{thickness}{A numeric value representing the thickness in pixels of the
rectangle outline (default: 1). If negative, then a filled circle is drawn.}
}
\value{
This function does not return anything. It modifies \code{image} in
 place.
}
\description{
\code{drawCirlc} draws a circle over an \code{\link{Image}}
 object. This operation is destructive: it changes irreversibly the
 \code{\link{Image}} object and cannot be undone.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}, \code{\link{drawEllipse}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
