% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawLine}
\alias{drawLine}
\title{Draw a Line on an \code{\link{Image}}}
\usage{
drawLine(image, pt1_x, pt1_y, pt2_x, pt2_y, color = "red", thickness = 1)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{pt1_x}{A numeric value representing the x coordinate of the first point
of the line.}

\item{pt1_y}{A numeric value representing the y coordinate of the first point
of the line.}

\item{pt2_x}{A numeric value representing the x coordinate of the second point
of the line.}

\item{pt2_y}{A numeric value representing the y coordinate of the second point
of the line.}

\item{color}{Any kind of R color specification compatible with
\code{\link{col2rgb}} representing the color of the line (default: "red").}

\item{thickness}{A numeric value representing the thickness in pixels of the
line (default: 1).}
}
\value{
This function does not return anything. It modifies \code{image} in
 place.
}
\description{
\code{drawLine} draws a line over an \code{\link{Image}} object.
 This operation is destructive: it changes irreversibly the \code{\link{Image}}
 object and cannot be undone.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}, \code{\link{drawArrow}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
