% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawPolyline}
\alias{drawPolyline}
\title{Draw Polygonal Lines on an Image}
\usage{
drawPolyline(image, line, closed = FALSE, color = "red", thickness = 1)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{line}{An m x 2 matrix (or an object that can be converted to an
m x 2 matrix) or a list of m x 2 matrices, with the first column containing
the x coordinates of the polygonal lines and the second column containing
the y coordinates of the polygonal lines.}

\item{closed}{A boolean indicating whether the drawn polylines are closed or
not (default: FALSE). If they are closed, the function draws a line from the
last vertex of each curve to its first vertex.}

\item{color}{A value or vector of any kind of R color specification compatible
with \code{\link{col2bgr}} representing the color to fill the polygon with
(default: "white").}

\item{thickness}{A numeric value representing the thickness in pixels of the
line (default: 1).}
}
\value{
This function does not return anything. It modifies \code{image} in
 place.
}
\description{
\code{drawPolyline} draws polygonal lines over an
\code{\link{Image}} object. This operation is destructive: it changes
 irreversibly the \code{\link{Image}} object and cannot be undone.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
poly <- data.frame(x = c(290, 290, 440, 440), y = c(170, 325, 325, 170))
drawPolyline(balloon, poly, closed = FALSE, color = "red", thickness = 3)

}
\seealso{
\code{\link{Image}}, \code{\link{selectROI}}, \code{\link{fillConvexPoly}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
