% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawRotatedRectangle}
\alias{drawRotatedRectangle}
\title{Draw Rotated Rectangles on an \code{\link{Image}}}
\usage{
drawRotatedRectangle(
  image,
  x,
  y,
  axis1,
  axis2,
  angle,
  color = "red",
  thickness = 1
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{x}{A numeric value or vector representing the x coordinates of the
center of each rectangle.}

\item{y}{A numeric value or vector representing the y coordinates of the
center of each rectangle.}

\item{axis1}{A numeric value or vector representing the length of the
first axis of each rectangle.}

\item{axis2}{A numeric value or vector representing the length of the
second axis of each rectangle.}

\item{angle}{A numeric value or vector representing the angle in degrees
between \code{axis1} and the horizontal.}

\item{color}{A value or vector of any kind of R color specification compatible
with \code{\link{col2bgr}} representing the color of each rectangle's outline
(default: "red").}

\item{thickness}{A numeric value or vector representing the thickness in
pixels of each rectangle's outline (default: 1). If negative, then a filled
rectangle is drawn.}
}
\value{
This function does not return anything. It modifies \code{image} in
 place.
}
\description{
\code{drawRotatedRectangle} draws rotated rectangles over an
 \code{\link{Image}} object. This operation is destructive: it changes
 irreversibly the \code{\link{Image}} object and cannot be undone.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
drawRotatedRectangle(balloon, 290, 170, 440, 325, 60, thickness = 3)

}
\seealso{
\code{\link{Image}}, \code{\link{drawRectangle}}, \code{\link{drawEllipse}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
