% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawText}
\alias{drawText}
\title{Draw Text on an \code{\link{Image}}}
\usage{
drawText(
  image,
  text,
  x,
  y,
  font_face = "simplex",
  font_scale = 1,
  italic = FALSE,
  color = "red",
  thickness = 1,
  bl_orig = TRUE
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{text}{A character string representing the text to be drawn.}

\item{x}{A numeric value representing the x coordinate of the bottom left
corner of the text string (or top left if \code{bl_orig == TRUE}).}

\item{y}{A numeric value representing the y coordinate of the bottom left
corner of the text string (or top left if \code{bl_orig == TRUE}).}

\item{font_face}{A character string representing the font type of the text
(default: "simplex"). See notes for a list of available font types.}

\item{font_scale}{A numeric value representing the scale factor by which the
font-specific base size is multiplied (default: 1).}

\item{italic}{A logical specifying whether the text should italicized
(default: FALSE).}

\item{color}{Any kind of R color specification compatible with
\code{\link{col2bgr}} representing the color of the line (default: "red").}

\item{thickness}{A numeric value representing the thickness in pixels of the
line (default: 1).}

\item{bl_orig}{A logical specifying the origin of the text drawing (default:
TRUE). If TRUE, the text is drawn right-side-up. If FALSE, it is drawn
upside-down.}
}
\value{
This function does not return anything. It modifies \code{image} in
 place.
}
\description{
\code{drawText} draws text over an \code{\link{Image}} object.
 This operation is destructive: it changes irreversibly the \code{\link{Image}}
 object and cannot be undone.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
drawText(balloon, "I'm a balloon", 50, 50, thickness = 3)

}
\seealso{
\code{\link{Image}}, \code{\link{getTextSize}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
