% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{extractChannel}
\alias{extractChannel}
\title{Extract Single Channel from Image}
\usage{
extractChannel(image, channel, target = "new")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{channel}{An integer specifying the index of the channel to extract.}

\item{target}{The location where the results should be stored. It can take 2
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  if \code{target} does not have the same dimensions and bit depth as
  \code{image}, an error may be thrown. \code{target} should also be a
  single-channel \code{\link{Image}} object or an error will be thrown.}
}}
}
\value{
If \code{target="new"}, the function returns a single-channel
 \code{\link{Image}} object. If \code{target} is an \code{\link{Image}}
 object, the function returns nothing and modifies that \code{\link{Image}}
 object in place.
}
\description{
\code{extractChannel} extracts a single color channel from the
 source image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
extractChannel(balloon, 2)

}
\seealso{
\code{\link{Image}}, \code{\link{split}}, \code{\link{merge}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
