% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{fillPoly}
\alias{fillPoly}
\title{Fill Polygon with Color in Image}
\usage{
fillPoly(image, polygon, color = "white")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{polygon}{An m x 2 matrix (or an object that can be converted to an
m x 2 matrix), with the first column containing the x coordinates of the
polygon and the second column containing the y coordinates of the polygon.}

\item{color}{A value or vector of any kind of R color specification compatible
with \code{\link{col2rgb}} representing the color to fill the polygon with
(default: "white").}
}
\value{
This function does not return anything. It modifies \code{image} in
 place.
}
\description{
\code{fillPolygon} fills all the pixels of an image withing a
 given polygon with a given color.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}, \code{\link{selectROI}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
