% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{findHomography}
\alias{findHomography}
\title{Perspective Transformation Between Two Planes}
\usage{
findHomography(
  from,
  to,
  from_dim,
  to_dim = from_dim,
  method = "RANSAC",
  ransac_reproj_th = 3,
  max_it = 2000,
  conf = 0.95
)
}
\arguments{
\item{from}{A Nx2 matrix indicating the location (x, y) of N points in the
source image.}

\item{to}{A Nx2 matrix indicating the location (x, y) of N points in the
destination image. The order of the points must correspond to the order in
\code{from}.}

\item{from_dim}{A vector which first two elements indicate the number of rows
and columns of the source image.}

\item{to_dim}{A vector which first two elements indicate the number of rows
and columns of the destination image. If not specified, \code{from_dim} will
be used as a default.}

\item{method}{A character string indicating the method used to compute a
homography matrix. It can be one of the followings: "LS" (least-square),
"RANSAC" (RANSAC-based robust method; the default), "LMEDS" (Least-Median
robust method), or "RHO" (PROSAC-based robust method).}

\item{ransac_reproj_th}{Maximum allowed reprojection error to treat a point
pair as an inlier (used in the RANSAC and RHO methods only). That is, if
`src_x`, `src_y`, `dst_x`, and `dst_y` are measured in pixels, it usually
makes sense to set this parameter somewhere in the range of 1 to 10.}

\item{max_it}{The maximum number of RANSAC iterations.}

\item{conf}{Confidence level, between 0 and 1.}

\item{template}{A grayscale \code{\link{Image}} object.}
}
\value{
A 3x3 homography matrix.
}
\description{
\code{findHomography} computes a perspective transformation
 between two planes.
}
\examples{
src <- cbind(x = c(0, 1, 1, 0), y = c(0, 0, 1, 1))
dst <- src + 1
findHomography(src, dst, c(10, 10))

}
\seealso{
\code{\link{getPerspectiveTransform}}, \code{\link{getAffineTransform}},
 \code{\link{warpPerspective}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
