% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{findNonZero}
\alias{findNonZero}
\title{Find Non-Zero Pixels in an Image}
\usage{
findNonZero(image)
}
\arguments{
\item{image}{An an 8-bit (8U) single-channel \code{\link{Image}} object.}
}
\value{
A matrix with two columns, corresponding to the x (columns) and y
 (rows) coordinates of the non-zero pixels.
}
\description{
\code{findNonZero} retrieves the locations of all non-zero
 pixels in an image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_gray <- changeColorSpace(balloon, "GRAY")
findNonZero(balloon_gray)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
