% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{findTransformECC}
\alias{findTransformECC}
\title{Enhanced Correlation Coefficient-based Geometric Transform}
\usage{
findTransformECC(
  template,
  image,
  warp_matrix = NULL,
  warp_mode = "affine",
  max_it = 200,
  epsilon = 0.001,
  filt_size = 0
)
}
\arguments{
\item{template}{A grayscale \code{\link{Image}} object.}

\item{image}{A grayscale \code{\link{Image}} object of the same dimensions as
\code{template}.}

\item{warp_matrix}{An initial mapping (warp) matrix. It must be a 3x3 matrix
when \code{warp_mode} is set to "homography", a 2x3 matrix otherwise. If set
to \code{NULL} (the default), it will be automatically initialized as an
identity matrix with the appropriate dimensions.}

\item{warp_mode}{A character string indicating the type of warping required
to transform \code{image} into \code{template}. It can be any of the following:
\itemize{
 \item{"translation":}{simple translational transformation.}
 \item{"euclidean":}{Euclidean (rigid) transformation (translation + rotation).}
 \item{"affine" (default):}{affine transformation (Euclidean + shear; this
  transformation will preserve parallelism between lines).}
 \item{"homography":}{homography transformation (affine + perspective; this
  transformation does not preserve parallelism between lines).}
}}

\item{max_it}{The maximum number of iterations (default: 200).}

\item{epsilon}{The convergene tolerance (default: 1e-3).}

\item{filt_size}{The size in pixels of a gaussian blur filter applied to the
images before computation of the transform. When set to 0 (the default), no
filtering is applied.}
}
\value{
A 2x3 or 3x3 (if \code{warp_mode = "homography"}) matrix.
}
\description{
\code{findTransformECC} computes the geometric transform between
 two images in terms of the Enhanced Correlation Coefficient criterion.
}
\examples{
file1 <- system.file("sample_img/balloon1.png", package = "Rvision")
file2 <- system.file("sample_img/balloon2.png", package = "Rvision")
balloon1 <- changeColorSpace(image(file1), "GRAY")
balloon2 <- changeColorSpace(image(file2), "GRAY")
findTransformECC(balloon1, balloon2)

}
\references{
Evangelidis, G. D., and Psarakis, E. Z. (2008). Parametric image
 alignment using enhanced correlation coefficient maximization. IEEE Trans.
 Pattern Anal. Mach. Intell. 30, 1858–1865. doi:10.1109/TPAMI.2008.113.
}
\seealso{
\code{\link{computeECC}}, \code{\link{findTransformORB}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
