% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{findTransformORB}
\alias{findTransformORB}
\title{ORB-based Geometric Transform}
\usage{
findTransformORB(
  template,
  image,
  warp_mode = "affine",
  max_features = 500,
  descriptor_matcher = "BruteForce-Hamming",
  match_frac = 0.15,
  homography_method = "RANSAC"
)
}
\arguments{
\item{template}{A grayscale \code{\link{Image}} object.}

\item{image}{A grayscale \code{\link{Image}} object of the same dimensions as
\code{template}.}

\item{warp_mode}{A character string indicating the type of warping required
to transform \code{image} into \code{template}. It can be any of the following:
\itemize{
 \item{"affine" (default):}{affine transformation (Euclidean + shear; this
  transformation will preserve parallelism between lines).}
 \item{"homography":}{homography transformation (affine + perspective; this
  transformation does not preserve parallelism between lines).}
}}

\item{max_features}{The maximum number of features to extract (default: 500).}

\item{descriptor_matcher}{A character string indicating the type of the
descriptor matcher to use. It can be one of the followings: "BruteForce",
"BruteForce-L1", "BruteForce-Hamming" (the default), or
"BruteForce-Hamming(2)".}

\item{match_frac}{The fraction of top matches to keep (default: 0.15).}

\item{homography_method}{A character string indicating the method used to
compute a homography matrix. It can be one of the followings: "LS"
(least-square), "RANSAC" (RANSAC-based robust method; the default), "LMEDS"
(Least-Median robust method), or "RHO" (PROSAC-based robust method).}
}
\value{
A 2x3 or 3x3 (if \code{warp_mode = "homography"}) matrix.
}
\description{
\code{findTransformORB} computes the geometric transform between
 two images in terms of the ORB feature detector.
}
\examples{
file1 <- system.file("sample_img/balloon1.png", package = "Rvision")
file2 <- system.file("sample_img/balloon2.png", package = "Rvision")
balloon1 <- changeColorSpace(image(file1), "GRAY")
balloon2 <- changeColorSpace(image(file2), "GRAY")
findTransformORB(balloon1, balloon2)

}
\references{
Evangelidis, G. D., and Psarakis, E. Z. (2008). Parametric image
 alignment using enhanced correlation coefficient maximization. IEEE Trans.
 Pattern Anal. Mach. Intell. 30, 1858–1865. doi:10.1109/TPAMI.2008.113.
}
\seealso{
\code{\link{findTransformECC}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
