% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{fitEllipse}
\alias{fitEllipse}
\title{Fit an Ellipse Around a Set of 2D Points}
\usage{
fitEllipse(x, y, method = "original")
}
\arguments{
\item{x}{A vector of x coordinates.}

\item{y}{A vector of y coordinates of the same lenght as \code{x}.}

\item{method}{A character string indicating the method to use in order to fit
the ellipse. It can take the following values:
\itemize{
   \item{'original': }{least square.}
   \item{'ams': }{Approximate Mean Square (AMS) proposed in Taubin (1991).}
   \item{'direct': }{Direct least square method proposed in Fitzgibbon, Pilu,
    and Fisher (1999).}
}}
}
\value{
A list containing the height and width (in pixels) of the ellipse,
 the angle (in degrees) of its main axis with respect to the x axis, and the
 x and y coordinates of its center.
}
\description{
\code{fitEllipse} calculates the ellipse that fits a set of 2D
 points.
}
\examples{
fitEllipse(rnorm(100), rnorm(100))

}
\references{
Taubin G. Estimation of planar curves, surfaces, and nonplanar
 space curves defined by implicit equations with applications to edge and
 range image segmentation. IEEE Trans Pattern Anal Mach Intell. 1991;13:
 1115–1138. doi:10.1109/34.103273

Fitzgibbon A, Pilu M, Fisher RB. Direct least square fitting of
 ellipses. IEEE Trans Pattern Anal Mach Intell. 1999;21: 476–480.
 doi:10.1109/34.765658
}
\seealso{
\code{\link{minAreaRect}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
