% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{flip}
\alias{flip}
\title{Flip an \code{\link{Image}}}
\usage{
flip(image, type = 0, in_place = FALSE)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{type}{An integer indicating the type of flipping to be performed. If
\code{type = 0} (the default), the image is flipped around its x-axis; if
\code{type = 1} (or any positive value, then it is flipped around its y-axis;
finally, if \code{type = -1} (or any negative value, then it is flipped
around both axes.)}

\item{in_place}{A logical indicating whether the change should be applied to
the image itself (TRUE, faster but destructive) or to a copy of it (FALSE,
the default, slower but non destructive).}
}
\value{
An \code{\link{Image}} object if \code{in_place=FALSE}. Otherwise, it
 returns nothing and modifies \code{image} in place.
}
\description{
\code{flip} returns a flipped version of an \code{\link{Image}}
 around one or both of its axes.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_flipped <- flip(balloon, -1)
plot(balloon_flipped)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
