% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{floodFill}
\alias{floodFill}
\title{Fills a Connected Component with a Given Color.}
\usage{
floodFill(
  image,
  seed = c(1, 1),
  color = "white",
  lo_diff = 0,
  up_diff = 0,
  connectivity = 4
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{seed}{A 2-element vector indicating the x and y coordinates of the seed
point from where to start the filling.}

\item{color}{A value or vector of any kind of R color specification
compatible with \code{\link{col2bgr}} representing the color of the border
(default: "white").}

\item{lo_diff}{Maximal lower brightness/color difference between the
currently observed pixel and one of its neighbors belonging to the component,
or a seed pixel being added to the component (see Details). It can be a
single value or a vector of the same length as the number \code{n} of
channels in \code{image}. If it is shorter, its elements will be recycled.
If it has more, only the first \code{n} elements will be used.}

\item{up_diff}{Maximal upper brightness/color difference between the
currently observed pixel and one of its neighbors belonging to the component,
or a seed pixel being added to the component (see Details). It can be a
single value or a vector of the same length as the number \code{n} of
channels in \code{image}. If it is shorter, its elements will be recycled.
If it has more, only the first \code{n} elements will be used.}

\item{connectivity}{The connetivity neighborhood to decide whether 2 pixels
are contiguous. This parameter can take two values:
\itemize{
 \item{4: }{the neighborhood of a pixel are the four pixels located above
  (north), below (south), to the left (west) and right (east) of the pixel.}
 \item{8 (the default): }{the neighborhood of a pixel includes the four
  4-neighbors and the four pixels along the diagonal directions (northeast,
  northwest, southeast, and southwest).}
}}
}
\value{
This function returns the number of pixels that were filled and
 modifies \code{image} in place.
}
\description{
\code{floodFill} fills a connected component starting from a
 seed point with a specified color.
}
\details{
The connectivity is determined by the color/brightness closeness of
 the neighbor pixels. The pixel at (x,y) is considered to belong to the
 repainted domain if:
 \itemize{
  \item{in case of a floating range:
   \itemize{
    \item{\code{image[x',y'] - lo_diff <= image[x,y] <= image[x',y'] + up_diff}}
   }
  }
  \item{in case of a fixed range:
   \itemize{
    \item{\code{image[seed$x,seed$y] − lo_diff <= image[x,y] <= image(seed$x,seed$y) + up_diff }}
   }
  }
 }
 where image[x′,y′] is the value of one of pixel neighbors that is already
 known to belong to the component. That is, to be added to the connected
 component, a color/brightness of the pixel should be close enough to:
 \itemize{
  \item{Color/brightness of one of its neighbors that already belong to the
   connected component in case of a floating range.}
  \item{Color/brightness of the seed point in case of a fixed range.}
 }
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
floodFill(dots, color = "green")

}
\seealso{
\code{\link{connectedComponents}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
