% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{fps}
\alias{fps}
\title{Framerate of a Video}
\usage{
fps(video)
}
\arguments{
\item{video}{A \code{\link{Video}} object.}
}
\value{
A numeric value.
}
\description{
Retrieve the framerate (in frames per second) of a
 \code{\link{Video}} object.
}
\examples{
# TODO

}
\seealso{
\code{\link{Video}}, \code{\link{video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
