% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\name{full}
\alias{full}
\alias{empty}
\title{Test for a Full/Empty Queue Object}
\usage{
full(x)

empty(x)
}
\arguments{
\item{x}{A \code{\link{Queue}} object.}
}
\value{
A logical.
}
\description{
\code{full} tests whether a \code{\link{Queue}} object is full,
 that is whether the number of frames it contains is equal to its capacity.
 \code{empty} tests whether a \code{\link{Queue}} object is empty, that is
 whether it contains no frames.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
buf <- queue(balloon)
full(buf)

}
\seealso{
\code{\link{Queue}}, \code{\link{queue}}, \code{\link{capacity}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
