% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{gaussianBlur}
\alias{gaussianBlur}
\title{Blurs an Image Using a Gaussian Filter}
\usage{
gaussianBlur(
  image,
  k_height = 5,
  k_width = 5,
  sigma_x = 1,
  sigma_y = 1,
  in_place = FALSE
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{k_height}{The half-height in pixels of the kernel (default: 5).}

\item{k_width}{The half-width in pixels of the kernel (default: 5).}

\item{sigma_x}{The standard deviation of the kernel along the x axis
(default: 1).}

\item{sigma_y}{The standard deviation of the kernel along the y axis
(default: 1).}

\item{in_place}{A logical indicating whether the change should be applied to
the image itself (TRUE, faster but destructive) or to a copy of it (FALSE,
the default, slower but non destructive).}
}
\value{
An \code{\link{Image}} object if \code{in_place=FALSE}. Otherwise, it
 returns nothing and modifies \code{image} in place.
}
\description{
\code{gaussianBlur} convolves the source image with the
 specified Gaussian kernel. The result is a blurred version of the source
 image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_blur <- gaussianBlur(balloon, 11, 11, 5, 5)
plot(balloon_blur)

}
\seealso{
\code{\link{Image}}, \code{\link{boxFilter}}, \code{\link{blur}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
