% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{getGaborKernel}
\alias{getGaborKernel}
\title{Gabor Filter Kernels}
\usage{
getGaborKernel(
  width = 31,
  height = 31,
  sigma = 5,
  theta = pi,
  lambda = 31,
  gamma = 5,
  psi = 0
)
}
\arguments{
\item{width}{The width in pixels of the kernel (default: 31).}

\item{height}{The height in pixels of the kernel (default: 31).}

\item{sigma}{The standard deviation of the Gaussian envelope (default: 5).}

\item{theta}{The orientation of the normal to the parallel stripes of a Gabor
function (default: \code{\link{pi}}).}

\item{lambda}{The wavelength of the sinusoidal factor (default: 31).}

\item{gamma}{The spatial aspect ratio (default: 5).}

\item{psi}{A phase offset (default: 0).}
}
\value{
A matrix of Gabor coefficients.
}
\description{
\code{getGaborKernel} is a convenience function to create Gabor
 kernels that can be used to filter images with \code{\link{filter2D}}.
}
\note{
For more details about Gabor filter equations and parameters,
 see: \href{https://en.wikipedia.org/wiki/Gabor_filter}{Gabor filter}.
}
\examples{
getGaborKernel()

}
\seealso{
\code{\link{filter2D}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
