% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib3d.R
\name{getOptimalNewCameraMatrix}
\alias{getOptimalNewCameraMatrix}
\title{Camera Matrix Correction Using the Free Scaling Parameter}
\usage{
getOptimalNewCameraMatrix(
  camera_matrix,
  dist_coefs,
  nrow,
  ncol,
  alpha = 0,
  center_principal_point = FALSE
)
}
\arguments{
\item{camera_matrix}{A 3x3 camera intrinsic matrix as returned by
\code{\link{calibrateCamera}}.}

\item{dist_coefs}{A single row matrix with 4, 5, 8, 12 or 14 elements as
returned by \code{\link{calibrateCamera}}.}

\item{nrow, ncol}{The number of rows and columns of the image to undistort.}

\item{alpha}{A numeric value corresponding to the free scaling parameter
between 0 (only valid pixels in the the source image are retained in the
undistorted image; the default) and 1 (all the source image pixels are
retained in the undistorted image).}

\item{center_principal_point}{A boolean that indicates whether in the new
camera intrinsic matrix the principal point should be at the image center or
not (the default). The principal point is chosen to best fit a subset of the
source image (determined by alpha) to the corrected image.}
}
\value{
A list:
 \itemize{
  \item{\code{camera_matrix}:}{the new 3x3 camera intrinsic matrix.}
  \item{\code{roi:}}{a 4-element list defining a rectangle that outlines
   the all-valid-pixels region in the undistorted image.}
 }
}
\description{
\code{getOptimalNewCameraMatrix} computes and returns an optimal
 new camera intrinsic matrix based on the free scaling parameter \code{alpha}.
 By varying this parameter, you may retrieve only sensible pixels
 \code{alpha = 0} , keep all the original image pixels if there is valuable
 information in the corners \code{alpha = 1}, or get something in between.
 When \code{alpha > 0}, the undistorted result is likely to have some black
 pixels corresponding to "virtual" pixels outside of the captured distorted
 image.
}
\examples{
# See the help vignette:
\dontrun{ vignette("", package = "Rvision") }

}
\seealso{
\code{\link{calibrateCamera}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
