% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{getRotationMatrix2D}
\alias{getRotationMatrix2D}
\title{Affine Matrix of 2D Rotation}
\usage{
getRotationMatrix2D(center, angle = 90, scale = 1)
}
\arguments{
\item{center}{A 2-elements vector indicating the location (row, column) of
the center of the rotation in the source image.}

\item{angle}{A numeric value indicating the rotation angle in degrees
(default: 90).}

\item{scale}{A numeric value indicating an isotropic scale factor (default: 1).}
}
\value{
A 2x3 matrix.
}
\description{
\code{getRotationMatrix2D} computes the affine matrix for the
 rotation of a 2D image.
}
\examples{
getRotationMatrix2D(c(50, 50), 45, 1)

}
\seealso{
\code{\link{rotateScale}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
