% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{getStructuringElement}
\alias{getStructuringElement}
\title{Structuring Elements}
\usage{
getStructuringElement(
  k_shape = "rectangle",
  k_height = 5,
  k_width = 5,
  anchor = c(-1, -1)
)
}
\arguments{
\item{k_shape}{A string corresponding to the shape of the structuring element.
Valid kernel shapes are:
\itemize{
  \item{"rectangle" (the default)}
  \item{"cross"}
  \item{"ellipse"}
}}

\item{k_height}{The half-height in pixels of the structuring element.}

\item{k_width}{The half-width in pixels of the structuring element.}

\item{anchor}{A 2-element numeric vector defining the anchor position within
the cross-shaped structuring element. The default value (−1,−1) means that
the anchor is at the center. Note that only the shape of a cross-shaped
kernel depends on the anchor position.}
}
\value{
A matrix of 0s and 1s.
}
\description{
\code{getStructuringElement} is a convenience function to create
 structuring elements with defined shapes that can be used to filter images
 with \code{\link{filter2D}}.
}
\examples{
getStructuringElement()

}
\seealso{
\code{\link{filter2D}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
