% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{getTextSize}
\alias{getTextSize}
\title{Calculate the Height and Width of a Text String}
\usage{
getTextSize(
  text,
  font_face = "simplex",
  font_scale = 1,
  italic = FALSE,
  thickness = 1
)
}
\arguments{
\item{text}{A character string representing the text to be drawn.}

\item{font_face}{A character string representing the font type of the text
(default: "simplex"). See notes for a list of available font types.}

\item{font_scale}{A numeric value representing the scale factor by which the
font-specific base size is multiplied (default: 1).}

\item{italic}{A logical specifying whether the text should italicized
(default: FALSE).}

\item{thickness}{A numeric value representing the thickness in pixels of the
line (default: 1).}
}
\value{
A two-element vector corresponding to the height and width of the
 text string.
}
\description{
\code{getTextSize} calculates the size of a box that contains
 the specified text string, the tight box surrounding it, and the baseline.
}
\examples{
getTextSize("I'm a balloon")

}
\seealso{
\code{\link{Image}}, \code{\link{drawText}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
