% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{grabCut}
\alias{grabCut}
\title{Segmentation with GrabCut Algorithm}
\usage{
grabCut(
  image,
  mask,
  rect = rep(1, 4),
  bgdModel,
  fgdModel,
  iter = 1,
  mode = "EVAL"
)
}
\arguments{
\item{image}{An 8-bit (8U), 3-channel \code{\link{Image}} object to segment.}

\item{mask}{An 8-bit (8U), single-channel \code{\link{Image}} object. Each
pixel can take any of the following 4 values:
\describe{
   \item{0: }{an obvious background pixels.}
   \item{1: }{an obvious foreground (object) pixel.}
   \item{2: }{a possible background pixel.}
   \item{3: }{a possible foreground pixel.}
}}

\item{rect}{A vector defining the region of interest containing a segmented
object. The pixels outside of the region of interest are marked as "obvious
background". \code{rect} must be a 4-element numeric vector which elements
correspond to - in this order - the x and y coordinates of the bottom left
corner of the region of interest, and to its width and height. The parameter
is only used when \code{mode="RECT"} (default: rep(1, 4)).}

\item{bgdModel}{A 1x65, single-channel, 64-bit (64F) \code{\link{Image}}
object to set and store the parameters of the background model.}

\item{fgdModel}{A 1x65, single-channel, 64-bit (64F) \code{\link{Image}}
object to set and store the parameters of the foreground model.}

\item{iter}{Number of iterations (default: 1) the algorithm should make
before returning the result. Note that the result can be refined with
further calls with \code{mode="MASK"} or \code{mode="MASK"}.}

\item{mode}{A character string indicating the operation mode of the function.
It can be any of the following:
\describe{
   \item{"RECT": }{The function initializes the state and the mask using the
    provided \code{rect}. After that it runs \code{iter} iterations of the
    algorithm.}
   \item{"MASK":}{The function initializes the state using the provided
    \code{mask}.}
   \item{"EVAL":}{The value means that the function should just resume.}
   \item{"FREEZE":}{The value means that the function should just run the
    grabCut algorithm (a single iteration) with the fixed model.}
}}
}
\value{
This function returns nothing. It modifies in place \code{mask},
 \code{bgdModel}, and \code{fgdModel}.
}
\description{
\code{grabCut} performs image segmentation (i.e., partition of
 the image into coherent regions) using the GrabCut method.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
mask <- zeros(nrow(balloon), ncol(balloon), 1)
bgdModel <- zeros(1, 65, 1, "64F")
fgdModel <- zeros(1, 65, 1, "64F")
grabCut(balloon, mask, c(290, 170, 160, 160), bgdModel, fgdModel, iter = 5, mode = "RECT")

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
