% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{histEq}
\alias{histEq}
\title{Histogram Equalization}
\usage{
histEq(image, target = "new")
}
\arguments{
\item{image}{An \code{\link{Image}} object to transform.}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{image} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  if \code{target} does not have the same dimensions, number of channels,
  and bit depth as \code{image}, nothing will be stored.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{image} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
\code{histEq} performs the histogram equalization of an image.
 The function equalizes the histogram of the input image using the following
 algorithm:
 \itemize{
  \item{Calculate the histogram of the image.}
  \item{Normalize the histogram so that the sum of histogram bins is 255.}
  \item{Compute the integral of the histogram.}
  \item{Transform the image using the integral of the histogram as a look-up
   table.}
 }
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_eq <- histEq(balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{histmatch}}, \code{\link{LUT}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
