% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{histmatch}
\alias{histmatch}
\title{Histogram Matching/Specification}
\usage{
histmatch(image, reference, in_place = FALSE)
}
\arguments{
\item{image}{An \code{\link{Image}} object to transform.}

\item{reference}{An \code{\link{Image}} object which histogram will be used
as a reference to transform \code{image}.}

\item{in_place}{logical indicating whether the change should be applied to
the image itself (TRUE, faster but destructive) or to a copy of it (FALSE,
the default, slower but non destructive).}
}
\value{
An \code{\link{Image}} object if \code{in_place=FALSE}. Otherwise, it
 returns nothing and modifies \code{image} in place.
}
\description{
\code{histmatch} transforms an \code{\link{Image}} object so
 that its histogram matches (approximately) that of another
 \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_matched <- histmatch(dots, balloon)
plot(dots_matched)

}
\seealso{
\code{\link{Image}}, \code{\link{histmatch}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
