% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{houghLinesP}
\alias{houghLinesP}
\title{Probabilistic Hough Line Detector}
\usage{
houghLinesP(
  image,
  rho,
  theta,
  threshold,
  min_line_length = 0,
  max_line_gap = 0
)
}
\arguments{
\item{image}{An 8-bit (8U) single-channel (grayscale) binary
\code{\link{Image}} object.}

\item{rho}{The distance resolution of the accumulator in pixels.}

\item{theta}{The angle resolution of the accumulator in radians.}

\item{threshold}{The accumulator threshold parameter. Only lines that get
more votes than that threshold are returned.}

\item{min_line_length}{The minimum line length. Line segments shorter than
that are rejected.}

\item{max_line_gap}{The maximum allowed gap between points on the same line
segment to link them.}
}
\value{
A matrix with 4 columns corresponding x and y coordinates of the
 extremities of each detected line.
}
\description{
\code{houghLinesP} finds line segments in a binary image using
 the probabilistic Hough transform.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_canny <- canny(balloon, 50, 50)
lines <- houghLinesP(balloon_canny, 1, pi / 180, 80, 0, 50)

}
