% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\docType{class}
\name{Image-class}
\alias{Image-class}
\alias{image}
\alias{Rcpp_Image}
\alias{Image}
\title{An S4 Class Containing an OpenCV Image}
\usage{
image(..., colorspace = "BGR")
}
\arguments{
\item{...}{When created from an image file, \code{image} takes one argument
that is a character string indicating the path to the image file. When
created from an array (e.g. a matrix), it takes this array as its single
argument. An \code{Image} object can also be created without any argument,
in which case it is empty and can be populated with an image later.}

\item{colorspace}{A string indicating the desired color space for the image.}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{Image} objects are the base objects of the
 \pkg{\link{Rvision}} package. They contain an
 \href{http://opencv.org/}{OpenCV} image that can originate from an image
 file, an array, a video file or a video stream. This image can be
 manipulated using the functions of \pkg{\link{Rvision}}.

Function for creating \code{\link{Image}} objects from arrays
 and image files.
}
\section{Slots}{

\describe{
\item{\code{depth}}{A function returning the bit depth of the object.}

\item{\code{dim,ncol,nrow,nchan}}{Functions returning the dimensions of the object.}

\item{\code{space}}{A character string indicating the colorspace of the object.}

\item{\code{pget,pset}}{Functions to get and set single pixel values.}

\item{\code{toR}}{A function to convert the object to an R array.}

\item{\code{fromGPU}}{A function retrieving the object from GPU memory.}

\item{\code{toGPU}}{A function sending the object to GPU memory.}

\item{\code{GPU}}{A boolean indicating whether the object is in GPU memory.}

\item{\code{write}}{A function to write the object to a file.}
}}

\note{
\code{Image} objects can be created from video files and video streams
 using the following functions: \code{\link{video}}, \code{\link{stream}}.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
noise <- image(array(sample(0:255, 100 * 100 * 3, replace = TRUE), dim = c(100, 100, 3)))

}
\seealso{
\code{\link{image}}

\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
