% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{imageMinMax}
\alias{imageMinMax}
\alias{bitMin}
\alias{bitMax}
\title{Element-Wise Minimums and Maximums}
\usage{
bitMin(e1, e2, target)

bitMax(e1, e2, target)
}
\arguments{
\item{e1, e2}{Either 2 \code{\link{Image}} objects or 1 \code{\link{Image}}
object and 1 numeric value/vector. If a vector and its length is less than
the number of channels of the image, then it is recycled to match it.}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{e1} if it is an
  \code{\link{Image}} object, otherwise into \code{e2} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{e1} or \code{e2} but will replace that of \code{target}.
  Note that if \code{target} does not have the same dimensions, number of
  channels, and bit depth as \code{e1} (if \code{e1} is an \code{\link{Image}}
  object, \code{e2} otherwise), an error will be thrown.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{e1} in place if it is an \code{\link{Image}} object, otherwise it
 modifies \code{e2} in place. If \code{target} is an \code{\link{Image}}
 object, the function returns nothing and modifies that \code{\link{Image}}
 object in place.
}
\description{
These functions calculates the per-element minimum or maximum of
 2 \code{\link{Image}} objects or of 1 \code{\link{Image}} object and 1
 numeric value/vector.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
bitMax(balloon, c(0, 0, 127), "self")
bitMin(balloon, c(127, 255, 255), "self")

}
\seealso{
\code{\link{Image}}, \code{\link{min.Rcpp_Image}}, \code{\link{max.Rcpp_Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
