% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{imhist}
\alias{imhist}
\title{Multichannel Histogram of an Image}
\usage{
imhist(
  image,
  nbins = 256,
  range = c(0, 255),
  mask = NULL,
  plot = FALSE,
  col = c("blue", "green", "red", "black"),
  xlab = "Pixel value",
  ylab = "Counts",
  lty = 1,
  ...
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{nbins}{An integer indicating the number of bins of the histogram
(default: 256).}

\item{range}{The range of pixel values to include in the histogram (default:
c(0, 256)). The lower boundary is inclusive but the higher one is exclusive.}

\item{mask}{A single-channel (GRAY) 8-bit (8U) \code{\link{Image}} object
with the same dimensions as \code{image}. This can be used to mask out
pixels that should not be considered when calculating the histogram (pixels
set to 0 in the mask will be ignored during the histogram calculation).}

\item{plot}{A logical indicating whether to plot the histogram (default:
FALSE).}

\item{col}{A value or vector of any kind of R color specification
compatible with \code{\link{col2rgb}} representing the color of the
histogram for each image channel (default:
c("blue", "green", "red", "black")).}

\item{xlab, ylab}{Character strings for the axis labels (default: "Pixel
value" for the x axis and "Counts" for the y axis).}

\item{lty}{A vector of line types, see \code{\link{par}}.}

\item{...}{Further arguments passed to \code{\link{matplot}}.}
}
\value{
If \code{plot=FALSE}, the function returns a \code{m x n} matrix,
 with \code{m = nbins} and \code{n} equal to the number of channels in the
 image + 1. The first column corresponds to the bin values. If
 \code{plot=TRUE}, the function plots the histogram and returned the
 aforementioned matrix silently.
}
\description{
\code{imhist} computes (or plots) the multichannel histogram of
 an \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
imhist(balloon, plot = TRUE)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
