% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{improfile}
\alias{improfile}
\title{Pixel Values Along a Line Segment}
\usage{
improfile(image, xi, yi, connectivity = 8, left_to_right = FALSE)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{xi, yi}{Two-element vectors containing the x and y coordinates of the
start and end points of a line segment. The points can lie outside of the
image boundaries but then the line will be clipped on the image boundaries.}

\item{connectivity}{The connectivity neighborhood to decide whether 2 pixels
are contiguous. This parameter can take two values:
\itemize{
 \item{4: }{the neighborhood of a pixel are the four pixels located above
  (north), below (south), to the left (west) and right (east) of the pixel.}
 \item{8 (the default): }{the neighborhood of a pixel includes the four
  4-neighbors and the four pixels along the diagonal directions (northeast,
  northwest, southeast, and southwest).}
}}

\item{left_to_right}{If `TRUE`, the line segment is traversed from the
leftmost endpoint to the rightmost endpoint, regardless of the order in
which they are specified in \code{xi} and \code{yi}. Otherwise (the default),
the line is traversed in the order in which the points are specified in
\code{xi} and \code{yi}.}
}
\value{
A matrix in which the first two columns indicate the coordinates of
 the points traversed by the line segment, and the other columns indicate the
 color values at each location.
}
\description{
\code{improfile} finds all pixels intersected by a line segment
 and returns their coordinates and color values.
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
improfile(dots, c(1, ncol(dots)), c(nrow(dots) / 2, nrow(dots) / 2))

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
