% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{inpaint}
\alias{inpaint}
\title{Reconstruct Image Region from Region Neighborhood}
\usage{
inpaint(image, mask, radius = 5, method = "NS")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{mask}{An 8-bit single-channel \code{\link{Image}} object. The region to
be reconstructed should be white.}

\item{radius}{Radius of the circular neighborhood of each point inpainted
that is considered by the algorithm (default: 5).}

\item{method}{The inpainting method to be used. It can only be one of the
following:
\itemize{
  \item{"NS": }{Navier-Stokes based method (the default).}
  \item{"Telea": }{Alexandru Telea's method.}
}}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{inpaint} reconstructs the selected image area from the
 pixel near the area boundary. The function may be used to remove dust and
 scratches from a scanned photo, or to remove undesirable objects from still
 images or videos.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
mask <- image(array(0, dim = c(nrow(balloon), ncol(balloon), 3)))
poly <- data.frame(x = c(290, 290, 440, 440), y = c(170, 325, 325, 170))
fillPoly(mask, poly, color = "white")
mask <- changeColorSpace(mask, "GRAY")
balloon_inpait <- inpaint(balloon, mask)
plot(balloon_inpait)

}
\references{
Telea, A. (2004). An image inpainting technique based on the fast
 marching method. Journal of graphics tools. doi: 10.1080/10867651.2004.10487596.
}
\seealso{
\code{\link{Image}}, \code{\link{selectROI}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
