% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R, R/zzz.R
\name{inPlaceComparison}
\alias{inPlaceComparison}
\alias{\%i>\%}
\alias{\%i<\%}
\alias{\%i>=\%}
\alias{\%i<=\%}
\alias{\%i==\%}
\alias{\%i!=\%}
\title{In Place Comparison Operators for Images}
\usage{
e1 \%i>\% e2

e1 \%i<\% e2

e1 \%i>=\% e2

e1 \%i<=\% e2

e1 \%i==\% e2

e1 \%i!=\% e2
}
\arguments{
\item{e1, e2}{Either 2 \code{\link{Image}} objects or 1 \code{\link{Image}}
object and 1 numeric value/vector. If a vector and its length is less than
the number of channels of the image, then it is recycled to match it.}
}
\value{
These operators do not return anything. They modify the image in
 place (destructive operation). If 2 images are passed to the operators, only
 the one of the left side of the operator is modified; the other is left
 untouched.
}
\description{
In Place Comparison Operators for Images
}
\examples{
balloon1 <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon2 <- image(system.file("sample_img/balloon2.png", package = "Rvision"))
balloon1 \%i>\% balloon2

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
