% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{insertChannel}
\alias{insertChannel}
\title{Insert Single Channel into Image}
\usage{
insertChannel(image, channel, insert)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{channel}{An integer specifying the index of the channel that will be
replaced by \code{insert}}

\item{insert}{A single-channel \code{\link{Image}} object with the same bit
depth as \code{image}.}
}
\value{
The function returns nothing and modifies \code{image} in place.
}
\description{
\code{insertChannel} insert a single color channel into the
 target image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
red <- extractChannel(balloon, 3)
insertChannel(balloon, 1, red)

}
\seealso{
\code{\link{Image}}, \code{\link{split}}, \code{\link{merge}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
