% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{invert}
\alias{invert}
\title{Invert Colors}
\usage{
invert(image)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{invert} returns an image which colors are the linear
 inverse of that of the original image.
}
\examples{
path_to_image <- system.file("sample_img", "bunny.png", package = "Rvision")
my_image <- image(filename = path_to_image)
plot(invert(my_image))

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
