% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoWriterClass.R
\name{invertFourcc}
\alias{invertFourcc}
\title{FOURCC Code to Codec Name}
\usage{
invertFourcc(x)
}
\arguments{
\item{x}{A integer.}
}
\value{
A 4-character strings corresponding to the video codec.
}
\description{
\code{invertFourcc} translates a
 \href{https://www.fourcc.org/codecs.php}{FOURCC} code into the 4-character
 name of the corresponding video codec.
}
\examples{
invertFourcc(1684633208)

}
\seealso{
\code{\link{VideoWriter}}, \code{\link{videoWriter}},
 \code{\link{codec}}, \code{\link{fourcc}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
