% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{isBlob}
\alias{isBlob}
\title{Test for a Blob Object}
\usage{
isBlob(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
A logical indicating whether the object is of class
 \code{blob} (TRUE) or not (FALSE).
}
\description{
Tests whether the object is of class \code{blob}.
}
\examples{
# TODO
}
\seealso{
\code{\link{simpleBlobDetector}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
