% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{isImage}
\alias{isImage}
\title{Test for an Image Object}
\usage{
isImage(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
A logical indicating whether the object is of class
 \code{\link{Image}} (TRUE) or not (FALSE).
}
\description{
Tests whether the object is of class \code{\link{Image}}
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
isImage(balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
