% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{isNet}
\alias{isNet}
\title{Test for a Net Object}
\usage{
isNet(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
A logical indicating whether \code{object} is a \code{Net} object.
}
\description{
Tests whether an object is a \code{Net} object as produced by
 \code{\link{readNet}}.
}
\examples{
\dontrun{
net <- readNet("model.onnx")
isNet(net)
}

}
\seealso{
\code{\link{readNet}}
}
\author{
Troy Hernandez, \email{troy.hernandez@pm.me}
}
