% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\name{isQueue}
\alias{isQueue}
\title{Test for a Queue object}
\usage{
isQueue(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
A logical indicating whether the object is of class
 \code{\link{Queue}} (TRUE) or not (FALSE).
}
\description{
Tests whether the object is of class \code{\link{Queue}}.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
buf <- queue(balloon)
isQueue(buf)

}
\seealso{
\code{\link{Queue}}, \code{\link{queue}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
