% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R
\name{isStream}
\alias{isStream}
\title{Test for a Stream Object}
\usage{
isStream(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
A logical indicating whether the object is of class
 \code{\link{Stream}} (TRUE) or not (FALSE).
}
\description{
Tests whether the object is of class \code{\link{Stream}}.
}
\examples{
\dontrun{
live <- stream(0)
isStream(live)
release(live)
}

}
\seealso{
\code{\link{Stream}}, \code{\link{stream}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
