% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoStackClass.R
\name{isVideoStack}
\alias{isVideoStack}
\title{Test for a VideoStack Object}
\usage{
isVideoStack(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
A logical indicating whether the object is of class
 \code{\link{VideoStack}} (TRUE) or not (FALSE).
}
\description{
Tests whether the object is of class \code{\link{VideoStack}}.
}
\examples{
path <- system.file("sample_vid/Balloon.mp4", package = "Rvision")
balloonStack <- videoStack(path, path)
isVideoStack(balloonStack)

}
\seealso{
\code{\link{VideoStack}}, \code{\link{videoStack}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
