% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoWriterClass.R
\name{isVideoWriter}
\alias{isVideoWriter}
\title{Test for a VideoWriter object}
\usage{
isVideoWriter(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
A logical indicating whether the object is of class
 \code{\link{VideoWriter}} (TRUE) or not (FALSE).
}
\description{
Tests whether the object is of class \code{\link{VideoWriter}}
}
\examples{
# TODO
}
\seealso{
\code{\link{VideoWriter}}, \code{\link{videoWriter}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
