% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{laplacian}
\alias{laplacian}
\title{Calculates the Laplacian of an Image}
\usage{
laplacian(image, k_size = 5, scale = 1, in_place = FALSE)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{k_size}{The half-size in pixels of the kernel (default: 5).}

\item{scale}{The scale factor for the computed Laplacian values (default: 1).}

\item{in_place}{A logical indicating whether the change should be applied to
the image itself (TRUE, faster but destructive) or to a copy of it (FALSE,
the default, slower but non destructive).}
}
\value{
An \code{\link{Image}} object if \code{in_place=FALSE}. Otherwise, it
 returns nothing and modifies \code{image} in place.
}
\description{
\code{laplacian} calculates the Laplacian of the source image by
 adding up the second x and y derivatives calculated using the Sobel operator.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_laplacian <- laplacian(balloon, 5)
plot(balloon_laplacian)

}
\seealso{
\code{\link{Image}}, \code{\link{sobel}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
