% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{linearPolar}
\alias{linearPolar}
\title{Linear-Polar Transformation}
\usage{
linearPolar(
  image,
  center = dim(image)[1:2]/2,
  max_radius = min(dim(image)[1:2])/2,
  interp_mode = "linear",
  inverse_map = FALSE,
  target = "new"
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{center}{description}

\item{max_radius}{description}

\item{interp_mode}{A character string indicating the interpolation method to
be used. It can be
any of the following:
\describe{
 \item{"nearest":}{nearest neighbor interpolation.}
 \item{"linear" (the default):}{bilinear interpolation.}
 \item{"cubic":}{bicubic interpolation.}
 \item{"area":}{resampling using pixel area relation. It may be a preferred
  method for image decimation, as it gives moiré-free results, but when the
  image is zoomed, it is similar to the nearest neighbor method.}
 \item{"lanczos4":}{Lanczos interpolation over 8x8 neighborhood.}
 \item{"linear_exact":}{bit exact bilinear interpolation.}
 \item{"nearest_exact":}{Bit exact nearest neighbor interpolation. This will
  produce same results as the nearest neighbor method in PIL, scikit-image
  or Matlab.}
 \item{"fill_outliers":}{fills all of the destination image pixels. If some
  of them correspond to outliers in the source image, they are set to zero.}
}}

\item{inverse_map}{A logical. TRUE if \code{image} is a semilog-polar 
transformed image and you want to inverse that transformation. If FALSE, 
then the image will be semilog-polar transformed (the default).}

\item{target}{The location where the results should be stored. It can take 3
values:
\describe{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  \code{target} must have the same bit depth and number of channels as
  \code{image} but can have different dimensions.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target} is an \code{\link{Image}} object, the function
 returns nothing and modifies that \code{\link{Image}} object in place.
}
\description{
\code{linearPolar} remaps an image to polar coordinates space.
}
\examples{
file <- system.file("sample_img/dots.png", package = "Rvision")
balloon <- changeColorSpace(image(file), "GRAY")

}
\seealso{
\code{\link{linearPolar}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
