% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{LUT}
\alias{LUT}
\title{Look-up Table Transform}
\usage{
LUT(image, lut, in_place = FALSE)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{lut}{A look-up table. This should be a vector with 256 elements, or a
\code{256 x n} table, with n corresponding to the number of channels in
\code{image}. If \code{lut} is a vector and \code{image} has more than one
channel, then \code{lut} is recycled for each channel.}

\item{in_place}{A logical indicating whether the change should be applied to
the image itself (TRUE, faster but destructive) or to a copy of it (FALSE,
the default, slower but non destructive).}
}
\value{
An \code{\link{Image}} object if \code{in_place=FALSE}. Otherwise, it
 returns nothing and modifies \code{image} in place.
}
\description{
\code{LUT} performs a look-up table transform of an
 \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
high_contrast_lut <- 255 * pbeta(0:255 / 255, 4, 4)
high_contrast_balloon <- LUT(balloon, high_contrast_lut)
plot(high_contrast_balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{histmatch}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
