% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{matchKeypoints}
\alias{matchKeypoints}
\title{Match Keypoints}
\usage{
matchKeypoints(
  source,
  target,
  descriptor_matcher = "BruteForce-Hamming",
  match_frac = 0.15
)
}
\arguments{
\item{source, target}{Single-channel \code{\link{Image}} objects
containing the BRIEF descriptors of the source and target images, as produced
by \code{\link{ORBkeypoints}}.}

\item{descriptor_matcher}{A character string indicating the type of the
descriptor matcher to use. It can be one of the followings: "BruteForce",
"BruteForce-L1", "BruteForce-Hamming" (the default), "BruteForce-Hamming(2)",
or "FlannBased".}

\item{match_frac}{The fraction of top matches to keep (default: 0.15).}
}
\value{
A three-column matrix with the identities of the keypoints matched
 between the source and target images, and the distance between them (a lower
 distance indicates a better match).
}
\description{
\code{matchKeypoints} matches keypoints detected in two separate
 images. This is useful to find common features for image registration, for
 instance.
}
\examples{
balloon1 <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon2 <- image(system.file("sample_img/balloon2.png", package = "Rvision"))
kp1 <- ORBkeypoints(balloon1, n_features = 40000)
kp2 <- ORBkeypoints(balloon2, n_features = 40000)
matchKeypoints(kp1$descriptors, kp2$descriptors, match_frac = 1)

}
\seealso{
\code{\link{ORBkeypoints}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
