% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{matchShapes}
\alias{matchShapes}
\title{Compare Two Shapes}
\usage{
matchShapes(x1, x2, method = "I1")
}
\arguments{
\item{x1}{Either a Nx2 matrix of the X-Y coordinates of a polygon (e.g., a
contour produced by \code{\link{findContours}}), or a single-channel
\code{\link{Image}} object.}

\item{x2}{Either a Nx2 matrix of the X-Y coordinates of a polygon (e.g., a
contour produced by \code{\link{findContours}}), or a single-channel
\code{\link{Image}} object.}

\item{method}{The comparison method to compute the difference between the two
shapes (see Notes; default: "I1").}
}
\value{
A numerical value.
}
\description{
\code{matchShapes} computes the difference between two shapes
 using the Hu invariants.
}
\note{
The available shape matching methods are defined as follows:
\itemize{
   \item{\eqn{I_1(A,B) = \sum _{i=1...7} \left | \frac{1}{m^A_i} - \frac{1}{m^B_i} \right |}}
   \item{\eqn{I_2(A,B) = \sum _{i=1...7} \left | m^A_i - m^B_i \right |}}
   \item{\eqn{I_3(A,B) = \max _{i=1...7} \frac{ \left| m^A_i - m^B_i \right| }{ \left| m^A_i \right| }}}
}

where

\itemize{
   \item{\eqn{A} denotes x1, \eqn{B} denotes x2.}
   \item{\eqn{m^A_i = \mathrm{sign} (h^A_i) \cdot \log{h^A_i}}}
   \item{\eqn{m^B_i = \mathrm{sign} (h^B_i) \cdot \log{h^B_i}}}
   \item{and \eqn{h^A_i, h^B_i} are the Hu invariants of \eqn{A} and \eqn{B}, respectively.}
}
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_gray <- changeColorSpace(dots, "GRAY")
dots_bin <- dots_gray < 200
contours <- findContours(dots_bin)
contour0 <- contours$contours[contours$contours[, 1] == 0, 2:3]
contour1 <- contours$contours[contours$contours[, 1] == 1, 2:3]
matchShapes(contour0, contour1)

}
\seealso{
\code{\link{findContours}}, \code{\link{huInvariants}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
