% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{matchTemplate}
\alias{matchTemplate}
\title{Template Matching}
\usage{
matchTemplate(image, template, method, mask = NULL)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{template}{An \code{\link{Image}} object of the same type as \code{image}.
\code{template} cannot be larger than \code{image}.}

\item{method}{A string indicating the comparison method to use. It can be any
of the following (see \url{https://bit.ly/2RjELvJ} for a full description of each
comparison method):
\itemize{
 \item "SQDIFF"
 \item "SQDIFF_NORMED"
 \item "CCORR"
 \item "CCORR_NORMED"
 \item "CCOEFF"
 \item "CCOEFF_NORMED"
}}

\item{mask}{An \code{\link{Image}} object of the same type and dimensions as
\code{template} (default: NULL). \code{mask} is currently only supported when
\code{SQDIFF} and \code{CCORR_NORMED} are used.}
}
\value{
A 16-bit grayscale \code{\link{Image}} object of the same dimensions
 as \code{image}.
}
\description{
\code{matchTemplate} compares a template against overlapping
 image regions using the specified \code{method}. After the function finishes
 the comparison, the best matches can be found as global minimums (when
 methods "SQDIFF" or "SQDIFF_NORMED" were used) or maximums (when methods
 "CCORR", "CCORR_NORMED", "CCOEFF" or "CCOEF_NORMED" were used) using the
 \code{\link{minMaxLoc}} function.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
sub <- subImage(balloon, 290, 170, 150, 150)
match <- matchTemplate(balloon, sub, method = "SQDIFF")
mm <- minMaxLoc(match)
plot(balloon)
points(mm[1, 2], mm[1, 3], col = "red", pch = 20, cex = 5)

}
\seealso{
\code{\link{Image}}, \code{\link{minMaxLoc}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
