% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{mean.list}
\alias{mean.list}
\title{Mean of Images in a List}
\usage{
\method{mean}{list}(x, ...)
}
\arguments{
\item{x}{A list of \code{\link{Image}} objects. All images must have the same
dimensions, number of channels, and bitdepth.}

\item{...}{Unused at the moment.}
}
\value{
An \code{\link{Image}} object.
}
\description{
Returns the pixelwise mean of \code{\link{Image}} objects stored
in a list.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
img_list <- lapply(1:10, function(x) readNext(balloon))
plot(mean(img_list))

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
