% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{mean.Rcpp_Image}
\alias{mean.Rcpp_Image}
\title{Mean Value of the Pixels in an Image}
\usage{
\method{mean}{Rcpp_Image}(x, ..., mask = NA)
}
\arguments{
\item{x}{An \code{\link{Image}} object.}

\item{...}{Unused at the moment.}

\item{mask}{A single-channel (GRAY) 8-bit (8U) \code{\link{Image}} object
with the same dimensions as \code{x}. This can be used to mask out pixels
that should not be considered when calculating the mean (pixels set to 0 in
the mask will be ignored during the mean calculation).}
}
\value{
A numeric value (for single-channel images) or a vector of numeric
 values (for multi-channel images).
}
\description{
Returns the mean of the pixel values of an \code{\link{Image}}
 object. If the \code{\link{Image}} object has more than one channel, it
 returns the mean for each channel.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
mean(balloon)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
