% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{merge}
\alias{merge}
\title{Merge Separate Channels into an Image}
\usage{
merge(x)
}
\arguments{
\item{x}{A list of single channel (grayscale) \code{\link{Image}} objects.}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{merge} returns an image from the combination of grayscale
 images corresponding to single channels (green, blue, red, or alpha).
}
\note{
Color images are usually represented by 3 channels (possibly 4) in the
 following order: green (1), blue (2), red (3), and possibly alpha (4).
}
\examples{
# TODO
}
\seealso{
\code{\link{split}}, \code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
