% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{min.Rcpp_Image}
\alias{min.Rcpp_Image}
\alias{max.Rcpp_Image}
\alias{range.Rcpp_Image}
\title{Maxima and Minima of an Image}
\usage{
\method{min}{Rcpp_Image}(x, ...)

\method{max}{Rcpp_Image}(x, ...)

\method{range}{Rcpp_Image}(x, ...)
}
\arguments{
\item{x}{An \code{\link{Image}} object.}

\item{...}{Unused at the moment.}
}
\value{
\code{min} and \code{max} return a matrix with 1 row and \code{nchan(x)}
 columns. \code{range} returns a matrix with 2 rows and \code{nchan(x)} columns.
}
\description{
Returns the maximum and minimum pixel values of an
 \code{\link{Image}} object. If the \code{\link{Image}} object has more than
 one channel, it returns the maximum and minimum of each channel.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
min(balloon)
max(balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{minMaxLoc}}.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
