% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{minMaxLoc}
\alias{minMaxLoc}
\title{Coordinates of the Maxima and Minima of an Image}
\usage{
minMaxLoc(x)
}
\arguments{
\item{x}{An \code{\link{Image}} object.}
}
\value{
A matrix (or a list of matrices for multi-channels images).
}
\description{
\code{minMaxLoc} returns the maximum and minimum pixel values of
 an \code{\link{Image}} object, as well as their coordinates in the image. If
 the \code{\link{Image}} object has more than one channel, it returns the
 values for each channel.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
minMaxLoc(balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{min.Rcpp_Image}}, \code{\link{max.Rcpp_Image}}.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
