% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{moments}
\alias{moments}
\title{Calculate the Moments of a Shape}
\usage{
moments(contours, id = NULL)
}
\arguments{
\item{contours}{A list of two matrices as produced by \code{\link{findContours}}.}

\item{id}{An optional vector indicating the identity of the specific contours
for which to run the function.}
}
\value{
A data frame with 3 columns:
 \itemize{
   \item{"id": }{the contour identity (indicates the set of moments belonging
    to the same contour).}
   \item{"moment": }{the name of the moment. See Note below.}
   \item{"value": }{the value of the moment.}
}
}
\description{
\code{moments} calculates all of the moments up to the third
 order of a polygon or rasterized shape.
}
\note{
The spatial moments \eqn{m_{ji}} are computed as:
 \deqn{m_{ji}= \sum _{x,y} \left ( \texttt{contour} (x,y) \cdot x^j \cdot y^i \right )}

The central moments \eqn{\mu_{ji}} are computed as:
 \deqn{{\mu_{ji}}= \sum _{x,y} \left ( \texttt{contour} (x,y) \cdot (x - \bar{x} )^j \cdot (y - \bar{y} )^i \right )}
where \eqn{(\bar{x}, \bar{y})} is the mass center:
 \deqn{\bar{x} = \frac{m_{10}}{m_{00}} , \; \bar{y} = \frac{m_{01}}{m_{00}}}

The normalized central moments \eqn{\eta_{ji}} are computed as:
 \deqn{\eta_{ji}= \frac{\mu_{ji}}{m_{00}^{(i+j)/2+1}} .}
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_gray <- changeColorSpace(dots, "GRAY")
dots_bin <- dots_gray < 200
contours <- findContours(dots_bin)
moments(contours, id = c(3, 5))

}
\seealso{
\code{\link{findContours}}, \code{\link{huInvariants}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
