% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{newDisplay}
\alias{newDisplay}
\title{Open New \code{\link{Image}} Display}
\usage{
newDisplay(window_name = "Display", height = 480, width = 640)
}
\arguments{
\item{window_name}{A character string representing the name of the display
window (default: "Display").}

\item{height}{An integer representing the height in pixels of the display
window.}

\item{width}{An integer representing the width in pixels of the display
window.}
}
\value{
This function does not return anything.
}
\description{
\code{newDisplay} creates a window to display \code{\link{Image}}
 objects.
}
\examples{
\dontrun{
newDisplay()
}

}
\seealso{
\code{\link{Image}}, \code{\link{display}}, \code{\link{destroyDisplay}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
