% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{nrow.Rcpp_Image}
\alias{nchan}
\alias{ncol.Rcpp_Image}
\alias{nrow.Rcpp_Image}
\title{The Number of Rows/Columns/Channels of an Image}
\usage{
nrow(image)
ncol(image)
nchan(image)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}
}
\value{
An numeric value.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link[=dim.Rcpp_Image]{dim}} which returns \emph{all}
 dimensions.
}

