% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{nrow.Rcpp_Video}
\alias{nrow.Rcpp_Video}
\alias{ncol.Rcpp_Video}
\alias{nframes}
\title{The Number of Rows/Columns/Frames of a Video}
\usage{
nrow(video)
ncol(video)
nframes(video)
}
\arguments{
\item{video}{A \code{\link{Video}} object.}
}
\value{
A numeric value.
}
\description{
nrow, ncol and nframes return the number of rows, columns or
 frames present in a \code{\link{Video}} object.
}
\examples{
# TODO

}
\seealso{
\code{\link[=dim.Rcpp_Video]{dim}} which returns \emph{all}
 dimensions.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
