% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{ones}
\alias{ones}
\title{Create a One-Filled Image}
\usage{
ones(nrow, ncol, colorspace = "BGR", bitdepth = "8U")
}
\arguments{
\item{nrow}{An integer indicating desired the number of rows for the image.}

\item{ncol}{An integer indicating desired the number of columns for the image.}

\item{colorspace}{A string indicating the desired color space for the image.
Options are "BGR" (Blue Green Red image, the default), "BGRA" (BGR image
with Alpha channel), and "GRAY" (grayscale image).}

\item{bitdepth}{A string indicating the desired bit depth for the image.
Options are "8U" (the default), "8S", "16U", "16S", "32S", and "32F".}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{ones} creates an \code{\link{Image}} object filled with
 ones.
}
\examples{
one <- ones(100, 100)

}
\seealso{
\code{\link{Image}}, \code{\link{zeros}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
