% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{pget}
\alias{pget}
\title{Return Pixel Value at Specified Locations}
\usage{
pget(image, x, y)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{x}{A vector of x locations (columns) in the image.}

\item{y}{A vector of y locations (rows) in the image.}
}
\value{
A matrix. The number of columns of the matrix depends on the number
 of channels in the image. Each row corresponds to a pair of x/y coordinates.
}
\description{
\code{pget} returns the values of the pixels at the specified x
 and y coordinates in the image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
pget(balloon, c(1, 100, 200), c(200, 100, 1))

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
