% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{pixelsInContour}
\alias{pixelsInContour}
\title{Which Pixels are Inside a Contour}
\usage{
pixelsInContour(contours, id = NULL)
}
\arguments{
\item{contours}{A list of two matrices as produced by \code{\link{findContours}}.}

\item{id}{An optional vector indicating the identity of the specific contours
for which to run the function.}
}
\value{
A matrix with 3 columns:
 \itemize{
   \item{"id": }{the contour identity (indicates the set of points belonging
    to the same contour).}
   \item{"x": }{the x coordinates of the points inside the contour.}
   \item{"y": }{the y coordinates of the points inside the contour.}
}
}
\description{
\code{pixelsInContour} determines the pixels that are inside a
 specified contour.
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_gray <- changeColorSpace(dots, "GRAY")
dots_bin <- dots_gray < 200
contours <- findContours(dots_bin)
pixelsInContour(contours, id = c(3, 5))

}
\seealso{
\code{\link{findContours}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
