% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{plot.blob}
\alias{plot.blob}
\title{Plot Blobs}
\usage{
\method{plot}{blob}(x, col = "red", asp = 1, ...)
}
\arguments{
\item{x}{A \code{blob} object.}

\item{col}{The color of the \code{blob} center and box (default: "red").}

\item{asp}{The y/x aspect ratio of the plotting window (default: 1).}

\item{...}{Additional arguments to be passed to \code{\link{plot}}.}
}
\description{
Plotting method for objects inheriting from class \code{blob}.
}
\examples{
# TODO
}
\seealso{
\code{\link{isBlob}}, \code{\link{simpleBlobDetector}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
