% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{plot.Image}
\alias{plot.Image}
\alias{plot.Rcpp_Image}
\title{Plotting \pkg{Rvision} Images}
\usage{
\method{plot}{Rcpp_Image}(image, min = 0, max = 255, ...)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{min}{The minimum value that a pixel in the image can take (default: 0).}

\item{max}{The maximum value that a pixel in the image can take (default: 255).}

\item{...}{Additional arguments to be passed to \code{\link{rasterImage}}.}
}
\description{
Plotting method for objects inheriting from class \code{\link{Image}}.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{Image}}
}

