% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opticalFlow.R
\name{plot.OF_array}
\alias{plot.OF_array}
\title{Plot Optical Flow Arrays}
\usage{
\method{plot}{OF_array}(array, gridsize = c(25, 25), thresh = 0,
  add = TRUE, arrow.ex = 0.05, xpd = TRUE, ...)
}
\arguments{
\item{array}{An object of class \code{OF_array}.}

\item{gridsize}{A 2-element vector indicating the number of optical flow
vectors to plot in each x-y dimension (default: c(25, 25)). Alternatively, a
numeric value that will be used for both dimensions.}

\item{thresh}{The minimal length of optical flow vectors that should be
plotted (default: 0).}

\item{add}{A logical indicating whether to plot the vector field over an
existing plot (default: FALSE).}

\item{arrow.ex}{Controls the length of the arrows. The length is in terms of
the fraction of the shorter axis in the plot. So with a default of .05, 20
arrows of maximum length can line up end to end along the shorter axis.}

\item{xpd}{If true does not clip arrows to fit inside the plot region,
default is not to clip.}
}
\description{
Plotting method for objects of class \code{OF_array} as produced
 by the \code{\link{farneback}} function.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{farneback}}, \code{\link{arrows}}
}

