% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opticalFlow.R
\name{plotOF}
\alias{plotOF}
\title{Plot Optical Flow Arrays}
\usage{
plotOF(
  of,
  gridsize = c(25, 25),
  thresh = 0,
  add = TRUE,
  arrow.ex = 0.05,
  xpd = TRUE,
  ...
)
}
\arguments{
\item{of}{An \code{\link{Image}} produced by the \code{\link{farneback}}
function.}

\item{gridsize}{A 2-element vector indicating the number of optical flow
vectors to plot in each x-y dimension (default: c(25, 25)). Alternatively, a
numeric value that will be used for both dimensions.}

\item{thresh}{The minimal length of optical flow vectors that should be
plotted (default: 0).}

\item{add}{A logical indicating whether to plot the vector field over an
existing plot (default: FALSE).}

\item{arrow.ex}{Controls the length of the arrows. The length is in terms of
the fraction of the shorter axis in the plot. So with a default of .05, 20
arrows of maximum length can line up end to end along the shorter axis.}

\item{xpd}{If true does not clip arrows to fit inside the plot region,
default is not to clip.}

\item{...}{Graphics arguments passed to the \code{\link{arrows}} function that
can change the color or arrow sizes. See \code{\link{arrows}} help for details.}
}
\description{
Plotting method for \code{\link{Image}} objects produced by the
 \code{\link{farneback}} function.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
balloon1 <- readFrame(balloon, 1)
balloon2 <- readFrame(balloon, 2)
changeColorSpace(balloon1, "GRAY", "self")
changeColorSpace(balloon2, "GRAY", "self")
of <- farneback(balloon1, balloon2)
plot(of)
plotOF(of, length = 0.05)

}
\seealso{
\code{\link{farneback}}, \code{\link{arrows}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
