% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{pow}
\alias{pow}
\title{Power Function}
\usage{
pow(x, y = 2, target = "new")
}
\arguments{
\item{x}{A 32- or 64-bit (32F or 64F) \code{\link{Image}} object.}

\item{y}{A numeric value representing the exponent of power. Note that for
a non-integer power exponent, the absolute values of the \code{\link{Image}}
object are used.}

\item{target}{The location where the results should be stored when passing a
sum of images to the function. It can take 3 values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{x} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast but will replace the
  content of \code{target}. Note that \code{target} should have the same
  dimensions, bitdepth, and number of channels as \code{x}, otherwise an
  error will be thrown.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{x} in place. If \code{target} is an \code{\link{Image}} object, the
 function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
\code{pow} raises every element of an \code{\link{Image}} object
 to a power.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
changeBitDepth(balloon, "32F", 1 , "self")
pow(balloon, 2)

}
\seealso{
\code{\link{exp}}, \code{\link{log}}, \code{\link{sqrt}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
