% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{pset}
\alias{pset}
\title{Set Pixel Value at Specified Locations}
\usage{
pset(image, x, y, color)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{x}{A vector of x locations (columns) in the image.}

\item{y}{A vector of y locations (rows) in the image.}

\item{color}{A value or vector of any kind of R color specification compatible
with \code{\link{col2bgr}}. It can also be a matrix with the same number of
columns as the number of elements in \code{x} and the same number of rows as
the number of channels in \code{image}.}
}
\value{
This functions returns nothing and changes the values of the pixels
 in \code{image} in place.
}
\description{
\code{pset} sets the values of the pixels at the specified x
 and y coordinates in the image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
pset(balloon, 1:100, 200:101, "red")

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
