% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{pyrDown}
\alias{pyrDown}
\title{Pyramid Downsampling}
\usage{
pyrDown(image, target = "new")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{target}{The location where the results should be stored. It can take 2
values:
\describe{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  if \code{target} does not have the same number of channels, and bit depth
  as \code{image}, an error may be thrown. The dimensions of \code{target}
  must satisfy the following conditions:
    \itemize{
      \item{\code{ abs(ncol(target) * 2 - ncol(image)) <= 2 }}
      \item{\code{ abs(nrow(target) * 2 - nrow(image)) <= 2 }}
    }
  }
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target} is an \code{\link{Image}} object, the function
 returns nothing and modifies that \code{\link{Image}} object in place.
}
\description{
\code{prDown} blurs an image and then downsamples it.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
small_balloon <- pyrDown(balloon)

}
\seealso{
\code{\link{pyrUp}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
