% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\name{queue}
\alias{queue}
\title{Create an Object of Class \code{Queue}}
\usage{
queue(x, size = 10, delay = 1000, overflow = "pause")
}
\arguments{
\item{x}{Either a \code{\link{Video}} or a \code{\link{Stream}} object.}

\item{size}{The number of frames that the \code{\link{Queue}} object can
store (default: 10). This number is fixed if \code{overflow} is set to
\code{"pause"} or \code{"replace"}. This number may increase if
\code{overflow} is set to \code{"grow"} and the queue becomes full.}

\item{delay}{Time in microseconds between two queue updates (default: 1000).
Increasing the delay will slow down the speed at which the queue fills up.
Decreasing it will fill up the queue faster but will use up more processing
resources.}

\item{overflow}{A character string indicating what the \code{\link{Queue}}
object should do when it is full. Three methods are available:
\itemize{
 \item{"pause": }{the queue will stop retrieving and storing new frames
  until a frame is read by the user.}
 \item{"replace": }{the oldest frame in the queue is discarded to make space
  for a new frame. }
 \item{"grow": }{the size of the queue is doubled every time it fills up.
  This should be used with extreme caution as it can lead to excessive
  memory usage.}
}}
}
\value{
A \code{\link{Queue}} object.
}
\description{
Function for creating \code{\link{Queue}} objects for
 \code{\link{Video}} and \code{\link{Stream}} objects.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
buf <- queue(balloon)

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
