% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\name{nrow.Rcpp_Queue}
\alias{nrow.Rcpp_Queue}
\alias{ncol.Rcpp_Queue}
\alias{length.Rcpp_Queue}
\title{The Number of Rows/Columns and Length of a Queue}
\usage{
nrow.Rcpp_Queue(x)

ncol.Rcpp_Queue(x)

\method{length}{Rcpp_Queue}(x)
}
\arguments{
\item{x}{A \code{\link{Queue}} object.}
}
\value{
A numeric value.
}
\description{
nrow, ncol and nframes return the number of rows, columns or
 frames present in a \code{\link{Queue}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
buf <- queue(balloon)
nrow(buf)
ncol(buf)
length(buf)

}
\seealso{
\code{\link[=dim.Rcpp_Queue]{dim}} which returns \emph{all}
 dimensions.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
