% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{randn}
\alias{randn}
\title{Random Normal Image}
\usage{
randn(image, mean = 127, sd = 1)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{mean}{A vector indicating the mean value (expectation) of the generated
random numbers. It can have as many elements as the number of channels in
the image. If it has less elements than the number of channels, it is
recycled to match the number of channels. If it has more elements than the
number of channels, the extra elements are ignored without warning (default:
0).}

\item{sd}{A vector indicating the standard deviation  of the generated
random numbers. It can have as many elements as the number of channels in
the image. If it has less elements than the number of channels, it is
recycled to match the number of channels. If it has more elements than the
number of channels, the extra elements are ignored without warning (default:
256).}
}
\value{
This function returns nothing and modifies image in place.
}
\description{
\code{randn} replaces the content of an \code{\link{Image}}
 object with normally-distributed random pixel values.
}
\examples{
rnd <- zeros(100, 100)
randn(rnd)

}
\seealso{
\code{\link{Image}}, \code{\link{randn}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
