% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{randu}
\alias{randu}
\title{Random Uniform Image}
\usage{
randu(image, low = 0, high = 256)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{low}{A vector indicating the inclusive lower boundary of the generated
random numbers. It can have as many elements as the number of channels in
the image. If it has less elements than the number of channels, it is
recycled to match the number of channels. If it has more elements than the
number of channels, the extra elements are ignored without warning (default:
0).}

\item{high}{A vector indicating the exclusive upper boundary of the generated
random numbers. It can have as many elements as the number of channels in
the image. If it has less elements than the number of channels, it is
recycled to match the number of channels. If it has more elements than the
number of channels, the extra elements are ignored without warning (default:
256).}
}
\value{
This function returns nothing and modifies image in place.
}
\description{
\code{randu} replaces the content of an \code{\link{Image}}
 object with uniformly-distributed random pixel values.
}
\examples{
rnd <- zeros(100, 100)
randu(rnd)

}
\seealso{
\code{\link{Image}}, \code{\link{randn}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
