% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R, R/VideoStackClass.R, R/generic.R
\name{readFrame.Rcpp_Video}
\alias{readFrame.Rcpp_Video}
\alias{readFrame.VideoStack}
\alias{readFrame}
\title{Read Specific Video Frame}
\usage{
\method{readFrame}{Rcpp_Video}(x, pos, target = "new")

\method{readFrame}{VideoStack}(x, pos, target = "new")

readFrame(x, pos, target = "new")
}
\arguments{
\item{x}{A \code{\link{Video}} or \code{\link{VideoStack}} object.}

\item{pos}{An integer corresponding to the position of the frame to read in
the video or video stack.}

\item{target}{The location where the results should be stored. It can take 2
values:
\describe{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This will replace the content of
  \code{target}. Note that \code{target} must have the same dimensions as
  \code{x}.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target} is an \code{\link{Image}} object, the function
 returns nothing and modifies that \code{\link{Image}} object in place.
}
\description{
Read a specific frame of a \code{\link{Video}} or
 \code{\link{VideoStack}} object and returns it as an \code{\link{Image}}
 object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
frame10 <- readFrame(balloon, 10)

}
\seealso{
\code{\link{Video}}, \code{\link{VideoStack}}, \code{\link{Image}},
 \code{\link{readNext}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
