% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{readHIS}
\alias{readHIS}
\title{Read HIS Image from Varex Imaging X-ray Panels}
\usage{
readHIS(x)
}
\arguments{
\item{x}{A character string naming the path to a HIS binary image file.}
}
\value{
A 16U grayscale \code{\link{Image}} object.
}
\description{
\code{readHIS} reads HIS binary files produced by the X-ray
 panels from Varex Imaging and returns \code{\link{Image}} objects that can
 be processed by this package.
}
\examples{
#TODO

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
