% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{readMulti}
\alias{readMulti}
\title{Read a Multi-Page Image}
\usage{
readMulti(file)
}
\arguments{
\item{file}{A character string naming the path to a multi-page image file.}
}
\value{
A list of \code{\link{Image}} objects.
}
\description{
\code{readMulti} reads a multi-page image and returns a list of
 \code{\link{Image}} objects, each corresponding to a different page.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}
}
