% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R
\name{readNext.Rcpp_Stream}
\alias{readNext.Rcpp_Stream}
\title{Read Next Stream Frame}
\usage{
\method{readNext}{Rcpp_Stream}(obj)
}
\arguments{
\item{obj}{A \code{\link{Stream}} object.}
}
\value{
An \code{\link{Image}} object.
}
\description{
Read the next frame of a \code{\link{Stream}} object and returns
 it as an \code{\link{Image}} object.
}
\examples{
# TODO

}
\seealso{
\code{\link{Stream}}, \code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
