% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{readNext.Rcpp_Video}
\alias{readNext.Rcpp_Video}
\title{Read Next Video Frame}
\usage{
\method{readNext}{Rcpp_Video}(obj)
}
\arguments{
\item{obj}{A \code{\link{Video}} object.}
}
\value{
An \code{\link{Image}} object.
}
\description{
Read the next frame of a \code{\link{Video}} object and returns
 it as an \code{\link{Image}} object.
}
\examples{
# TODO

}
\seealso{
\code{\link{Video}}, \code{\link{Image}}, \code{\link{readFrame}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
