% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R, R/VideoClass.R, R/generic.R
\name{readNext.Rcpp_Stream}
\alias{readNext.Rcpp_Stream}
\alias{readNext.Rcpp_Video}
\alias{readNext}
\title{Read Next Frame of an Object}
\usage{
\method{readNext}{Rcpp_Stream}(obj)

\method{readNext}{Rcpp_Video}(obj)

readNext(obj)
}
\arguments{
\item{obj}{A \code{\link{Video}} or \code{\link{Stream}} object.}
}
\value{
An \code{\link{Image}} object.
}
\description{
Read the next frame of a \code{\link{Video}} or
 \code{\link{Stream}} object and returns it as an \code{\link{Image}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
plot(readNext(balloon))
release(balloon)

}
\seealso{
\code{\link{Video}}, \code{\link{Stream}},
\code{\link{Image}}, \code{\link{readFrame}}, \code{\link{readStream}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
