% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R, R/VideoClass.R, R/generic.R
\name{readStream.Rcpp_Stream}
\alias{readStream.Rcpp_Stream}
\alias{readStream.Rcpp_Video}
\alias{readStream}
\title{Read Next Frame of an Object in Place}
\usage{
\method{readStream}{Rcpp_Stream}(obj, image, ...)

\method{readStream}{Rcpp_Video}(obj, image, pos = -1, ...)

readStream(obj, image, ...)
}
\arguments{
\item{obj}{A \code{\link{Video}} or \code{\link{Stream}} object.}

\item{image}{An \code{\link{Image}} to update in place.}

\item{...}{Not used.}

\item{pos}{An integer corresponding to the number of the frame to read in the
video. If -1 (the default), then the next frame is read.}
}
\value{
This function returns nothing and modifies \code{image} in place.
}
\description{
Read the next frame of a \code{\link{Video}} or
 \code{\link{Stream}} object and use it to update an \code{\link{Image}}
 object in place.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
frame <- readNext(balloon)
plot(frame)
readStream(balloon, frame)
plot(frame)
release(balloon)

}
\seealso{
\code{\link{Video}}, \code{\link{Stream}},
\code{\link{Image}}, \code{\link{readFrame}}, \code{\link{readNext}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
