% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R
\name{release.Rcpp_Stream}
\alias{release.Rcpp_Stream}
\title{Release Stream from Memory}
\usage{
\method{release}{Rcpp_Stream}(obj)
}
\arguments{
\item{obj}{A \code{\link{Stream}} object.}
}
\value{
If successful, the \code{\link{Stream}} object is cleared from memory
}
\description{
Close a \code{\link{Stream}} object.
}
\examples{
# TODO
}
\seealso{
\code{\link{Stream}}, \code{\link{stream}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
