% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R, R/VideoClass.R,
%   R/VideoWriterClass.R, R/generic.R
\name{release.Rcpp_Stream}
\alias{release.Rcpp_Stream}
\alias{release.Rcpp_Video}
\alias{release.Rcpp_VideoWriter}
\alias{release}
\title{Release Object from Memory}
\usage{
\method{release}{Rcpp_Stream}(x)

\method{release}{Rcpp_Video}(x)

\method{release}{Rcpp_VideoWriter}(obj)

release(x)
}
\arguments{
\item{x}{A \code{\link{Video}} or \code{\link{Stream}} object.}
}
\value{
If successful, the \code{\link{Stream}} or
\code{\link{Video}} object is cleared from memory
}
\description{
Close a \code{\link{Stream}} or \code{\link{Video}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
release(balloon)

}
\seealso{
\code{\link{Stream}}, \code{\link{stream}}, \code{\link{Video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
