% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\name{reset}
\alias{reset}
\title{Reset a Queue}
\usage{
reset(x)
}
\arguments{
\item{x}{A \code{\link{Queue}} object.}
}
\value{
This function returns nothing.
}
\description{
\code{reset} flush a \code{\link{Queue}} object from all the
 frames it contains and starts filling it up again with new frames.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
buf <- queue(balloon)
reset(buf)

}
\seealso{
\code{\link{Queue}}, \code{\link{queue}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
