% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{rotate}
\alias{rotate}
\title{Image Rotation In 90-Degree Increments}
\usage{
rotate(image, rotation = "CLOCKWISE", target = "new")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{rotation}{A character string indicating the desired rotation:
\itemize{
 \item{"CLOCKWISE" (default):}{rotate by 90 degrees clockwise.}
 \item{"COUNTER":}{rotate by 90 degrees counterclockwise.}
 \item{"180":}{rotate by 180 degrees.}
}}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  \code{target} must have the same bit depth and number of channels as
  \code{image}. The dimensions must be the same if \code{rotation="CLOCKWISE"},
  or inverted if \code{rotation="CLOCKWISE"} or \code{rotation="COUNTER"}.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target} is an \code{\link{Image}} object, the function
 returns nothing and modifies that \code{\link{Image}} object in place.
}
\description{
\code{rotate} rotates an image in multiples of 90 degrees.
}
\examples{
img <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
img_rotated <- rotate(img)

}
\seealso{
\code{\link{rotateScale}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
