% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rvision-package.R
\docType{package}
\name{Rvision}
\alias{Rvision}
\alias{Rvision-package}
\title{Rvision: A basic (for now) computer vision library for R.}
\description{
The \code{Rvision} package uses the OpenCV library to provide R users with
some basic functions to read and manipulate images, videos and camera streams.
}
\details{
\tabular{ll}{
 Package: \tab Rvision\cr
 Type: \tab Package\cr
 Version: \tab 0.1.0\cr
 Date: \tab 2016-07-06\cr
 License: \tab GPL-3\cr
}
}
\references{
More about OpenCV at \url{http://opencv.org/}.
}
\author{
Simon Garnier <garnier@njit.edu>

Maintainer: Simon Garnier <garnier@njit.edu>
}
