% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ack.R
\docType{data}
\name{RvisionAck}
\alias{RvisionAck}
\title{Image and video samples}
\description{
\code{\link{Rvision}} provides a few image and video samples for illustrating
the examples of its different functions. Below are references to the original
material from which these samples have been extracted.
}
\references{
\itemize{
 \item{"Big Buck Bunny", © copyright 2008, Blender Foundation /
   \url{www.bigbuckbunny.org} - Distributed under the Creative Commons
   Attribution 3.0 license.}
 \item{EC Funded CAVIAR project/IST 2001 37540 -
   \url{http://homepages.inf.ed.ac.uk/rbf/CAVIAR/}}
 \item{Pixabay / \url{https://pixabay.com/en/dots-blue-background-fabric-313325/}
   - Distributed under the Creative Commons CC0 Public Domain license.}
}
}
