% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{selectROI}
\alias{selectROI}
\title{Select a Region of Interest in an Image}
\usage{
selectROI(image, window_name = "Display", scale = 1, return_mask = TRUE)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{window_name}{A character string representing the name of the display
window (default: "Display").}

\item{scale}{The scaling of the display relative to the image size (default: 1).}

\item{return_mask}{Should the function return a mask of the ROI. The mask is
an 8-bit single-channel image with the pixels inside the ROI painted white
and the pixels outside the ROI painted black (default: TRUE).}
}
\value{
If \code{return_mask == FALSE}, a data frame with the following two
 columns:
 \itemize{
  \item{x: }{the x coordinates of the ROI polygon.}
  \item{y: }{the y coordinates of the ROI polygon.}
 }
 If \code{return_mask == TRUE}, a list containing the data frame containing
 the coordinates of the ROIand an 8-bit single-channel image corresponding to
 the mask of the ROI.
}
\description{
\code{selectROI} allows the user to select a region of interest
 (ROI) in an image. An ROI is a polygonal region selected by clicking on the
 locations of the vertices of the polygon.
}
\examples{
\dontrun{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
selectROI(balloon)
}

}
\seealso{
\code{\link{Image}}, \code{\link{newDisplay}}, \code{\link{display}},
 \code{\link{click}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
