% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R
\name{setProp.Rcpp_Stream}
\alias{setProp.Rcpp_Stream}
\alias{getProp.Rcpp_Stream}
\title{Set/Get Stream Properties}
\usage{
setProp(obj, property, value)
getProp(obj, property)
}
\arguments{
\item{obj}{A \code{\link{Stream}} object.}

\item{property}{A character string specifying the name of the property to
modify (see details below for a complete list).}

\item{value}{The new value of the property.}
}
\value{
\code{setProp} returns TRUE is the property was set successfully.
 \code{getProp} returns a numeric value or a character string depending on
 \code{property}.
}
\description{
Set or get the values of various properties of the
 \code{\link{Stream}} object.
}
\note{
Setting stream properties depends on a lot of things, mainly your
 operating system, the camera drivers installed on your coputer and the
 camera itself. As a consequence, setting stream values might not work at all
 with your installation.

 Stream properties are:
 \itemize{
   \item{\code{FRAME_WIDTH}: Width in pixels of the frames in the video stream.}
   \item{\code{FRAME_HEIGHT}: Height in pixels of the frames in the video stream.}
   \item{\code{BRIGHTNESS}: Brightness of the image}
   \item{\code{CONTRAST}: Contrast of the image }
   \item{\code{SATURATION}: Saturation of the image}
   \item{\code{HUE}: Hue of the image}
   \item{\code{GAIN}: Gain of the image}
   \item{\code{EXPOSURE}: Exposure}
 }
}
\examples{
# TODO
}
\seealso{
\code{\link{Stream}}, \code{\link{stream}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
