% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{setProp.Rcpp_Video}
\alias{getProp.Rcpp_Video}
\alias{setProp.Rcpp_Video}
\title{Set/Get Video Properties}
\usage{
setProp(video, property, value)
getProp(video, property)
}
\arguments{
\item{video}{A \code{\link{Video}} object.}

\item{property}{A character string specifying the name of the property to
modify (see details below for a complete list).}

\item{value}{The new value of the property.}
}
\value{
\code{setProp} returns TRUE is the property was set successfully.
 \code{getProp} returns a numeric value or a character string depending on
 \code{property}.
}
\description{
Set or get the values of various properties of the
 \code{\link{Video}} object.
}
\note{
Video properties are:
\itemize{
 \item{\code{POS_MSEC}: Current position of the video file in milliseconds.}
 \item{\code{POS_FRAMES}: 0-based index of the frame to be decoded/captured next.}
 \item{\code{POS_AVI_RATIO} Relative position of the video file: 0=start of the film, 1=end of the film.}
 \item{\code{FRAME_WIDTH}: Width in pixels of the frames in the video stream.}
 \item{\code{FRAME_HEIGHT}: Height in pixels of the frames in the video stream.}
 \item{\code{FPS}: Frame rate in frames per second.}
 \item{\code{FOURCC}: 4-character \href{http://www.fourcc.org/codecs.php}{FOURCC} code of the codec}
 \item{\code{FRAME_COUNT}: Number of frames in the video file.}
}
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{Video}}, \code{\link{video}}
}

