% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R, R/VideoClass.R,
%   R/VideoWriterClass.R, R/generic.R
\name{setProp.Rcpp_Stream}
\alias{setProp.Rcpp_Stream}
\alias{getProp.Rcpp_Stream}
\alias{setProp.Rcpp_Video}
\alias{getProp.Rcpp_Video}
\alias{setProp.Rcpp_VideoWriter}
\alias{getProp.Rcpp_VideoWriter}
\alias{setProp}
\alias{getProp}
\title{Set/Get Video Properties}
\usage{
\method{setProp}{Rcpp_Stream}(x, property, value)

\method{getProp}{Rcpp_Stream}(x, property)

\method{setProp}{Rcpp_Video}(x, property, value)

\method{getProp}{Rcpp_Video}(x, property)

\method{setProp}{Rcpp_VideoWriter}(x, property, value)

\method{getProp}{Rcpp_VideoWriter}(x, property)

setProp(x, property, value)

getProp(x, property)
}
\arguments{
\item{x}{A \code{\link{Video}} or \code{\link{Stream}} object.}

\item{property}{A character string specifying the name of the property to
modify (see details below for a complete list).}

\item{value}{The new value of the property.}
}
\value{
\code{setProp} returns TRUE is the property was set successfully.
 \code{getProp} returns a numeric value or a character string depending on
 \code{property}.
}
\description{
Set or get the values of various properties of the
 \code{\link{Video}} or \code{\link{Stream}} object.
}
\note{
Video properties are:
\itemize{
 \item{\code{POS_MSEC}: Current position of the video file in milliseconds.}
 \item{\code{POS_FRAMES}: 0-based index of the frame to be decoded/captured next.}
 \item{\code{POS_AVI_RATIO} Relative position of the video file: 0=start of the film, 1=end of the film.}
 \item{\code{FRAME_WIDTH}: Width in pixels of the frames in the video stream.}
 \item{\code{FRAME_HEIGHT}: Height in pixels of the frames in the video stream.}
 \item{\code{FPS}: Frame rate in frames per second.}
 \item{\code{FOURCC}: 4-character \href{http://www.fourcc.org/codecs.php}{FOURCC} code of the codec}
 \item{\code{FRAME_COUNT}: Number of frames in the video file.}
}

Setting stream properties depends on a lot of things, mainly your
 operating system, the camera drivers installed on your coputer and the
 camera itself. As a consequence, setting stream values might not work at all
 with your installation.

 Stream properties are:
 \itemize{
   \item{\code{FRAME_WIDTH}: Width in pixels of the frames in the video stream.}
   \item{\code{FRAME_HEIGHT}: Height in pixels of the frames in the video stream.}
   \item{\code{BRIGHTNESS}: Brightness of the image}
   \item{\code{CONTRAST}: Contrast of the image }
   \item{\code{SATURATION}: Saturation of the image}
   \item{\code{HUE}: Hue of the image}
   \item{\code{GAIN}: Gain of the image}
   \item{\code{EXPOSURE}: Exposure}
 }
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
getProp(balloon, "FPS")
setProp(balloon, "POS_FRAMES", 25)

}
\seealso{
\code{\link{Video}}, \code{\link{video}},
\code{\link{Stream}}, \code{\link{stream}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
