% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{spatialGradient}
\alias{spatialGradient}
\title{First Order Derivatives of an Image with the Sobel Operator}
\usage{
spatialGradient(image, k_size)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{k_size}{The half-size in pixels of the kernel (default: 5).}
}
\value{
image An \code{\link{Image}} object.
}
\description{
\code{spatialGradient} calculates the first order derivative of
 an image in both x and y using a Sobel operator.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}, \code{\link{sobel}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
