% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{split}
\alias{split}
\title{Split an Image into Separate Channels}
\usage{
split(image)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}
}
\value{
A list of single channel (grayscale) \code{\link{Image}} objects.
}
\description{
\code{split} returns a list of grayscale images corresponding to
 each of the channels (green, blue, red, or alpha) of an image.
}
\note{
Color images are usually represented by 3 channels (possibly 4) in the
 following order: green (1), blue (2), red (3), and possibly alpha (4).
}
\examples{
# TODO
}
\seealso{
\code{\link{merge}} \code{\link{Image}}
}
