% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R
\name{stream}
\alias{stream}
\title{Create an Object of Class \code{Stream}}
\usage{
stream(...)
}
\arguments{
\item{...}{\code{stream} takes one argument that is a number indicating the
index of the camera to use (0 is usually the default webcam on most
computers). A \code{Stream} object can also be created without any argument,
in which case it is empty and can be populated with a stream later.}
}
\value{
A \code{\link{Stream}} object.
}
\description{
Function for creating \code{\link{Stream}} objects from video
 streams.
}
\examples{
# TODO
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
