% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R
\name{nrow.Rcpp_Stream}
\alias{nrow.Rcpp_Stream}
\alias{ncol.Rcpp_Stream}
\title{The Number of Rows/Columns of a Stream}
\usage{
nrow.Rcpp_Stream(x)

ncol.Rcpp_Stream(x)
}
\arguments{
\item{x}{A \code{\link{Stream}} object.}
}
\value{
A numeric value.
}
\description{
\code{nrow} and \code{ncol} return the number of rows and columns
 present in a \code{\link{Stream}} object.
}
\examples{
\dontrun{
live <- stream(0)
nrow(live)
ncol(live)
release(live)
}

}
\seealso{
\code{\link[=dim.Rcpp_Stream]{dim}} which returns \emph{all}
 dimensions.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
