% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoStackClass.R
\name{[[.VideoStack}
\alias{[[.VideoStack}
\alias{[[<-.VideoStack}
\alias{[.VideoStack}
\alias{[<-.VideoStack}
\title{Extract or Replace Videos in Video Stacks}
\usage{
\method{[[}{VideoStack}(x, i)

\method{[[}{VideoStack}(x, i) <- value

\method{[}{VideoStack}(x, i)

\method{[}{VideoStack}(x, i) <- value
}
\arguments{
\item{x}{An \code{\link{VideoStack}} object.}

\item{i}{An index specifying the element to extract or replace. Indices are
numeric vectors which values are coerced to integer as by
\code{\link{as.integer}} (and hence truncated towards zero) or logical
vectors which are recycled if necessary to match the dimensions of the stack.}

\item{value}{A \code{\link{Video}} object or a vector of \code{\link{Video}}
objects to replace the \code{\link{Video}} objects in the stack, or a
\code{NULL} value to remove the \code{\link{Video}} objects in the stack.}
}
\description{
Operators acting on \code{\link{VideoStack}} objects to extract
 or replace the \code{\link{Video}} objects they contain.
}
\examples{
path <- system.file("sample_vid/Balloon.mp4", package = "Rvision")
vid <- video(path)
balloonStack <- videoStack(path)
balloonStack[[2]] <- vid

}
\seealso{
\code{\link{VideoStack}}, \code{\link{Video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
