% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{subImage}
\alias{subImage}
\title{Extract Subimage}
\usage{
subImage(x, row, column, width, height)
}
\arguments{
\item{x}{An \code{\link{Image}} object.}

\item{row}{The index of the upper row of the subimage.}

\item{column}{The index of the leftmost column of the subimage.}

\item{width}{The width of the subimage.}

\item{height}{The height of the subimage.}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{subImage} extracts a portion of an \code{\link{Image}} and
 returns it as an \code{\link{Image}} object.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
