% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sum}
\alias{sum}
\title{Sum Generic for additional arguments}
\usage{
sum(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{is an object of class \code{Rcpp_Image}.}

\item{...}{further arguments passed to summary methods}

\item{na.rm}{logical: should missing values be removed?}
}
\description{
Overloaded Sum to pass additional arguments
}
