% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ximgproc.R
\name{thinning}
\alias{thinning}
\title{Thinning}
\usage{
thinning(image, method = "Zhang-Suen", target = "new")
}
\arguments{
\item{image}{An 8-bit (8U) single-channel, binary \code{\link{Image}} object.}

\item{method}{A string indicating the binarization method to use. It can be
any of the following:
\itemize{
 \item{"Zhang-Suen (the default)"}
 \item{"Guo-Hall"}
}}

\item{target}{The location where the results should be stored. It can take 3
values:
\describe{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{image} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  if \code{target} does not have the same dimensions, number of channels, and
  bit depth as \code{image}, an error may be thrown.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{image} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
\code{thinning} applies a binary blob thinning operation to
 achieve a skeletonization of a binary \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_gray <- changeColorSpace(balloon, "GRAY")
balloon_th <- niBlackThreshold(balloon_gray)
skel <- thinning(balloon_th)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
