% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{tile}
\alias{tile}
\title{Repeat Image}
\usage{
tile(image, ny = 1, nx = 1, target = "new")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{ny}{A non-zero, positive integer specifying how many times the image is
repeated along the vertical axis.}

\item{nx}{A non-zero, positive integer specifying how many times the image is
repeated along the horizontal axis.}

\item{target}{The location where the results should be stored. It can take 2
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  if \code{target} does not the appropriate dimensions, number of channels,
  or bitdepth, it will be coerced automatically into the appropriate format.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target} is an \code{\link{Image}} object, the function
 returns nothing and modifies that \code{\link{Image}} object in place.
}
\description{
\code{tile} duplicates an \code{\link{Image}} object one or more
 times along each of its two axes.
}
\examples{
rnd <- zeros(360, 1, 1)
randu(rnd)
rnd_tiled <- tile(rnd, 1, 640)

}
\seealso{
\code{\link{Image}}, \code{\link{reduce}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
