% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R
\name{timelapse}
\alias{timelapse}
\title{Make Timelapse from \code{Stream} Object}
\usage{
timelapse(x, outputFolder, interval = 1, duration = Inf, format = "png")
}
\arguments{
\item{x}{The \code{Stream} object to use.}

\item{outputFolder}{The path to the folder where the timelapse images will be
saved. If it does not exist, it will be created. Note: the function will
overwrite files present in this folder if they have the same names as the
timelapse images.}

\item{interval}{The interval in seconds between two successive images
(default: 1).}

\item{duration}{The duration in seconds of the timelapse. If infinite (the
default), the timelapse will run until the user interrupts the function
manually.}

\item{format}{A character string corresponding to the format of the images
(default: "png").}
}
\value{
This function does not return anything. It saves captured images in
 \code{outputFolder}.
}
\description{
Generates a timelapse sequence from a \code{Stream} object with
 a given duration and interval between images.
}
\examples{
\dontrun{
live <- stream(0)
timelapse(live, "./out")
release(live)
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
