% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib3d.R
\name{undistortPoints}
\alias{undistortPoints}
\title{Transform Coordinates to Compensate for Lens Distortion}
\usage{
undistortPoints(
  points,
  camera_matrix,
  dist_coefs,
  new_camera_matrix = camera_matrix
)
}
\arguments{
\item{points}{A 2xN matrix of X/Y coordinates.}

\item{camera_matrix}{A 3x3 camera intrinsic matrix as returned by
\code{\link{calibrateCamera}}.}

\item{dist_coefs}{A single row matrix with 4, 5, 8, 12 or 14 elements as
returned by \code{\link{calibrateCamera}}.}

\item{new_camera_matrix}{A 3x3 camera intrinsic matrix as returned by
\code{\link{getOptimalNewCameraMatrix}} if you chose to execute this
optional step (default: \code{camera_matrix}).}
}
\value{
A 2xN matrix of transformed X/Y coordinates.
}
\description{
\code{undistortPoints} transforms a set of coordinates
 representing points in an image to compensate for radial and tangential lens
 distortion.
}
\examples{
# See the help vignette:
\dontrun{ vignette("z8_calib", package = "Rvision") }

}
\seealso{
\code{\link{undistort}}, \code{\link{calibrateCamera}},
 \code{\link{getOptimalNewCameraMatrix}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
