% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\docType{class}
\name{Video-class}
\alias{Rcpp_Video}
\alias{Video}
\alias{Video-class}
\title{An S4 class containing an OpenCV video}
\usage{
Video(...)
}
\arguments{
\item{...}{\code{Video} takes one argument that is a character string
indicating the path to the video file. A \code{Video} object can also be
created without any argument, in which case it is empty and can be populated
with a video later.}
}
\value{
A \code{Video} object.
}
\description{
\code{Video} objects contains an \href{http://opencv.org/}{OpenCV}
 video that originates from a video file.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{Image}}, \code{\link{Stream}}
}

