% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{video}
\alias{video}
\title{Create an Object of Class \code{Video}}
\usage{
video(...)
}
\arguments{
\item{...}{\code{video} takes one argument that is a character string
indicating the path to the video file. A \code{Video} object can also be
created without any argument, in which case it is empty and can be populated
with a video later.}
}
\value{
A \code{\link{Video}} object.
}
\description{
Function for creating \code{\link{Video}} objects from video
 files.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
