% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{nrow.Rcpp_Video}
\alias{nrow.Rcpp_Video}
\alias{ncol.Rcpp_Video}
\alias{nframes}
\title{The Number of Rows/Columns/Frames of a Video}
\usage{
nrow.Rcpp_Video(x)

ncol.Rcpp_Video(x)

nframes(x)
}
\arguments{
\item{x}{A \code{\link{Video}} object.}
}
\value{
A numeric value.
}
\description{
nrow, ncol and nframes return the number of rows, columns or
 frames present in a \code{\link{Video}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
nrow(balloon)
ncol(balloon)
nframes(balloon)

}
\seealso{
\code{\link[=dim.Rcpp_Video]{dim}} which returns \emph{all}
 dimensions.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
