% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoStackClass.R
\docType{class}
\name{VideoStack-class}
\alias{VideoStack-class}
\alias{VideoStack}
\title{An S4 Class Containing a Video Stack}
\description{
A \code{VideoStack} object contains a stack of
 \code{\link{Video}} objects with similar dimensions and frame rate.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list of \code{\link{Video}} objects.}

\item{\code{nframes}}{A vector of the number of frames in each video of the stack.}
}}

\seealso{
\code{\link{Video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
