% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoStackClass.R
\name{videoStack}
\alias{videoStack}
\title{Create an Object of Class \code{VideoStack}}
\usage{
videoStack(..., api = "ANY")
}
\arguments{
\item{...}{Character strings (separately or in a vector or list), each
corresponding to the path to a video file, or \code{\link{Video}} objects
(separately or in a vector or list). All videos must have the same
dimensions. If left empty, an empty \code{\link{VideoStack}} object will be
created and videos can be added to it later.}

\item{api}{A character string corresponding to the API to use for reading the
video from a file (see Note; default: "ANY").}
}
\value{
A \code{\link{VideoStack}} object.
}
\description{
Function for creating \code{\link{VideoStack}} objects from
 multiple video files.
}
\note{
Hereafter is a list of all supported APIs. Note that not all APIs will
 be available on your computer (actually most of them will not be).
 \itemize{
  \item{"ANY": }{automatically select an API.}
  \item{"VFW": }{Video For Windows.}
  \item{"V4L", "V4L2": }{Video For Linux.}
  \item{"FIREWIRE", "FIREWARE", "IEEE1394", "DC1394", "CMU1394": }{IEEE 1394 drivers.}
  \item{"QT": }{Quicktime.}
  \item{"UNICAP": }{Unicap drivers.}
  \item{"DSHOW": }{DirectShow.}
  \item{"PVAPI": }{PvAPI, Prosilica GigE SDK.}
  \item{"OPENNI": }{OpenNI (for Kinect).}
  \item{"OPENNI_ASUS": }{OpenNI (for Asus Xtion).}
  \item{"XIAPI": }{XIMEA Camera API.}
  \item{"AVFOUNDATION": }{AVFoundation framework for iOS and OSX > Lion.}
  \item{"GIGANETIX": }{Smartek Giganetix GigEVisionSDK.}
  \item{"MSMF": }{Microsoft Media Foundation.}
  \item{"WINRT": }{Microsoft Windows Runtime using Media Foundation.}
  \item{"INTELPERC": }{Intel Perceptual Computing SDK.}
  \item{"OPENNI2": }{OpenNI2 (for Kinect).}
  \item{"OPENNI2_ASUS": }{OpenNI2 (for Asus Xtion and Occipital Structure sensors).}
  \item{"GPHOTO2": }{gPhoto2 connection.}
  \item{"GSTREAMER": }{GStreamer.}
  \item{"FFMPEG": }{FFMPEG library.}
  \item{"IMAGES": }{OpenCV Image Sequence.}
  \item{"ARAVIS": }{Aravis SDK.}
  \item{"OPENCV_MJPEG": }{Built-in OpenCV MotionJPEG codec.}
  \item{"INTEL_MFX": }{Intel MediaSDK.}
 }
}
\examples{
path <- system.file("sample_vid/Balloon.mp4", package = "Rvision")
balloonStack <- videoStack(path, path)

}
\seealso{
\code{\link{Video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
