% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoStackClass.R
\name{nrow.VideoStack}
\alias{nrow.VideoStack}
\alias{ncol.VideoStack}
\alias{nframes.VideoStack}
\title{The Number of Rows/Columns/Frames of a VideoStack}
\usage{
nrow.VideoStack(x)

nrow.VideoStack(x)

\method{nframes}{VideoStack}(x)
}
\arguments{
\item{x}{A \code{\link{VideoStack}} object.}
}
\value{
A numeric value.
}
\description{
nrow, ncol and nframes return the number of rows, columns or
 frames present in a \code{\link{VideoStack}} object.
}
\examples{
path <- system.file("sample_vid/Balloon.mp4", package = "Rvision")
balloonStack <- videoStack(path, path)
nrow(balloonStack)
ncol(balloonStack)
nframes(balloonStack)

}
\seealso{
\code{\link[=dim.VideoStack]{dim}} which returns \emph{all}
 dimensions.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
