% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoWriterClass.R
\docType{class}
\name{VideoWriter-class}
\alias{VideoWriter-class}
\alias{videoWriter}
\alias{Rcpp_VideoWriter}
\alias{VideoWriter}
\title{An S4 Class Containing an OpenCV Video Writer}
\usage{
videoWriter(
  outputFile,
  fourcc,
  fps,
  height,
  width,
  isColor = TRUE,
  api = "ANY"
)
}
\arguments{
\item{outputFile}{An character string corresponding to the path to an output
file.}

\item{fourcc}{A 4-character string corresponding to the fourcc code of the
codec to be used. A list of fourcc codes can be obtained at this archived
page of the fourcc site \href{https://www.fourcc.org/codecs.php}{https://www.fourcc.org/codecs.php}.
Alternatively, the integer value corresponding to a fourcc code.}

\item{fps}{A numeric value corresponding to the framerate of the output video.}

\item{height}{An integer value corresponding to the height of the video in
pixels.}

\item{width}{An integer value corresponding to the width of the video in
pixels.}

\item{isColor}{A logical indicating whether the output video is a color
(default: TRUE) or grayscale (FALSE) video.}

\item{api}{A character string corresponding to the API to use for reading the
video from the file (see Note; default: "ANY").}
}
\value{
A \code{\link{VideoWriter}} object.
}
\description{
A \code{VideoWriter} object contains an \href{http://opencv.org/}{OpenCV}
 video writer to an output file.

Function for creating \code{\link{VideoWriter}} objects.
}
\section{Slots}{

\describe{
\item{\code{dim,ncol,nrow}}{Functions returning the dimensions of the object.}

\item{\code{codec}}{Function returning the codec of the object.}

\item{\code{fps}}{Function returning the frame rate of the object.}

\item{\code{api}}{Function returning the api use to write frames to the output file.}

\item{\code{output}}{Function returning the path to the output file.}

\item{\code{get,set}}{Functions to access and set internal properties of the object.}

\item{\code{open,isOpened}}{Functions to open a new video write or check the status
of the video writer.}

\item{\code{write}}{Function to write a frame to the output file.}

\item{\code{release}}{Function to release the object from memory.}
}}

\note{
Hereafter is a list of all supported APIs. Note that not all APIs will
 be available on your computer (actually most of them will not be).
 \describe{
  \item{"ANY": }{automatically select an API.}
  \item{"VFW": }{Video For Windows.}
  \item{"V4L", "V4L2": }{Video For Linux.}
  \item{"FIREWIRE", "FIREWARE", "IEEE1394", "DC1394", "CMU1394": }{IEEE 1394 drivers.}
  \item{"QT": }{Quicktime.}
  \item{"UNICAP": }{Unicap drivers.}
  \item{"DSHOW": }{DirectShow.}
  \item{"PVAPI": }{PvAPI, Prosilica GigE SDK.}
  \item{"OPENNI": }{OpenNI (for Kinect).}
  \item{"OPENNI_ASUS": }{OpenNI (for Asus Xtion).}
  \item{"XIAPI": }{XIMEA Camera API.}
  \item{"AVFOUNDATION": }{AVFoundation framework for iOS and OSX > Lion.}
  \item{"GIGANETIX": }{Smartek Giganetix GigEVisionSDK.}
  \item{"MSMF": }{Microsoft Media Foundation.}
  \item{"WINRT": }{Microsoft Windows Runtime using Media Foundation.}
  \item{"INTELPERC": }{Intel Perceptual Computing SDK.}
  \item{"OPENNI2": }{OpenNI2 (for Kinect).}
  \item{"OPENNI2_ASUS": }{OpenNI2 (for Asus Xtion and Occipital Structure sensors).}
  \item{"GPHOTO2": }{gPhoto2 connection.}
  \item{"GSTREAMER": }{GStreamer.}
  \item{"FFMPEG": }{FFMPEG library.}
  \item{"IMAGES": }{OpenCV Image Sequence.}
  \item{"ARAVIS": }{Aravis SDK.}
  \item{"OPENCV_MJPEG": }{Built-in OpenCV MotionJPEG codec.}
  \item{"INTEL_MFX": }{Intel MediaSDK.}
 }
}
\examples{
\dontrun{
writer <- videoWriter("test.mp4", "H264", 25, 1080, 1920)
release(writer)
}

}
\seealso{
\code{\link{videoWriter}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
