% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoWriterClass.R
\name{videoWriter}
\alias{videoWriter}
\title{Create an object of class \code{VideoWriter}}
\usage{
videoWriter(...)
}
\arguments{
\item{...}{\code{videoWriter} takes six arguments:
\describe{
  \item{output_file}{A character string indicating the path to the output file.}
  \item{fourcc}{A 4-character string corresponding to the fourcc code of the
    codec to be used. A list of fourcc codes can be obtained at
    \href{http://www.fourcc.org/codecs.php}{http://www.fourcc.org/codecs.php}.}
  \item{fps}{A numeric value corresponding to the framerate of the output
    video.}
  \item{height, width}{Integer values corresponding to the height and width
    of the video in pixels.}
  \item{is_color}{A logical indicating whether the output video is a color
    (TRUE) or grayscale (FALSE) video.}
}
A \code{VideoWriter} object can also be created without any argument,
in which case it is empty and can be populated with a video writer later.}
}
\value{
A \code{\link{VideoWriter}} object.
}
\description{
Function for creating \code{\link{VideoWriter}} objects.
}
\examples{
# TODO
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
