% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoWriterClass.R
\name{nrow.Rcpp_VideoWriter}
\alias{nrow.Rcpp_VideoWriter}
\alias{ncol.Rcpp_VideoWriter}
\title{The Number of Rows/Columns of a Video Writer}
\usage{
nrow.Rcpp_VideoWriter(x)

ncol.Rcpp_VideoWriter(x)
}
\arguments{
\item{x}{A \code{\link{VideoWriter}} object.}
}
\value{
A numeric value.
}
\description{
nrow, ncol return the number of rows or columns present in a
 \code{\link{VideoWriter}} object.
}
\examples{
\dontrun{
writer <- videoWriter("test.mp4", "H264", 25, 1080, 1920)
nrow(writer)
ncol(writer)
release(writer)
}

}
\seealso{
\code{\link[=dim.Rcpp_VideoWriter]{dim}} which returns \emph{all}
 dimensions.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
