% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{watershed}
\alias{watershed}
\title{Image Segmentation Using the Watershed Algorithm}
\usage{
watershed(image, markers)
}
\arguments{
\item{image}{An 8-bit (8U), BGR \code{\link{Image}} object.}

\item{markers}{A signed 32-bit (32S) single-channel (GRAY) \code{\link{Image}}
object (see Details).}
}
\value{
This function does not return anything. It modifies \code{markers} in
 place.
}
\description{
\code{watershed} implements one of the variants of watershed,
 non-parametric marker-based segmentation algorithm, described in Meyer (1992).
}
\details{
Before passing \code{image} to the function, you have to roughly
 outline the desired regions in the \code{markers} image with positive (>0)
 indices. So, every region is represented as one or more connected components
 with the pixel values 1, 2, 3, and so on. Such markers can be retrieved from
 a binary mask using \code{\link{findContours}}. The markers are "seeds" of
 the future image regions. All the other pixels in \code{markers}, whose
 relation to the outlined regions is not known and should be defined by the
 algorithm, should be set to 0's. In the function output, each pixel in
 \code{markers} is set to a value of the "seed" components or to -1 at
 boundaries between the regions.
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_gray <- changeColorSpace(dots, "GRAY")
dots_bin <- dots_gray < 200
contours <- findContours(dots_bin)
rows <- contours$contours$y
cols <- contours$contours$x
colors <- contours$contours$id + 2
markers <- array(0.0, dim = c(nrow(dots_bin), ncol(dots_bin), 1))
markers[1:2, 1:2, 1] <- 1
markers[cbind(rows, cols, 1)] <- colors
markers <- image(markers)
changeBitDepth(markers, "32S", in_place = TRUE)
watershed(dots, markers)
plot(markers)

}
\references{
Meyer F. Color image segmentation. 1992 International Conference
 on Image Processing. 1992. Available:
 https://ieeexplore.ieee.org/abstract/document/785528/
}
\seealso{
\code{\link{findContours}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
