% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{write.Image}
\alias{write.Image}
\title{Image Output}
\usage{
write.Image(image, file)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{file}{A character string naming the path to a file.}
}
\value{
A logical indicating whether writing was successful (TRUE) or not
 (FALSE).
}
\description{
Writes the content of an \code{\link{Image}} object to a file.
}
\note{
The function will guess the format of the output file using the file
 extension provided by the user.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{Image}}, \code{\link{image}}
}

