% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoWriterClass.R
\name{writeFrame}
\alias{writeFrame}
\title{Write Frame to Output Video}
\usage{
writeFrame(x, frame)
}
\arguments{
\item{x}{A \code{\link{VideoWriter}} object.}

\item{frame}{An \code{\link{Image}} object.}
}
\value{
This function does not return anything.
}
\description{
Write an \code{\link{Image}} object to an output video using a
 \code{\link{VideoWriter}} object.
}
\examples{
# TODO
}
\seealso{
\code{\link{VideoWriter}}, \code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
