% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{writeMulti}
\alias{writeMulti}
\title{Wrtie a Multi-Page Image}
\usage{
writeMulti(x, img_list, overwrite = FALSE)
}
\arguments{
\item{x}{A character string naming the path to a multi-page image file.}

\item{img_list}{A list of \code{\link{Image}} objects.}

\item{overwrite}{Should the file be overwritten if it already exists?
(default: FALSE)}
}
\value{
A message indicating whether writing was successful or not.
}
\description{
\code{writeMulti} writes a list of \code{\link{Image}} objects
 to a multi-page image.
}
\examples{
balloon <- readMulti(system.file("sample_img/multipage.tif", package = "Rvision"))
\dontrun{
writeMulti("~/Desktop/balloon.tiff", balloon)
}


}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
