% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoWriterClass.R
\name{writerOuput}
\alias{writerOuput}
\title{Output File of a Video Writer}
\usage{
writerOuput(x)
}
\arguments{
\item{x}{A \code{\link{VideoWriter}} object.}
}
\value{
A character string corresponding to the address of the output file on
 the hard drive.
}
\description{
Retrieve the output file of a \code{\link{VideoWriter}} object.
}
\examples{
\dontrun{
writer <- videoWriter("test.mp4", "H264", 25, 1080, 1920)
writerOuput(writer)
release(writer)
}

}
\seealso{
\code{\link{VideoWriter}}, \code{\link{videoWriter}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
