% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{addWeighted}
\alias{addWeighted}
\title{Weighted Sum of Two Images}
\usage{
addWeighted(image1, image2, weight = c(0.5, 0.5), in_place = FALSE)
}
\arguments{
\item{image1}{An \code{\link{Image}} object.}

\item{image2}{An \code{\link{Image}} object.}

\item{weight}{A 2-element vector of the respective weight of each image
(default: c(0.5, 0.5)). If the two weights do not add up to 1, they will be
rescaled accordingly.}

\item{in_place}{A logical indicating whether the change should be applied to
the image itself (TRUE, faster but destructive) or to a copy of it (FALSE,
the default, slower but non destructive).}
}
\value{
An \code{\link{Image}} object if \code{in_place=FALSE}. Otherwise, it
 returns nothing and modifies \code{image1} in place.
}
\description{
This function computes the weighted sum of two
 \code{\link{Image}} objects.
}
\examples{
balloon1 <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon2 <- image(system.file("sample_img/balloon2.png", package = "Rvision"))
plot(addWeighted(balloon1, balloon2))

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
