% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{addWeighted}
\alias{addWeighted}
\title{Weighted Sum of Two Images}
\usage{
addWeighted(image1, image2, weight = c(0.5, 0.5))
}
\arguments{
\item{image1}{An \code{\link{Image}} object.}

\item{image2}{An \code{\link{Image}} object.}

\item{weight}{A 2-element vector of the respective weight of each image
(default: c(0.5, 0.5)). If the two weights do not add up to 1, they will be
rescaled accordingly.}
}
\value{
An \code{\link{Image}} object.
}
\description{
This function computes the weighted sum of two
 \code{\link{Image}} objects.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
