% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{addWeighted}
\alias{addWeighted}
\title{Weighted Sum of Two Images}
\usage{
addWeighted(e1, e2, weight = c(0.5, 0.5), target = "new")
}
\arguments{
\item{e1}{An \code{\link{Image}} object.}

\item{e2}{An \code{\link{Image}} object.}

\item{weight}{A 2-element vector of the respective weight of each image
(default: c(0.5, 0.5)). If the two weights do not add up to 1, they will be
rescaled accordingly.}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{e1} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{e1} but will replace that of \code{target}. Note that
  if \code{target} does not have the same dimensions, number of channels, and
  bit depth as \code{e1}, nothing will be stored.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{e1} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
This function computes the weighted sum of two
 \code{\link{Image}} objects.
}
\examples{
balloon1 <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon2 <- image(system.file("sample_img/balloon2.png", package = "Rvision"))
balloon_weighted <- addWeighted(balloon1, balloon2)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
