% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{as.array.Rcpp_Image}
\alias{as.array.Rcpp_Image}
\alias{as.matrix.Rcpp_Image}
\title{Convert Image to Array or Matrix}
\usage{
\method{as.array}{Rcpp_Image}(x, ...)
}
\arguments{
\item{x}{An \code{\link{Image}} object.}

\item{...}{Ignored.}
}
\value{
A matrix or array of the same dimensions as the \code{\link{Image}}
 object.
}
\description{
Attempts to turn its argument into a matrix or an array.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_mat <- as.matrix(balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{matrix}}, \code{\link{array}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
