% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{bilateralFilter}
\alias{bilateralFilter}
\title{Edge-Preserving Noise Reduction with a Bilateral Filter}
\usage{
bilateralFilter(image, d = 5, sigma_color = 25, sigma_space = 25)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{d}{The diameter in pixels of the filter neighborhood (default: 5).}

\item{sigma_color}{The filter standard deviation in the color space
(see Note; default: 25).}

\item{sigma_space}{The filter standard deviation in the coordinate space
(see Note; default: 25).}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{bilateralFilter} applies the bilateral filter to an image.
 This filter can reduce unwanted noise very well while keeping edges fairly
 sharp. However, it is very slow compared to most filters.
}
\note{
A larger value of \code{sigma_color} means that farther colors within
 the pixel neighborhood will be mixed together, resulting in larger areas of
 semi-equal color.

A larger value of \code{sigma_space} means that farther pixels will
 influence each other as long as their colors are close enough. When
 \code{d > 0}, it specifies the neighborhood size regardless of
 \code{sigma_space}. Otherwise, \code{d} is proportional to \code{sigma_space}.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_noisy <- balloon + image(array(sample(0:30, nrow(balloon) * ncol(balloon), replace = TRUE),
                                       dim = c(nrow(balloon), ncol(balloon), 3)))
plot(balloon_noisy)
balloon_bilateral <- bilateralFilter(balloon_noisy, 25)
plot(balloon_bilateral)

}
\seealso{
\code{\link{Image}}, \code{\link{gaussianBlur}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
