% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{blur}
\alias{blur}
\title{Blurs an Image Using a Normalized Box Filter}
\usage{
blur(image, k_height = 5, k_width = 5, in_place = FALSE)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{k_height}{The half-height in pixels of the kernel (default: 5).}

\item{k_width}{The half-width in pixels of the kernel (default: 5).}

\item{in_place}{A logical indicating whether the change should be applied to
the image itself (TRUE, faster but destructive) or to a copy of it (FALSE,
the default, slower but non destructive).}
}
\value{
An \code{\link{Image}} object if \code{in_place=FALSE}. Otherwise, it
 returns nothing and modifies \code{image} in place.
}
\description{
\code{blur} convolves the source image with the specified
 normalized box kernel (a matrix of 1s divided by the number of pixels in the
 kernel). The result is a blurred version of the source image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_blur <- blur(balloon, 11, 11)
plot(balloon_blur)

}
\seealso{
\code{\link{Image}}, \code{\link{boxFilter}}, \code{\link{gaussianBlur}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
