% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{boxFilter}
\alias{boxFilter}
\title{Blurs an Image Using a Box Filter}
\usage{
boxFilter(image, k_height = 5, k_width = 5)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{k_height}{The half-height in pixels of the kernel.}

\item{k_width}{The half-width in pixels of the kernel.}
}
\value{
image An \code{\link{Image}} object.
}
\description{
\code{gaussianBlur} convolves the source image with the
 specified box kernel (a matrix of 1s). The result is a blurred version of
 the source image.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}, \code{\link{gaussianBlur}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
