% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{changeBitDepth}
\alias{changeBitDepth}
\title{Convert image bit depth}
\usage{
changeBitDepth(image, bitdepth)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{bitdepth}{A scalar corresponding to the colorspace the image should
be converted to. Options are 8 and 16, for 8 and 16 bits respectively.}
}
\value{
An \code{\link{Image}} object.
}
\description{
This function takes an \code{\link{Image}} object and modifies
 its bit depth (e.g. from 16 bits to 8 bits).
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
