% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{changeBitDepth}
\alias{changeBitDepth}
\title{Convert image bit depth}
\usage{
changeBitDepth(image, bitdepth, scale = 1)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{bitdepth}{A scalar corresponding to the colorspace the image should
be converted to. Options are "8U", "8S", "16U", "16S", "32S", and "32F"
respectively.}

\item{scale}{A scaling factor (default: 1).}
}
\value{
An \code{\link{Image}} object.
}
\description{
This function takes an \code{\link{Image}} object and modifies
 its bit depth (e.g. from 16 bits to 8 bits).
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_16 <- changeBitDepth(balloon, 16)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
