% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{changeBitDepth}
\alias{changeBitDepth}
\title{Convert Image to New Bit Depth}
\usage{
changeBitDepth(image, bitdepth, scale = 1, target = "new", in_place = NULL)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{bitdepth}{A string corresponding to the bit depth the image should
be converted to. Options are "8U", "8S", "16U", "16S", "32S", "32F", and
"64F". Converting from a higher bit depth to a lower one (e.g., "16U" to
"8U" may result in data loss).}

\item{scale}{A scaling factor (default: 1).}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{image} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  \code{target} must have the same dimensions as \code{image}.}
}}

\item{in_place}{Deprecated. Use \code{target} instead.}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{image} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
This function takes an \code{\link{Image}} object and modifies
 its bit depth (e.g. from 16 bits to 8 bits).
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_16 <- changeBitDepth(balloon, "16U")

}
\seealso{
\code{\link{Image}}, \code{\link{changeColorSpace}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
