% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{changeColorSpace}
\alias{changeColorSpace}
\title{Convert image between colorspaces}
\usage{
changeColorSpace(image, colorspace)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{colorspace}{A string corresponding to the colorspace the image should
be converted to. Options are "BGR" (Blue Green Red image), "BGRA" (BGR image
with Alpha channel), and "GRAY" (grayscale image). Converting from a Bayer mosaic to BGR is also possible with "BayerBG2BGR","BayerGB2BGR", "BayerRG2BGR", "BayerGR2BGR".}
}
\value{
An \code{\link{Image}} object.
}
\description{
This function takes an \code{\link{Image}} object and converts
 it to another colorspace (e.g BGR to grayscale).
}
\examples{
# TODO

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
