% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{changeColorSpace}
\alias{changeColorSpace}
\title{Convert Image to New Color Space}
\usage{
changeColorSpace(image, colorspace, target = "new", in_place = NULL)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{colorspace}{A string corresponding to the color space the image should
be converted to. Not all conversions between color spaces are possible. All
available color space conversions can be found in \code{\link{cc_table}}.}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{image} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  \code{target} must have the same dimensions as \code{image}.}
}}

\item{in_place}{Deprecated. Use \code{target} instead.}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{image} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
This function takes an \code{\link{Image}} object and converts
 it to another color space (e.g BGR to grayscale).
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
grey_balloon <- changeColorSpace(balloon, "GRAY")

}
\seealso{
\code{\link{Image}}, \code{\link{changeBitDepth}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
