% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R, R/VideoWriterClass.R, R/generic.R
\name{codec.Rcpp_Video}
\alias{codec.Rcpp_Video}
\alias{codec.Rcpp_VideoWriter}
\alias{codec}
\title{Codec of a Video}
\usage{
\method{codec}{Rcpp_Video}(x)

\method{codec}{Rcpp_VideoWriter}(x)

codec(obj, property)
}
\arguments{
\item{x}{A \code{\link{Video}} object.}
}
\value{
A character string corresponding to the
 \href{http://www.fourcc.org/codecs.php}{FOURCC} code of the codec.
}
\description{
Retrieve the codec of a \code{\link{Video}} object.
}
\examples{
# TODO
}
\seealso{
\code{\link{Video}}, \code{\link{video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
