% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{codec}
\alias{codec}
\title{Codec of a Video}
\usage{
codec(x)
}
\arguments{
\item{x}{A \code{\link{Video}} object.}
}
\value{
A character string corresponding to the
 \href{http://www.fourcc.org/codecs.php}{FOURCC} code of the codec.
}
\description{
Retrieve the codecof a \code{\link{Video}} object.
}
\examples{
# TODO
}
\seealso{
\code{\link{Video}}, \code{\link{video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
