% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{dim.Rcpp_Image}
\alias{dim.Rcpp_Image}
\title{Dimensions of an Image}
\usage{
\method{dim}{Rcpp_Image}(x)
}
\arguments{
\item{x}{An \code{\link{Image}} object.}
}
\value{
A vector with 3 values corresponding to the number of rows, columns
 and channels of the image (in this order).
}
\description{
Retrieve the dimensions an \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
dim(balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
