% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{distanceTransform}
\alias{distanceTransform}
\title{Distance Transform}
\usage{
distanceTransform(image, distance_type = "L1", mask_size = 3, target = "new")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{distance_type}{A character string indicating the type of distance
to be calculated. It can be any of the following:
\itemize{
 \item{"L1" (the default):}{\code{distance = |x1-x2| + |y1-y2|}.}
 \item{"L2":}{the simple euclidean distance.}
 \item{"C":}{\code{distance = max(|x1-x2|,|y1-y2|)}.}
 \item{"L12":}{L1-L2 metric. \code{distance = 2(sqrt(1+x*x/2) - 1))}.}
 \item{"FAIR":}{\code{distance = c^2(|x|/c-log(1+|x|/c)), c = 1.3998}.}
 \item{"WELSCH":}{\code{distance = c^2/2(1-exp(-(x/c)^2)), c = 2.9846}.}
 \item{"HUBER":}{\code{distance = |x|<c ? x^2/2 : c(|x|-c/2), c=1.345}.}
}}

\item{mask_size}{A numeric value indicating the size of the distance
transform mask. It can be any of the following:
\itemize{
 \item{0:}{used only to indicate the Felzenszwalb algorithm when
  \code{distance_type = "L2"}.}
 \item{3 (the default):}{3x3 mask.}
 \item{5:}{5x5 mask.}
}}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{image} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  \code{target} must have the same dimensions as \code{image}, must have a
  single channel, and its bit depth must be either "8U" or "32F".}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{image} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.

An \code{\link{Image}} object.
}
\description{
\code{distanceTransform} calculates the distance to the closest
 zero pixel for each pixel of the source image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
changeColorSpace(balloon, "GRAY", target = "self")
bin <- balloon < 200
dst <- distanceTransform(bin)

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
