% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawCircle}
\alias{drawCircle}
\title{Draw Circles on an \code{\link{Image}}}
\usage{
drawCircle(image, x, y, radius, color = "red", thickness = 1)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{x}{A numeric value or vector representing the x coordinates of the
centers of each circle.}

\item{y}{A numeric value or vector representing the y coordinates of the
centers of each circle.}

\item{radius}{A numeric value or vector representing the radii of each circle.}

\item{color}{A value or vector of any kind of R color specification compatible
with \code{\link{col2bgr}} representing the color of each circle's outline
(default: "red").}

\item{thickness}{A numeric value or vector representing the thickness in
pixels of each circle's outline (default: 1). If negative, then a filled
rectangle is drawn.}
}
\value{
This function does not return anything. It modifies \code{image} in
 place.
}
\description{
\code{drawCirlc} draws circles over an \code{\link{Image}}
 object. This operation is destructive: it changes irreversibly the
 \code{\link{Image}} object and cannot be undone.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
drawCircle(balloon, 365, 245, 90, thickness = 3)

}
\seealso{
\code{\link{Image}}, \code{\link{drawEllipse}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
