% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{fillPoly}
\alias{fillPoly}
\title{Fill Polygon with Color in Image}
\usage{
fillPoly(image, polygon, color = "white")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{polygon}{An m x 2 matrix (or an object that can be converted to an
m x 2 matrix) or a list of m x 2 matrices, with the first column containing
the x coordinates of the polygons and the second column containing the y
coordinates of the polygons.}

\item{color}{A value or vector of any kind of R color specification compatible
with \code{\link{col2bgr}} representing the color to fill the polygon with
(default: "white").}
}
\value{
This function does not return anything. It modifies \code{image} in
 place.
}
\description{
\code{fillPoly} fills all the pixels of an image withing a
 given polygon with a given color.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
poly <- data.frame(x = c(290, 290, 440, 440), y = c(170, 325, 325, 170))
fillPoly(balloon, poly, color = "red")

}
\seealso{
\code{\link{Image}}, \code{\link{selectROI}}, \code{\link{fillConvexPoly}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
