% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter2D}
\alias{filter2D}
\title{Image Convolution with Kernel}
\usage{
filter2D(image, kernel)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{kernel}{A matrix representing the convolution kernel.}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{filter2D} applies an arbitrary linear filter to an image.
}
\note{
For color images, the same kernel is applied to each channel of the
 image. If you want to apply different kernels to each channel, first split
 the image into separate channels with the \code{\link{split}} and process
 them individually before merging them using the \code{\link{merge}} function.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}, \code{\link{split}}, \code{\link{merge}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
