% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{findNonZero}
\alias{findNonZero}
\title{Find Non-Zero Pixels in an Image}
\usage{
findNonZero(image)
}
\arguments{
\item{image}{An an 8-bit (8U) single-channel \code{\link{Image}} object.}
}
\value{
A data frame with two columns, corresponding to the x (columns) and
 y (rows) coordinates of the non-zero pixels.
}
\description{
\code{findNonZero} retrieves the locations of all non-zero
 pixels in an image.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
