% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{findNonZero}
\alias{findNonZero}
\title{Find Non-Zero Pixels in an Image}
\usage{
findNonZero(image, values = FALSE)
}
\arguments{
\item{image}{A single-channel \code{\link{Image}} object.}

\item{values}{A logical indicating whether the values of the non-zero pixels
should also be returned (default: FALSE).}
}
\value{
If \code{values=FALSE}, a matrix with two columns, corresponding to
 the x (columns) and y (rows) coordinates of the non-zero pixels. If
 \code{values=TRUE}, an additional column corresponding to the pixel values
 will also be returned.
}
\description{
\code{findNonZero} retrieves the locations of all non-zero
 pixels in an image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_gray <- changeColorSpace(balloon, "GRAY")
findNonZero(balloon_gray)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
