% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{flip}
\alias{flip}
\title{Flip an \code{\link{Image}}}
\usage{
flip(image, type = 0, target = "new", in_place = NULL)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{type}{An integer indicating the type of flipping to be performed. If
\code{type = 0} (the default), the image is flipped around its x-axis; if
\code{type = 1} (or any positive value, then it is flipped around its y-axis;
finally, if \code{type = -1} (or any negative value, then it is flipped
around both axes.)}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{image} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  if \code{target} does not have the same dimensions, number of channels, and
  bit depth as \code{image}, nothing will be stored.}
}}

\item{in_place}{Deprecated. Use \code{target} instead.}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{image} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
\code{flip} returns a flipped version of an \code{\link{Image}}
 around one or both of its axes.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_flipped <- flip(balloon, -1)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
