% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R, R/VideoWriterClass.R, R/generic.R
\name{fps.Rcpp_Video}
\alias{fps.Rcpp_Video}
\alias{fps.Rcpp_VideoWriter}
\alias{fps}
\title{Framerate of a Video}
\usage{
\method{fps}{Rcpp_Video}(obj)

\method{fps}{Rcpp_VideoWriter}(obj)

fps(obj)
}
\arguments{
\item{obj}{A \code{\link{Video}} object.}
}
\value{
A numeric value.
}
\description{
Retrieve the framerate (in frames per second) of a
 \code{\link{Video}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
fps(balloon)
release(balloon)

}
\seealso{
\code{\link{Video}}, \code{\link{video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
