% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R, R/VideoWriterClass.R, R/generic.R
\name{fps.Rcpp_Video}
\alias{fps.Rcpp_Video}
\alias{fps.Rcpp_VideoWriter}
\alias{fps}
\title{Framerate of a Video}
\usage{
\method{fps}{Rcpp_Video}(x)

\method{fps}{Rcpp_VideoWriter}(x)

fps(obj, property)
}
\arguments{
\item{x}{A \code{\link{Video}} object.}
}
\value{
A numeric value.
}
\description{
Retrieve the framerate (in frames per second) of a
 \code{\link{Video}} object.
}
\examples{
# TODO
}
\seealso{
\code{\link{Video}}, \code{\link{video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
