% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R, R/VideoWriterClass.R, R/generic.R
\name{fps.Rcpp_Video}
\alias{fps.Rcpp_Video}
\alias{fps.Rcpp_VideoWriter}
\alias{fps}
\title{Frame Rate of a Video}
\usage{
\method{fps}{Rcpp_Video}(x)

\method{fps}{Rcpp_VideoWriter}(x)

fps(x)
}
\arguments{
\item{x}{A \code{\link{Video}} or \code{\link{VideoWriter}} object.}
}
\value{
A numeric value.
}
\description{
Retrieve the frame rate (in frames per second) of a
 \code{\link{Video}} or \code{\link{VideoWriter}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
fps(balloon)

}
\seealso{
\code{\link{Video}}, \code{\link{video}}, \code{\link{VideoWriter}},
 \code{\link{videoWriter}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
