% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{frame}
\alias{frame}
\title{Video Reader Position}
\usage{
frame(video)
}
\arguments{
\item{video}{A \code{\link{Video}} object.}
}
\value{
A numeric value.
}
\description{
Retrieve the index of the frame to be read next in a
 \code{\link{Video}} object.
}
\note{
Frame index starts at 0 (i.e. the first image has index 0).
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{Video}}, \code{\link{video}}
}

