% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{histmatch}
\alias{histmatch}
\title{Histogram Matching/Specification}
\usage{
histmatch(image, reference, target = "new")
}
\arguments{
\item{image}{An \code{\link{Image}} object to transform.}

\item{reference}{An \code{\link{Image}} object which histogram will be used
as a reference to transform \code{image}.}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{image} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  if \code{target} does not have the same dimensions, number of channels,
  and bit depth as \code{image}, nothing will be stored.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{image} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
\code{histmatch} transforms an \code{\link{Image}} object so
 that its histogram matches (approximately) that of another
 \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_matched <- histmatch(dots, balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{histmatch}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
