% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\docType{class}
\name{Image-class}
\alias{Image}
\alias{Image-class}
\alias{Rcpp_Image}
\title{An S4 class containing an OpenCV image}
\format{An object of class \code{C++Class} of length 1.}
\usage{
Image
}
\description{
\code{Image} objects are the base objects of the \pkg{\link{Rvision}}
 package. They contain an \href{http://opencv.org/}{OpenCV} image that can
 originate from an image file, an array, a video file or a video stream.
 This image can be manipulated using the functions of \pkg{\link{Rvision}}.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{image}}, \code{\link{Video}}, \code{\link{Stream}}
}
\keyword{datasets}

