% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\docType{class}
\name{Image-class}
\alias{Image-class}
\alias{Image}
\alias{Rcpp_Image2}
\title{An S4 Class Containing an OpenCV Image}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
Image
}
\description{
\code{Image} objects are the base objects of the \pkg{\link{Rvision}}
 package. They contain an \href{http://opencv.org/}{OpenCV} image that can
 originate from an image file, an array, a video file or a video stream.
 This image can be manipulated using the functions of \pkg{\link{Rvision}}.
}
\section{Slots}{

\describe{
\item{\code{depth}}{A function returning the bit depth of the object.}

\item{\code{dim,ncol,nrow,nchan}}{Functions returning the dimensions of the object.}

\item{\code{space}}{A character string indicating the colorspace of the object.}

\item{\code{pget,pset}}{Functions to get and set single pixel values.}

\item{\code{toR}}{A function to convert the object to an R array.}

\item{\code{fromGPU}}{A function retrieving the object from GPU memory.}

\item{\code{toGPU}}{A function sending the object to GPU memory.}

\item{\code{GPU}}{A boolean indicating whether the object is in GPU memory.}

\item{\code{write}}{A function to write the object to a file.}
}}

\seealso{
\code{\link{image}}, \code{\link{Video}}, \code{\link{Stream}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\keyword{datasets}
