% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{inpaint}
\alias{inpaint}
\title{Reconstruct Image Region from Region Neighborhood}
\usage{
inpaint(
  image,
  mask,
  radius = 5,
  method = "NS",
  target = "new",
  in_place = NULL
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{mask}{An 8-bit single-channel \code{\link{Image}} object. The region to
be reconstructed should be white.}

\item{radius}{Radius of the circular neighborhood of each point inpainted
that is considered by the algorithm (default: 5).}

\item{method}{The inpainting method to be used. It can only be one of the
following:
\itemize{
  \item{"NS": }{Navier-Stokes based method (the default).}
  \item{"Telea": }{Alexandru Telea's method.}
}}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{image} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  \code{target} must have the same dimensions, number of channels, and bit
  depth,  as \code{image}, otherwise an error will be thrown.}
}}

\item{in_place}{Deprecated. Use \code{target} instead.}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{image} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
\code{inpaint} reconstructs the selected image area from the
 pixel near the area boundary. The function may be used to remove dust and
 scratches from a scanned photo, or to remove undesirable objects from still
 images or videos.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
mask <- zeros(nrow(balloon), ncol(balloon), 3)
poly <- data.frame(x = c(290, 290, 440, 440), y = c(170, 325, 325, 170))
fillPoly(mask, poly, color = "white")
changeColorSpace(mask, "GRAY", in_place = TRUE)
balloon_inpait <- inpaint(balloon, mask)

}
\references{
Telea, A. (2004). An image inpainting technique based on the fast
 marching method. Journal of graphics tools. doi: 10.1080/10867651.2004.10487596.
}
\seealso{
\code{\link{Image}}, \code{\link{selectROI}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
