% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{inRange}
\alias{inRange}
\title{Range Thresholding}
\usage{
inRange(image, low = 0, up = 255, target = "new")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{low}{A vector indicating the inclusive lower end of the thresholding
range. It can have as many elements as the number of channels in the image.
If it has less elements than the number of channels, it is recycled to match
the number of channels. If it has more elements than the number of channels,
the extra elements are ignored without warning (default: 0).}

\item{up}{A vector indicating the inclusive upper end of the thresholding
range. Itcan have as many elements as the number of channels in the image.
If it has less elements than the number of channels, it is recycled to match
the number of channels. If it has more elements than the number of channels,
the extra elements are ignored without warning (default: 255).}

\item{target}{The location where the results should be stored. It can take 2
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. In this case, \code{target} must be a
  single channel image with an 8U bit depth. Note that this will replace the
  content of \code{target}. }
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target} is an \code{\link{Image}} object, the function
 returns nothing and modifies that \code{\link{Image}} object in place.

An \code{\link{Image}} object.
}
\description{
\code{inRange} performs range thresholding on an
 \code{\link{Image}} object. Pixels which values are within the desired
 range are turned white while pixels which values are outside are turned
 black. This operation is performed separately on each channel for
 multi-channel images.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
bw <- inRange(balloon, low = c(10, 20, 30), up = c(120, 130, 140))

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
