% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{invert}
\alias{invert}
\title{Invert Colors}
\usage{
invert(image, in_place = FALSE)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{in_place}{A logical indicating whether the change should be applied to
the image itself (TRUE, faster but destructive) or to a copy of it (FALSE,
the default, slower but non destructive).}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{invert} returns an image which colors are the linear
 inverse of that of the original image.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_inv <- invert(balloon)
plot(balloon_inv)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
