% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{isBlob}
\alias{isBlob}
\title{Test for a Blob Object}
\usage{
isBlob(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
A logical indicating whether the object is of class
 \code{blob} (TRUE) or not (FALSE).
}
\description{
Tests whether the object is of class \code{blob}.
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
blobs <- simpleBlobDetector(invert(dots), min_threshold = 25, max_threshold = 220,
                            filter_by_area = TRUE, min_area = 200, max_area = Inf,
                            filter_by_color = FALSE)
isBlob(blobs)

}
\seealso{
\code{\link{simpleBlobDetector}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
