% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{laplacian}
\alias{laplacian}
\title{Calculates the Laplacian of an Image}
\usage{
laplacian(image, k_size = 5, scale = 1)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{k_size}{The half-size in pixels of the kernel (default: 5).}

\item{scale}{The scale factor for the computed Laplacian values (default: 1).}
}
\value{
image An \code{\link{Image}} object.
}
\description{
\code{laplacian} calculates the Laplacian of the source image by
 adding up the second x and y derivatives calculated using the Sobel operator.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}, \code{\link{sobel}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
