% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{medianBlur}
\alias{medianBlur}
\title{Blurs an Image Using a Median Filter}
\usage{
medianBlur(image, k_size)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{k_size}{The half-size in pixels of the kernel.}
}
\value{
image An \code{\link{Image}} object.
}
\description{
\code{medianBlur} smoothes an image using a median filter, i.e.
 the value of each pixel is replaced by the median value of all the pixels in
 its neighborhood.
}
\examples{
# TODO
}
\seealso{
\code{\link{Image}}, \code{\link{gaussianBlur}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
