% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{moments}
\alias{moments}
\title{Calculate the Moments of a Shape}
\usage{
moments(x, binary = FALSE)
}
\arguments{
\item{x}{Either a Nx2 matrix of the X-Y coordinates of a polygon (e.g., a
contour produced by \code{\link{findContours}}), or a single-channel
\code{\link{Image}} object.}

\item{binary}{If set to TRUE (default: FALSE), all non-zero image pixels are
treated as 1's. The parameter is used for images only.}
}
\value{
A data frame with 2 columns:
 \describe{
   \item{"moment": }{the name of the moment. See Note below.}
   \item{"value": }{the value of the moment.}
}
}
\description{
\code{moments} calculates all of the moments up to the third
 order of a polygon or rasterized shape.
}
\note{
The spatial moments \eqn{m_{ji}} are computed as:
 \deqn{m_{ji}= \sum _{x,y} \left ( \texttt{contour} (x,y) \cdot x^j \cdot y^i \right )}

The central moments \eqn{\mu_{ji}} are computed as:
 \deqn{{\mu_{ji}}= \sum _{x,y} \left ( \texttt{contour} (x,y) \cdot (x - \bar{x} )^j \cdot (y - \bar{y} )^i \right )}
where \eqn{(\bar{x}, \bar{y})} is the mass center:
 \deqn{\bar{x} = \frac{m_{10}}{m_{00}} , \; \bar{y} = \frac{m_{01}}{m_{00}}}

The normalized central moments \eqn{\eta_{ji}} are computed as:
 \deqn{\eta_{ji}= \frac{\mu_{ji}}{m_{00}^{(i+j)/2+1}} .}
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_gray <- changeColorSpace(dots, "GRAY")
dots_bin <- dots_gray < 200
contours <- findContours(dots_bin)
contour0 <- contours$contours[contours$contours[, 1] == 0, 2:3]
moments(contour0)

}
\seealso{
\code{\link{findContours}}, \code{\link{huInvariants}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
