% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{nrow.Rcpp_Image}
\alias{nrow.Rcpp_Image}
\alias{nrow.Image}
\alias{ncol.Rcpp_Image}
\alias{ncol.Image}
\alias{nchan}
\title{The Number of Rows/Columns/Channels of an Image}
\usage{
nrow.Rcpp_Image(x)
ncol.Rcpp_Image(x)
nchan(x)
}
\arguments{
\item{x}{An \code{\link{Image}} object.}
}
\value{
A numeric value.
}
\description{
nrow, ncol and nchan return the number of rows, columns or
 channels present in an \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
nrow(balloon)
ncol(balloon)
nchan(balloon)

}
\seealso{
\code{\link[=dim.Rcpp_Image]{dim}} which returns \emph{all}
 dimensions.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
