% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{ones}
\alias{ones}
\title{Create a One-Filled Image}
\usage{
ones(nrow, ncol, nchan = 3, bitdepth = "8U", colorspace = "BGR")
}
\arguments{
\item{nrow}{An integer indicating desired the number of rows for the image.}

\item{ncol}{An integer indicating desired the number of columns for the image.}

\item{nchan}{An integer indicating the desired number of channels for the image.}

\item{bitdepth}{A string indicating the desired bit depth for the image.
Options are "8U" (the default), "8S", "16U", "16S", "32S", "32F", and "64F".}

\item{colorspace}{A string indicating the desired color space for the image.}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{ones} creates an \code{\link{Image}} object filled with
 ones.
}
\examples{
one <- ones(100, 100)

}
\seealso{
\code{\link{Image}}, \code{\link{zeros}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
