% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{plot.blob}
\alias{plot.blob}
\title{Plot Blobs}
\usage{
\method{plot}{blob}(x, col = "red", asp = 1, ...)
}
\arguments{
\item{x}{A \code{blob} object.}

\item{col}{The color of the \code{blob} center and box (default: "red").}

\item{asp}{The y/x aspect ratio of the plotting window (default: 1).}

\item{...}{Additional arguments to be passed to \code{\link{plot}}.}
}
\description{
Plotting method for objects inheriting from class \code{blob}.
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
blobs <- simpleBlobDetector(invert(dots), min_threshold = 25, max_threshold = 220,
                            filter_by_area = TRUE, min_area = 200, max_area = Inf,
                            filter_by_color = FALSE)
plot(dots)
plot(blobs, pch = 20)

}
\seealso{
\code{\link{isBlob}}, \code{\link{simpleBlobDetector}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
