% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{plot.Image}
\alias{plot.Image}
\alias{plot.Rcpp_Image}
\title{Plot \pkg{Rvision} Images}
\usage{
\method{plot}{Rcpp_Image}(x, ...)
}
\arguments{
\item{x}{An \code{\link{Image}} object.}

\item{...}{Additional arguments to be passed to \code{\link{rasterImage}}.}
}
\description{
Plotting method for objects inheriting from class \code{\link{Image}}.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
plot(balloon)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
