% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\docType{class}
\name{Queue-class}
\alias{Queue-class}
\alias{Queue}
\alias{Rcpp_Queue}
\title{An S4 Class Containing a Queue of Images}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
Queue
}
\description{
A \code{Queue} object contains a self-filling, asynchronous
 queue of images retrieved from a \code{\link{Video}} or \code{\link{Stream}}
 object. Retrieving an image from a video or camera stream is generally fast
 but it still comes with a time penalty that can become significant when
 processing large numbers of images and/or videos. A \code{Queue} object can
 help save some of that lost time by reading and storing images from a video
 or stream in parallel of the main R thread, for instance while other
 operations are being performed by R on previously read images. These
 pre-loaded frames are, therefore, immediately available when needed,
 effectively eliminating the wait time between two frame reads (as long as
 the queue reads and stores new frames faster than R can process them).
}
\section{Slots}{

\describe{
\item{\code{dim,ncol,nrow}}{Functions returning the dimensions of the frames in the
queue.}

\item{\code{length}}{Function returning the number of frames in the queue.}

\item{\code{capacity}}{Function returning the maximum number of frames that the queue
can hold at any one time.}

\item{\code{frame}}{Function returning the index of the frame to be read next.}

\item{\code{full,empty}}{Function returning the state of the queue (full or empty or
neither).}

\item{\code{readNext}}{Function to access the next frame in the queue.}

\item{\code{reset}}{Function to empty the queue and start collecting new frames.}
}}

\author{
Simon Garnier, \email{garnier@njit.edu}
}
\keyword{datasets}
