% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{readFrame}
\alias{readFrame}
\title{Read Specific Video Frame}
\usage{
readFrame(x, pos)
}
\arguments{
\item{x}{A \code{\link{Video}} object.}

\item{pos}{An integer corresponding to the number of the frame to read.}
}
\value{
An \code{\link{Image}} object.
}
\description{
Read a specific frame of a \code{\link{Video}} object and
 returns it as an \code{\link{Image}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
plot(readFrame(balloon, 10))
plot(readNext(balloon))
release(balloon)
}
\seealso{
\code{\link{Video}}, \code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
