% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{readMulti}
\alias{readMulti}
\title{Read a Multi-Page Image}
\usage{
readMulti(x, colorspace = "BGR")
}
\arguments{
\item{x}{A character string naming the path to a multi-page image file.}

\item{colorspace}{A string indicating the desired color space for the images.}
}
\value{
A list of \code{\link{Image}} objects.
}
\description{
\code{readMulti} reads a multi-page image and returns a list of
 \code{\link{Image}} objects, each corresponding to a different page.
}
\examples{
balloon <- readMulti(system.file("sample_img/multipage.tif", package = "Rvision"))

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
