% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R, R/VideoClass.R, R/generic.R
\name{readNext.Rcpp_Stream}
\alias{readNext.Rcpp_Stream}
\alias{readNext.Rcpp_Video}
\alias{readNext}
\title{Read Next of an Object Frame}
\usage{
\method{readNext}{Rcpp_Stream}(obj)

\method{readNext}{Rcpp_Video}(obj)

readNext(obj)
}
\arguments{
\item{obj}{A \code{\link{Video}} or \code{\link{Stream}} object.}
}
\value{
An \code{\link{Image}} object.
}
\description{
Read the next frame of a \code{\link{Video}}
or \code{\link{Stream}} object and returns
 it as an \code{\link{Image}} object.
}
\examples{
# TODO
}
\seealso{
\code{\link{Video}}, \code{\link{Stream}},
\code{\link{Image}}, \code{\link{readFrame}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
