% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R, R/StreamClass.R, R/VideoClass.R,
%   R/generic.R
\name{readNext.Rcpp_Queue}
\alias{readNext.Rcpp_Queue}
\alias{readNext.Rcpp_Stream}
\alias{readNext.Rcpp_Video}
\alias{readNext}
\title{Read Next Frame of an Object}
\usage{
\method{readNext}{Rcpp_Queue}(x, target = "new")

\method{readNext}{Rcpp_Stream}(x, target = "new")

\method{readNext}{Rcpp_Video}(x, target = "new")

readNext(x, target = "new")
}
\arguments{
\item{x}{A \code{\link{Video}}, \code{\link{Stream}}, or \code{\link{Queue}}
object.}

\item{target}{The location where the results should be stored. It can take 2
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This will replace the content of
  \code{target}. Note that \code{target} must have the same dimensions as
  \code{x}.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target} is an \code{\link{Image}} object, the function
 returns nothing and modifies that \code{\link{Image}} object in place.
}
\description{
Read the next frame of a \code{\link{Video}}, \code{\link{Stream}},
 or \code{\link{Queue}} object and returns it as an \code{\link{Image}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
frame_next <- readNext(balloon)

}
\seealso{
\code{\link{Video}}, \code{\link{Stream}}, \code{\link{Queue}},
 \code{\link{Image}}, \code{\link{readFrame}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
