% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\name{release.Rcpp_Video}
\alias{release.Rcpp_Video}
\title{Release Video from Memory}
\usage{
\method{release}{Rcpp_Video}(obj)
}
\arguments{
\item{obj}{A \code{\link{Video}} object.}
}
\value{
If successful, the \code{\link{Video}} object is cleared from memory
}
\description{
Close a \code{\link{Video}} object.
}
\examples{
# TODO

}
\seealso{
\code{\link{Video}}, \code{\link{video}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
