% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R, R/StreamClass.R, R/VideoClass.R,
%   R/VideoWriterClass.R, R/generic.R
\name{release.Rcpp_Queue}
\alias{release.Rcpp_Queue}
\alias{release.Rcpp_Stream}
\alias{release.Rcpp_Video}
\alias{release.Rcpp_VideoWriter}
\alias{release}
\title{Release Object from Memory}
\usage{
\method{release}{Rcpp_Queue}(x)

\method{release}{Rcpp_Stream}(x)

\method{release}{Rcpp_Video}(x)

\method{release}{Rcpp_VideoWriter}(x)

release(x)
}
\arguments{
\item{x}{A \code{\link{Video}}, \code{\link{Stream}}, \code{\link{VideoWriter}},
or \code{\link{Queue}}, object.}
}
\value{
If successful, the object is cleared from memory
}
\description{
Close a \code{\link{Video}}, \code{\link{Stream}},
 \code{\link{VideoWriter}} or \code{\link{Queue}} object.
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
release(balloon)

}
\seealso{
\code{\link{Video}}, \code{\link{Stream}}, \code{\link{VideoWriter}},
 \code{\link{Queue}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
