% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{resize}
\alias{resize}
\title{Resize an \code{\link{Image}}}
\usage{
resize(
  image,
  height = NULL,
  width = NULL,
  fx = NULL,
  fy = NULL,
  interpolation = "linear",
  target = "new"
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{height}{A positive integer representing the new height in pixels of the
image (default: NULL).}

\item{width}{A positive integer representing the new width in pixels of the
image (default: NULL).}

\item{fx}{A positive numeric representing the ratio by which the width of
the image must be resized (default: NULL). Ignored if \code{width} is set.}

\item{fy}{A positive numeric representing the ratio by which the height of
the image must be resized (default: NULL). Ignored if \code{height} is set.}

\item{interpolation}{A character string representing the type of interpolation
to use during resizing (default: "linear"). See notes for all accepted
interpolation methods.}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast but will replace the
  content of \code{target}. Note that \code{target} must have the same bit
  depth and number of channels as \code{image} but that its dimensions must
  match that of the resized image, otherwise an error is thrown.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target} is an \code{\link{Image}} object, the function
 returns nothing and modifies that \code{\link{Image}} object in place.
}
\description{
\code{resize} returns a resized version of an \code{\link{Image}}.
}
\note{
The following interpolation methods are supported:
 \itemize{
  \item{"nearest": }{nearest neighbor interpolation.}
  \item{"linear": }{bilinear interpolation.}
  \item{"cubic": }{bicubic interpolation.}
  \item{"area": }{resampling using pixel area relation.}
  \item{"Lanczos": }{Lanczos interpolation over 8x8 neighborhood.}
  \item{"exact": }{bit exact bilinear interpolation.}
 }
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_resized <- resize(balloon, fx = 0.2, fy = 0.5)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
