% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{rotateScale}
\alias{rotateScale}
\title{Image Rotation and Scaling}
\usage{
rotateScale(
  image,
  center = (dim(image)[2:1] - 1)/2,
  angle = 90,
  scale = 1,
  ...
)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{center}{A 2-elements vector indicating the location (row, column) of
the center of the rotation in the source image. It defaults to the center of
the image.}

\item{angle}{A numeric value indicating the rotation angle in degrees
(default: 90).}

\item{scale}{A numeric value indicating an isotropic scale factor (default: 1).}

\item{...}{Additional parameters for the \code{\link{warpAffine}} function.}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{rotateScale} rotates (clockwise) and scales an image using
 the \code{\link{warpAffine}} function.
}
\examples{
img <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
img_rotated <- rotateScale(img, c(50, 50), 45, 1)

}
\seealso{
\code{\link{warpAffine}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
