% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rvision-package.R
\docType{package}
\name{Rvision}
\alias{Rvision}
\title{Rvision: Computer vision library for R}
\description{
The \code{Rvision} package contains an OpenCV-based computer vision library
providing methods to read and manipulate video and image files, as well as
camera streams.
}
\details{
\tabular{ll}{
 Package: \tab Rvision\cr
 Type: \tab Package\cr
 Version: \tab 0.1.0\cr
 Date: \tab 2016-07-06\cr
 License: \tab GPL-3\cr
}
}
\references{
More about OpenCV at \url{http://opencv.org/}.
}
\author{
Simon Garnier <garnier@njit.edu>

John Muschelli <muschellij2@gmail.com>

Maintainer: Simon Garnier <garnier@njit.edu>
}
