% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{scharr}
\alias{scharr}
\title{Calculates an Image Derivatives Using a Scharr Operator}
\usage{
scharr(image, dx = 1, dy = 1, scale = 1)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{dx}{Order of the x derivative (default: 1).}

\item{dy}{Order of the y derivative (default: 1),}

\item{scale}{The scale factor for the computed derivative values (default: 1).}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{scharr} calculates the derivatives of an image using a
 Scharr operator.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_scharr <- scharr(balloon, 1, 1, 1)
plot(balloon_scharr)

}
\seealso{
\code{\link{Image}}, \code{\link{sobel}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
