% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{scharr}
\alias{scharr}
\title{Calculates an Image Derivatives Using a Scharr Operator}
\usage{
scharr(image, dx = 1, dy = 0, scale = 1, target = "new", in_place = NULL)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{dx}{Order of the x derivative. It can only take 2 values: 1 (the
default) or 0. If \code{dx=1} then \code{dy} must be zero.}

\item{dy}{Order of the y derivative. It can only take 2 values: 1 or 0 (the
default). If \code{dy=1} then \code{dx} must be zero.}

\item{scale}{The scale factor for the computed derivative values (default: 1).}

\item{target}{The location where the results should be stored. It can take 3
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{"self":}{the results are stored back into \code{image} (faster but
  destructive).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast and will not replace the
  content of \code{image} but will replace that of \code{target}. Note that
  if \code{target} does not have the same dimensions, number of channels, and
  bit depth as \code{image}, an error may be thrown.}
}}

\item{in_place}{Deprecated. Use \code{target} instead.}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target="self"}, the function returns nothing and modifies
 \code{image} in place. If \code{target} is an \code{\link{Image}} object,
 the function returns nothing and modifies that \code{\link{Image}} object in
 place.
}
\description{
\code{scharr} calculates the x or y derivative of an image using
 a Scharr operator.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_scharr <- scharr(balloon)

}
\seealso{
\code{\link{Image}}, \code{\link{sobel}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
