% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{spatialGradient}
\alias{spatialGradient}
\title{First Order Derivatives of an Image with the Sobel Operator}
\usage{
spatialGradient(image, k_size = 5)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{k_size}{The half-size in pixels of the kernel (default: 5).}
}
\value{
A list containing two \code{\link{Image}} objects, one of for the
 derivative along the x axis and the other for the derivative along the y axis.
}
\description{
\code{spatialGradient} calculates the first order derivative of
 an image in both x and y using a Sobel operator.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_gradient <- spatialGradient(balloon, 5)

}
\seealso{
\code{\link{Image}}, \code{\link{sobel}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
