% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{split}
\alias{split}
\title{Split an Image into Separate Channels}
\usage{
split(x)
}
\arguments{
\item{x}{An \code{\link{Image}} object.}
}
\value{
A list of single channel (grayscale) \code{\link{Image}} objects.
}
\description{
\code{split} returns a list of grayscale images corresponding to
 each of the channels (blue, green, red, or alpha) of an image.
}
\note{
Color images are usually represented by 3 channels (possibly 4) in the
 following order: blue (1), green (2), red (3), and possibly alpha (4).
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_chan <- split(balloon)

}
\seealso{
\code{\link{merge}}, \code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
