% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R
\docType{class}
\name{Stream-class}
\alias{Stream-class}
\alias{Stream}
\alias{Rcpp_Stream}
\title{An S4 Class Containing an OpenCV Stream}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
Stream
}
\description{
A \code{Stream} object contains an \href{http://opencv.org/}{OpenCV}
 stream that originates from a camera connected to the computer.
}
\section{Slots}{

\describe{
\item{\code{dim,ncol,nrow}}{Functions returning the dimensions of the object.}

\item{\code{get,set}}{Functions to access and set internal properties of the object.}

\item{\code{isOpened}}{Function to check the status of the camera stream.}

\item{\code{readNext}}{Functions to access the next frame in the stream.}

\item{\code{release}}{Function to release the object from memory.}
}}

\author{
Simon Garnier, \email{garnier@njit.edu}
}
\keyword{datasets}
