% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamClass.R
\docType{class}
\name{Stream-class}
\alias{Stream-class}
\alias{Stream}
\alias{Rcpp_Stream}
\title{An S4 Class Containing an OpenCV Stream}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
Stream
}
\description{
\code{Stream} objects contains an \href{http://opencv.org/}{OpenCV}
 stream that originates from a camera connected to the computer.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\keyword{datasets}
