% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{subImage}
\alias{subImage}
\title{Extract Subimage}
\usage{
subImage(image, x, y, width, height, border = TRUE, ...)
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{x, y}{The coordinates of the bottom-left corner of the subimage within
the original image.}

\item{width}{The width of the subimage.}

\item{height}{The height of the subimage.}

\item{border}{If the subimage extends beyond the boundaries of the source
image and \code{border = TRUE} (the default), a border is created to account
for the missing pixels.}

\item{...}{Additional parameters to be passed to \code{\link{border}} if the
the subimage extends beyond the boundaries of the source image and
\code{border = TRUE}. These parameters control for the type and color of the
border.}
}
\value{
An \code{\link{Image}} object.
}
\description{
\code{subImage} extracts a portion of an \code{\link{Image}} and
 returns it as an \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_sub <- subImage(balloon, 290, 170, 150, 150)
plot(balloon_sub)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
