% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageClass.R
\name{subImage}
\alias{subImage}
\title{Extract Subimage}
\usage{
subImage(image, x, y, width, height, target = "new")
}
\arguments{
\item{image}{An \code{\link{Image}} object.}

\item{x, y}{The coordinates of the bottom-left corner of the subimage within
the original image.}

\item{width}{The width of the subimage. Ignored if \code{target} is an
\code{\link{Image}} object.}

\item{height}{The height of the subimage. Ignored if \code{target} is an
\code{\link{Image}} object.}

\item{target}{The location where the results should be stored. It can take 2
values:
\itemize{
 \item{"new":}{a new \code{\link{Image}} object is created and the results
  are stored inside (the default).}
 \item{An \code{\link{Image}} object:}{the results are stored in another
  existing \code{\link{Image}} object. This is fast but will replace the
  content of \code{target}. Note that if \code{target} does not have the
  same number of channels and bit depth as \code{image}, an error will be
  thrown.}
}}
}
\value{
If \code{target="new"}, the function returns an \code{\link{Image}}
 object. If \code{target} is an \code{\link{Image}} object, the function
 returns nothing and modifies that \code{\link{Image}} object in place.
}
\description{
\code{subImage} extracts a portion of an \code{\link{Image}} and
 returns it as an \code{\link{Image}} object.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_sub <- subImage(balloon, 290, 170, 150, 150)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
