% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sum}
\alias{sum}
\title{Sum Generic for Image objects}
\usage{
sum(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An \code{\link{Image}} object or a list of \code{\link{Image}}
objects.}

\item{...}{Further arguments passed to summary methods.

#' @examples
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
sum(balloon)}
}
\description{
Overloaded \code{\link[base]{sum}} to handle \code{\link{Image}}
 objects.
}
