% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sum}
\alias{sum}
\title{Sum Generic for Image objects}
\usage{
sum(x, ...)
}
\arguments{
\item{x}{An \code{\link{Image}} object or a list of \code{\link{Image}}
objects.}

\item{...}{Further arguments passed to summary methods. Unused if \code{x} is
an image or a list of images.}
}
\value{
If \code{x} is an \code{\link{Image}} object, the function returns a
 numeric value (for single-channel images) or a vector of numeric values (for
 multi-channel images).

 If \code{x} is a list of \code{\link{Image}} objects and \code{target="new"},
 the function returns an \code{\link{Image}} object. If \code{target} is an
 \code{\link{Image}} object, the function returns nothing and modifies that
 \code{\link{Image}} object in place.
}
\description{
Overloaded \code{\link[base]{sum}} to handle \code{\link{Image}}
 objects and lists of \code{\link{Image}} objects.
}
\examples{
balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
sum(balloon)

balloon_vid <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
img_list <- lapply(1:10, function(x) readNext(balloon_vid))
sum(img_list)

}
\seealso{
\code{\link{Image}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
