% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\docType{class}
\name{Video-class}
\alias{Video-class}
\alias{video}
\alias{Rcpp_Video}
\alias{Video}
\title{An S4 Class Containing an OpenCV Video}
\usage{
video(filename, api = "ANY")
}
\arguments{
\item{filename}{An character string corresponding to the path to a video file.}

\item{api}{A character string corresponding to the API to use for reading the
video from the file (see Note; default: "ANY").}
}
\value{
A \code{\link{Video}} object.
}
\description{
A \code{Video} object contains an \href{http://opencv.org/}{OpenCV}
 video that originates from a video file.

Function for creating \code{\link{Video}} objects from video
 files.
}
\section{Slots}{

\describe{
\item{\code{dim,ncol,nrow}}{Functions returning the dimensions of the object.}

\item{\code{nframes}}{Function returning the number of frames in the object.}

\item{\code{codec}}{Function returning the codec of the object.}

\item{\code{fps}}{Function returning the frame rate of the object.}

\item{\code{frame}}{Function returning the index of the frame to be read next.}

\item{\code{get,set}}{Functions to access and set internal properties of the object.}

\item{\code{isOpened}}{Function to check the status of the video stream.}

\item{\code{readNext,readFrame}}{Functions to access the next or an arbitrary frame
in the stream.}

\item{\code{release}}{Function to release the object from memory.}
}}

\note{
Hereafter is a list of all supported APIs. Note that not all APIs will
 be available on your computer (actually most of them will not be).
 \itemize{
  \item{"ANY": }{automatically select an API.}
  \item{"VFW": }{Video For Windows.}
  \item{"V4L", "V4L2": }{Video For Linux.}
  \item{"FIREWIRE", "FIREWARE", "IEEE1394", "DC1394", "CMU1394": }{IEEE 1394 drivers.}
  \item{"QT": }{Quicktime.}
  \item{"UNICAP": }{Unicap drivers.}
  \item{"DSHOW": }{DirectShow.}
  \item{"PVAPI": }{PvAPI, Prosilica GigE SDK.}
  \item{"OPENNI": }{OpenNI (for Kinect).}
  \item{"OPENNI_ASUS": }{OpenNI (for Asus Xtion).}
  \item{"XIAPI": }{XIMEA Camera API.}
  \item{"AVFOUNDATION": }{AVFoundation framework for iOS and OSX > Lion.}
  \item{"GIGANETIX": }{Smartek Giganetix GigEVisionSDK.}
  \item{"MSMF": }{Microsoft Media Foundation.}
  \item{"WINRT": }{Microsoft Windows Runtime using Media Foundation.}
  \item{"INTELPERC": }{Intel Perceptual Computing SDK.}
  \item{"OPENNI2": }{OpenNI2 (for Kinect).}
  \item{"OPENNI2_ASUS": }{OpenNI2 (for Asus Xtion and Occipital Structure sensors).}
  \item{"GPHOTO2": }{gPhoto2 connection.}
  \item{"GSTREAMER": }{GStreamer.}
  \item{"FFMPEG": }{FFMPEG library.}
  \item{"IMAGES": }{OpenCV Image Sequence.}
  \item{"ARAVIS": }{Aravis SDK.}
  \item{"OPENCV_MJPEG": }{Built-in OpenCV MotionJPEG codec.}
  \item{"INTEL_MFX": }{Intel MediaSDK.}
 }
}
\examples{
balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))

}
\seealso{
\code{\link{video}}, \code{\link{queue}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
