% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\docType{class}
\name{Video-class}
\alias{Video-class}
\alias{Video}
\alias{Rcpp_Video}
\title{An S4 Class Containing an OpenCV Video}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
Video
}
\description{
\code{Video} objects contains an \href{http://opencv.org/}{OpenCV}
 video that originates from a video file.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\keyword{datasets}
