% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoClass.R
\docType{class}
\name{Video-class}
\alias{Video-class}
\alias{Video}
\alias{Rcpp_Video}
\title{An S4 Class Containing an OpenCV Video}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
Video
}
\description{
A \code{Video} object contains an \href{http://opencv.org/}{OpenCV}
 video that originates from a video file.
}
\section{Slots}{

\describe{
\item{\code{dim,ncol,nrow}}{Functions returning the dimensions of the object.}

\item{\code{nframes}}{Function returning the number of frames in the object.}

\item{\code{codec}}{Function returning the codec of the object.}

\item{\code{fps}}{Function returning the frame rate of the object.}

\item{\code{frame}}{Function returning the index of the frame to be read next.}

\item{\code{get,set}}{Functions to access and set internal properties of the object.}

\item{\code{isOpened}}{Function to check the status of the video stream.}

\item{\code{readNext,readFrame}}{Functions to access the next or an arbitrary frame
in the stream.}

\item{\code{release}}{Function to release the object from memory.}
}}

\author{
Simon Garnier, \email{garnier@njit.edu}
}
\keyword{datasets}
