% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VideoWriterClass.R
\docType{class}
\name{VideoWriter-class}
\alias{VideoWriter-class}
\alias{VideoWriter}
\alias{Rcpp_VideoWriter}
\title{An S4 Class Containing an OpenCV Video Writer}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
VideoWriter
}
\description{
\code{VideoWriter} objects contains an \href{http://opencv.org/}{OpenCV}
 video writer to an output file.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\keyword{datasets}
