% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{watershed}
\alias{watershed}
\title{Image Segmentation Using the Watershed Algorithm}
\usage{
watershed(image, markers)
}
\arguments{
\item{image}{An 8-bit (8U), BGR \code{\link{Image}} object.}

\item{markers}{A signed 32-bit (32S) single-channel (GRAY) \code{\link{Image}}
object (see Details).}
}
\value{
This function does not return anything. It modifies \code{markers} in
 place.
}
\description{
\code{watershed} implements one of the variants of watershed,
 non-parametric marker-based segmentation algorithm, described in Meyer (1992).
}
\details{
Before passing \code{image} to the function, you have to roughly
 outline the desired regions in the \code{markers} image with positive (>0)
 indices. So, every region is represented as one or more connected components
 with the pixel values 1, 2, 3, and so on. Such markers can be retrieved from
 a binary mask using \code{\link{findContours}}. The markers are "seeds" of
 the future image regions. All the other pixels in \code{markers}, whose
 relation to the outlined regions is not known and should be defined by the
 algorithm, should be set to 0's. In the function output, each pixel in
 \code{markers} is set to a value of the "seed" components or to -1 at
 boundaries between the regions.
}
\examples{
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
bw <- inRange(dots, 0, 250)
medianBlur(bw, target = "self")
sure_bg <- morph(bw, "dilate", k_shape = "ellipse", iterations = 3)
dt <- distanceTransform(bw, "L2")
sure_fg <- dt > 20
unknown <- sure_bg - sure_fg
markers <- connectedComponents(sure_fg, table = FALSE)$labels + 1
markers <- markers * (invert(unknown) / 255)
watershed(dots, markers)

}
\references{
Meyer F. Color image segmentation. 1992 International Conference
 on Image Processing. 1992. Available:
 https://ieeexplore.ieee.org/abstract/document/785528/
}
\seealso{
\code{\link{findContours}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
